Delta Method Sampling Variance-Covariance Matrix for the Elements of the Standardized Matrix of Lagged Coefficients Over a Specific Time Interval or a Range of Time Intervals

This function computes the delta method sampling variance-covariance matrix for the elements of the standardized matrix of lagged coefficients \(\boldsymbol{\beta}\) over a specific time interval \(\Delta t\) or a range of time intervals using the first-order stochastic differential equation model's drift matrix \(\boldsymbol{\Phi}\) and process noise covariance matrix \(\boldsymbol{\Sigma}\).

Usage

DeltaBetaStd(phi, vcov_phi_vec, sigma, vcov_sigma_vech, delta_t, ncores = NULL)

Arguments

phi

Numeric matrix. The drift matrix (\(\boldsymbol{\Phi}\)). phi should have row and column names pertaining to the variables in the system.

vcov_phi_vec

Numeric matrix. The sampling variance-covariance matrix of \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\).

sigma

Numeric matrix. The process noise covariance matrix (\(\boldsymbol{\Sigma}\)).

vcov_sigma_vech

Numeric matrix. The sampling variance-covariance matrix of \(\mathrm{vech} \left( \boldsymbol{\Sigma} \right)\).

delta_t

Numeric. Time interval (\(\Delta t\)).

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of replications R is a large value.

Value

Returns an object of class ctmeddelta which is a list with the following elements:

call

Function call.

args

Function arguments.

fun

Function used ("DeltaBetaStd").

output

A list the length of which is equal to the length of delta_t.

Each element in the output list has the following elements:

delta_t

Time interval.

jacobian

Jacobian matrix.

est

Estimated elements of the matrix of lagged coefficients.

vcov

Sampling variance-covariance matrix of estimated elements of the matrix of lagged coefficients.

Details

See TotalStd().

Delta Method

Let \(\boldsymbol{\theta}\) be a vector that combines \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\), that is, the elements of the \(\boldsymbol{\Phi}\) matrix in vector form sorted column-wise and \(\mathrm{vech} \left( \boldsymbol{\Sigma} \right)\), that is, the unique elements of the \(\boldsymbol{\Sigma}\) matrix in vector form sorted column-wise. Let \(\hat{\boldsymbol{\theta}}\) be a vector that combines \(\{\mathrm{vec} \left( \hat{\boldsymbol{\Phi}} \right)\) and \(\mathrm{vech} \left( \hat{\boldsymbol{\Sigma}} \right)\). By the multivariate central limit theory, the function \(\mathbf{g}\) using \(\hat{\boldsymbol{\theta}}\) as input can be expressed as:

$$ \sqrt{n} \left( \mathbf{g} \left( \hat{\boldsymbol{\theta}} \right) - \mathbf{g} \left( \boldsymbol{\theta} \right) \right) \xrightarrow[]{ \mathrm{D} } \mathcal{N} \left( 0, \mathbf{J} \boldsymbol{\Gamma} \mathbf{J}^{\prime} \right) $$

where \(\mathbf{J}\) is the matrix of first-order derivatives of the function \(\mathbf{g}\) with respect to the elements of \(\boldsymbol{\theta}\) and \(\boldsymbol{\Gamma}\) is the asymptotic variance-covariance matrix of \(\hat{\boldsymbol{\theta}}\).

From the former, we can derive the distribution of \(\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)\) as follows:

$$ \mathbf{g} \left( \hat{\boldsymbol{\theta}} \right) \approx \mathcal{N} \left( \mathbf{g} \left( \boldsymbol{\theta} \right) , n^{-1} \mathbf{J} \boldsymbol{\Gamma} \mathbf{J}^{\prime} \right) $$

The uncertainty associated with the estimator \(\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)\) is, therefore, given by \(n^{-1} \mathbf{J} \boldsymbol{\Gamma} \mathbf{J}^{\prime}\) . When \(\boldsymbol{\Gamma}\) is unknown, by substitution, we can use the estimated sampling variance-covariance matrix of \(\hat{\boldsymbol{\theta}}\), that is, \(\hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right)\) for \(n^{-1} \boldsymbol{\Gamma}\). Therefore, the sampling variance-covariance matrix of \(\mathbf{g} \left( \hat{\boldsymbol{\theta}} \right)\) is given by

$$ \mathbf{g} \left( \hat{\boldsymbol{\theta}} \right) \approx \mathcal{N} \left( \mathbf{g} \left( \boldsymbol{\theta} \right) , \mathbf{J} \hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right) \mathbf{J}^{\prime} \right) . $$

References

Bollen, K. A. (1987). Total, direct, and indirect effects in structural equation models. Sociological Methodology, 17, 37. doi:10.2307/271028

Deboeck, P. R., & Preacher, K. J. (2015). No need to be discrete: A method for continuous time mediation analysis. Structural Equation Modeling: A Multidisciplinary Journal, 23 (1), 61–75. doi:10.1080/10705511.2014.973960

Ryan, O., & Hamaker, E. L. (2021). Time to intervene: A continuous-time approach to network analysis and centrality. Psychometrika, 87 (1), 214–252. doi:10.1007/s11336-021-09767-0

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.00843, 0.00040, -0.00151,
    -0.00600, -0.00033, 0.00110,
    0.00324, 0.00020, -0.00061,
    0.00040, 0.00374, 0.00016,
    -0.00022, -0.00273, -0.00016,
    0.00009, 0.00150, 0.00012,
    -0.00151, 0.00016, 0.00389,
    0.00103, -0.00007, -0.00283,
    -0.00050, 0.00000, 0.00156,
    -0.00600, -0.00022, 0.00103,
    0.00644, 0.00031, -0.00119,
    -0.00374, -0.00021, 0.00070,
    -0.00033, -0.00273, -0.00007,
    0.00031, 0.00287, 0.00013,
    -0.00014, -0.00170, -0.00012,
    0.00110, -0.00016, -0.00283,
    -0.00119, 0.00013, 0.00297,
    0.00063, -0.00004, -0.00177,
    0.00324, 0.00009, -0.00050,
    -0.00374, -0.00014, 0.00063,
    0.00495, 0.00024, -0.00093,
    0.00020, 0.00150, 0.00000,
    -0.00021, -0.00170, -0.00004,
    0.00024, 0.00214, 0.00012,
    -0.00061, 0.00012, 0.00156,
    0.00070, -0.00012, -0.00177,
    -0.00093, 0.00012, 0.00223
  ),
  nrow = 9
)
sigma <- matrix(
  data = c(
    0.24, 0.02, -0.05,
    0.02, 0.07, 0.02,
    -0.05, 0.02, 0.08
  ),
  nrow = 3
)
vcov_sigma_vech <- matrix(
  data = c(
    0.00057, 0.00001, -0.00009,
    0.00000, 0.00000, 0.00001,
    0.00001, 0.00012, 0.00001,
    0.00000, -0.00002, 0.00000,
    -0.00009, 0.00001, 0.00014,
    0.00000, 0.00000, -0.00005,
    0.00000, 0.00000, 0.00000,
    0.00010, 0.00001, 0.00000,
    0.00000, -0.00002, 0.00000,
    0.00001, 0.00005, 0.00001,
    0.00001, 0.00000, -0.00005,
    0.00000, 0.00001, 0.00012
  ),
  nrow = 6
)

# Specific time interval ----------------------------------------------------
DeltaBetaStd(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  sigma = sigma,
  vcov_sigma_vech = vcov_sigma_vech,
  delta_t = 1
)
#> 
#> Elements of the matrix of lagged coefficients
#> 
#> $`1`
#>       interval     est     se       z     p    2.5%   97.5%
#>  [1,]        1  0.6998 0.0471 14.8688 0e+00  0.6075  0.7920
#>  [2,]        1  0.6418 0.0648  9.9039 0e+00  0.5148  0.7688
#>  [3,]        1 -0.0949 0.0286 -3.3145 9e-04 -0.1509 -0.0388
#>  [4,]        1  0.0000 0.0339  0.0000 1e+00 -0.0664  0.0664
#>  [5,]        1  0.5999 0.0326 18.3826 0e+00  0.5359  0.6639
#>  [6,]        1  0.2954 0.0331  8.9153 0e+00  0.2304  0.3603
#>  [7,]        1  0.0000 0.0441  0.0000 1e+00 -0.0864  0.0864
#>  [8,]        1  0.0000 0.0421  0.0000 1e+00 -0.0824  0.0824
#>  [9,]        1  0.5001 0.0274 18.2776 0e+00  0.4464  0.5537
#> 

# Range of time intervals ---------------------------------------------------
delta <- DeltaBetaStd(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  sigma = sigma,
  vcov_sigma_vech = vcov_sigma_vech,
  delta_t = 1:5
)
plot(delta)

# Methods -------------------------------------------------------------------
# DeltaBetaStd has a number of methods including
# print, summary, confint, and plot
print(delta)
#> 
#> Elements of the matrix of lagged coefficients
#> 
#> $`1`
#>       interval     est     se       z     p    2.5%   97.5%
#>  [1,]        1  0.6998 0.0471 14.8688 0e+00  0.6075  0.7920
#>  [2,]        1  0.6418 0.0648  9.9039 0e+00  0.5148  0.7688
#>  [3,]        1 -0.0949 0.0286 -3.3145 9e-04 -0.1509 -0.0388
#>  [4,]        1  0.0000 0.0339  0.0000 1e+00 -0.0664  0.0664
#>  [5,]        1  0.5999 0.0326 18.3826 0e+00  0.5359  0.6639
#>  [6,]        1  0.2954 0.0331  8.9153 0e+00  0.2304  0.3603
#>  [7,]        1  0.0000 0.0441  0.0000 1e+00 -0.0864  0.0864
#>  [8,]        1  0.0000 0.0421  0.0000 1e+00 -0.0824  0.0824
#>  [9,]        1  0.5001 0.0274 18.2776 0e+00  0.4464  0.5537
#> 
#> $`2`
#>       interval    est     se      z      p    2.5%  97.5%
#>  [1,]        2 0.4897 0.0548 8.9377 0.0000  0.3823 0.5971
#>  [2,]        2 0.8341 0.0952 8.7638 0.0000  0.6476 1.0207
#>  [3,]        2 0.0758 0.0347 2.1824 0.0291  0.0077 0.1438
#>  [4,]        2 0.0000 0.0400 0.0000 1.0000 -0.0784 0.0784
#>  [5,]        2 0.3599 0.0504 7.1405 0.0000  0.2611 0.4587
#>  [6,]        2 0.3249 0.0389 8.3549 0.0000  0.2487 0.4011
#>  [7,]        2 0.0000 0.0529 0.0000 1.0000 -0.1037 0.1037
#>  [8,]        2 0.0000 0.0668 0.0000 1.0000 -0.1309 0.1309
#>  [9,]        2 0.2501 0.0318 7.8668 0.0000  0.1878 0.3124
#> 
#> $`3`
#>       interval    est     se      z     p    2.5%  97.5%
#>  [1,]        3 0.3427 0.0546 6.2779 0e+00  0.2357 0.4496
#>  [2,]        3 0.8147 0.1108 7.3554 0e+00  0.5976 1.0317
#>  [3,]        3 0.2378 0.0440 5.3996 0e+00  0.1515 0.3241
#>  [4,]        3 0.0000 0.0388 0.0000 1e+00 -0.0761 0.0761
#>  [5,]        3 0.2159 0.0609 3.5452 4e-04  0.0965 0.3352
#>  [6,]        3 0.2688 0.0360 7.4696 0e+00  0.1983 0.3393
#>  [7,]        3 0.0000 0.0481 0.0000 1e+00 -0.0942 0.0942
#>  [8,]        3 0.0000 0.0795 0.0000 1e+00 -0.1558 0.1558
#>  [9,]        3 0.1251 0.0299 4.1799 0e+00  0.0664 0.1837
#> 
#> $`4`
#>       interval    est     se      z      p    2.5%  97.5%
#>  [1,]        4 0.2398 0.0536 4.4747 0.0000  0.1348 0.3448
#>  [2,]        4 0.7086 0.1169 6.0634 0.0000  0.4796 0.9377
#>  [3,]        4 0.3271 0.0529 6.1856 0.0000  0.2234 0.4307
#>  [4,]        4 0.0000 0.0355 0.0000 1.0000 -0.0697 0.0697
#>  [5,]        4 0.1295 0.0650 1.9937 0.0462  0.0022 0.2568
#>  [6,]        4 0.1982 0.0328 6.0417 0.0000  0.1339 0.2625
#>  [7,]        4 0.0000 0.0391 0.0000 1.0000 -0.0767 0.0767
#>  [8,]        4 0.0000 0.0811 0.0000 1.0000 -0.1589 0.1589
#>  [9,]        4 0.0625 0.0310 2.0161 0.0438  0.0017 0.1233
#> 
#> $`5`
#>       interval    est     se      z      p    2.5%  97.5%
#>  [1,]        5 0.1678 0.0527 3.1821 0.0015  0.0644 0.2712
#>  [2,]        5 0.5790 0.1174 4.9307 0.0000  0.3488 0.8092
#>  [3,]        5 0.3501 0.0578 6.0524 0.0000  0.2367 0.4635
#>  [4,]        5 0.0000 0.0313 0.0000 1.0000 -0.0613 0.0613
#>  [5,]        5 0.0777 0.0642 1.2092 0.2266 -0.0482 0.2036
#>  [6,]        5 0.1374 0.0313 4.3912 0.0000  0.0760 0.1987
#>  [7,]        5 0.0000 0.0301 0.0000 1.0000 -0.0591 0.0591
#>  [8,]        5 0.0000 0.0750 0.0000 1.0000 -0.1470 0.1470
#>  [9,]        5 0.0313 0.0341 0.9180 0.3586 -0.0355 0.0980
#> 
summary(delta)
#>          97.5% interval         est         se         z            p
#> 1   0.79201437        1  0.69977250 0.04706304 14.868833 5.252120e-50
#> 2   0.76880726        1  0.64179702 0.06480233  9.903919 4.002769e-23
#> 3  -0.03876505        1 -0.09485636 0.02861854 -3.314507 9.180492e-04
#> 4   0.06636564        1  0.00000000 0.03386064  0.000000 1.000000e+00
#> 5   0.66385674        1  0.59989538 0.03263394 18.382559 1.812145e-75
#> 6   0.36031867        1  0.29538149 0.03313182  8.915341 4.863180e-19
#> 7   0.08643452        1  0.00000000 0.04410006  0.000000 1.000000e+00
#> 8   0.08242751        1  0.00000000 0.04205563  0.000000 1.000000e+00
#> 9   0.55369804        1  0.50007360 0.02735991 18.277603 1.247910e-74
#> 10  0.59706425        2  0.48968155 0.05478810  8.937735 3.972267e-19
#> 11  1.02066951        2  0.83412297 0.09517855  8.763770 1.888264e-18
#> 12  0.14380151        2  0.07576192 0.03471471  2.182415 2.907889e-02
#> 13  0.07839563        2  0.00000000 0.03999851  0.000000 1.000000e+00
#> 14  0.45865526        2  0.35987447 0.05039929  7.140467 9.301408e-13
#> 15  0.40113055        2  0.32491048 0.03888851  8.354923 6.547527e-17
#> 16  0.10370813        2  0.00000000 0.05291328  0.000000 1.000000e+00
#> 17  0.13090085        2  0.00000000 0.06678737  0.000000 1.000000e+00
#> 18  0.31237753        2  0.25007360 0.03178830  7.866843 3.637020e-15
#> 19  0.44964534        3  0.34266568 0.05458246  6.277945 3.430781e-10
#> 20  1.03174117        3  0.81466268 0.11075637  7.355448 1.902882e-13
#> 21  0.32414644        3  0.23782161 0.04404409  5.399626 6.678002e-08
#> 22  0.07607906        3  0.00000000 0.03881656  0.000000 1.000000e+00
#> 23  0.33523862        3  0.21588703 0.06089478  3.545247 3.922457e-04
#> 24  0.33930497        3  0.26877941 0.03598309  7.469603 8.043719e-14
#> 25  0.09418709        3  0.00000000 0.04805552  0.000000 1.000000e+00
#> 26  0.15583195        3  0.00000000 0.07950756  0.000000 1.000000e+00
#> 27  0.18369339        3  0.12505520 0.02991799  4.179933 2.915945e-05
#> 28  0.34481734        4  0.23978802 0.05358737  4.474711 7.651461e-06
#> 29  0.93769708        4  0.70863419 0.11687097  6.063389 1.332824e-09
#> 30  0.43069193        4  0.32706056 0.05287412  6.185646 6.184870e-10
#> 31  0.06965220        4  0.00000000 0.03553749  0.000000 1.000000e+00
#> 32  0.25682686        4  0.12950963 0.06495896  1.993715 4.618325e-02
#> 33  0.26246867        4  0.19817852 0.03280170  6.041715 1.524844e-09
#> 34  0.07671862        4  0.00000000 0.03914287  0.000000 1.000000e+00
#> 35  0.15894648        4  0.00000000 0.08109663  0.000000 1.000000e+00
#> 36  0.12333269        4  0.06253681 0.03101888  2.016088 4.379072e-02
#> 37  0.27115007        5  0.16779706 0.05273210  3.182067 1.462281e-03
#> 38  0.80915611        5  0.57900162 0.11742792  4.930698 8.193631e-07
#> 39  0.46350849        5  0.35012636 0.05784909  6.052409 1.426957e-09
#> 40  0.06125183        5  0.00000000 0.03125151  0.000000 1.000000e+00
#> 41  0.20361734        5  0.07769223 0.06424869  1.209242 2.265698e-01
#> 42  0.19866728        5  0.13735859 0.03128052  4.391187 1.127336e-05
#> 43  0.05909092        5  0.00000000 0.03014898  0.000000 1.000000e+00
#> 44  0.14702432        5  0.00000000 0.07501379  0.000000 1.000000e+00
#> 45  0.09804159        5  0.03127301 0.03406623  0.918006 3.586157e-01
#>            2.5%
#> 1   0.607530630
#> 2   0.514786782
#> 3  -0.150947681
#> 4  -0.066365636
#> 5   0.535934031
#> 6   0.230444312
#> 7  -0.086434524
#> 8  -0.082427513
#> 9   0.446449155
#> 10  0.382298845
#> 11  0.647576437
#> 12  0.007722334
#> 13 -0.078395634
#> 14  0.261093683
#> 15  0.248690407
#> 16 -0.103708126
#> 17 -0.130900849
#> 18  0.187769671
#> 19  0.235686018
#> 20  0.597584186
#> 21  0.151496781
#> 22 -0.076079059
#> 23  0.096535451
#> 24  0.198253844
#> 25 -0.094187087
#> 26 -0.155831947
#> 27  0.066417023
#> 28  0.134758701
#> 29  0.479571304
#> 30  0.223429200
#> 31 -0.069652205
#> 32  0.002192406
#> 33  0.133888372
#> 34 -0.076718620
#> 35 -0.158946481
#> 36  0.001740919
#> 37  0.064444048
#> 38  0.348847121
#> 39  0.236744221
#> 40 -0.061251826
#> 41 -0.048232879
#> 42  0.076049907
#> 43 -0.059090919
#> 44 -0.147024324
#> 45 -0.035495583
confint(delta, level = 0.95)
#>         97.5 % interval        2.5 %
#> 1   0.79201437        1  0.607530630
#> 2   0.76880726        1  0.514786782
#> 3  -0.03876505        1 -0.150947681
#> 4   0.06636564        1 -0.066365636
#> 5   0.66385674        1  0.535934031
#> 6   0.36031867        1  0.230444312
#> 7   0.08643452        1 -0.086434524
#> 8   0.08242751        1 -0.082427513
#> 9   0.55369804        1  0.446449155
#> 10  0.59706425        2  0.382298845
#> 11  1.02066951        2  0.647576437
#> 12  0.14380151        2  0.007722334
#> 13  0.07839563        2 -0.078395634
#> 14  0.45865526        2  0.261093683
#> 15  0.40113055        2  0.248690407
#> 16  0.10370813        2 -0.103708126
#> 17  0.13090085        2 -0.130900849
#> 18  0.31237753        2  0.187769671
#> 19  0.44964534        3  0.235686018
#> 20  1.03174117        3  0.597584186
#> 21  0.32414644        3  0.151496781
#> 22  0.07607906        3 -0.076079059
#> 23  0.33523862        3  0.096535451
#> 24  0.33930497        3  0.198253844
#> 25  0.09418709        3 -0.094187087
#> 26  0.15583195        3 -0.155831947
#> 27  0.18369339        3  0.066417023
#> 28  0.34481734        4  0.134758701
#> 29  0.93769708        4  0.479571304
#> 30  0.43069193        4  0.223429200
#> 31  0.06965220        4 -0.069652205
#> 32  0.25682686        4  0.002192406
#> 33  0.26246867        4  0.133888372
#> 34  0.07671862        4 -0.076718620
#> 35  0.15894648        4 -0.158946481
#> 36  0.12333269        4  0.001740919
#> 37  0.27115007        5  0.064444048
#> 38  0.80915611        5  0.348847121
#> 39  0.46350849        5  0.236744221
#> 40  0.06125183        5 -0.061251826
#> 41  0.20361734        5 -0.048232879
#> 42  0.19866728        5  0.076049907
#> 43  0.05909092        5 -0.059090919
#> 44  0.14702432        5 -0.147024324
#> 45  0.09804159        5 -0.035495583
plot(delta)