Generate Random Drift Matrices Using the Monte Carlo Method

This function generates random drift matrices \(\boldsymbol{\Phi}\) using the Monte Carlo method.

Usage

MCPhi(phi, vcov_phi_vec, R, test_phi = TRUE, ncores = NULL, seed = NULL)

Arguments

phi

Numeric matrix. The drift matrix (\(\boldsymbol{\Phi}\)). phi should have row and column names pertaining to the variables in the system.

vcov_phi_vec

Numeric matrix. The sampling variance-covariance matrix of \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\).

R

Positive integer. Number of replications.

test_phi

Logical. If test_phi = TRUE, the function runs TestPhi() on the generated drift matrix \(\boldsymbol{\Phi}\). If the TestPhi() returns FALSE, the function generates a new drift matrix \(\boldsymbol{\Phi}\) and runs the test recursively until TestPhi() returns TRUE.

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of replications R is a large value.

seed

Random seed.

Value

Returns a list of simulated drift matrices.

Details

Monte Carlo Method

Let \(\boldsymbol{\theta}\) be \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\), that is, the elements of the \(\boldsymbol{\Phi}\) matrix in vector form sorted column-wise. Let \(\hat{\boldsymbol{\theta}}\) be \(\mathrm{vec} \left( \hat{\boldsymbol{\Phi}} \right)\). Based on the asymptotic properties of maximum likelihood estimators, we can assume that estimators are normally distributed around the population parameters. $$ \hat{\boldsymbol{\theta}} \sim \mathcal{N} \left( \boldsymbol{\theta}, \mathbb{V} \left( \hat{\boldsymbol{\theta}} \right) \right) $$ Using this distributional assumption, a sampling distribution of \(\hat{\boldsymbol{\theta}}\) which we refer to as \(\hat{\boldsymbol{\theta}}^{\ast}\) can be generated by replacing the population parameters with sample estimates, that is, $$ \hat{\boldsymbol{\theta}}^{\ast} \sim \mathcal{N} \left( \hat{\boldsymbol{\theta}}, \hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right) \right) . $$

Linear Stochastic Differential Equation Model

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t}, \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\mathbf{y}_{i, t}\), \(\boldsymbol{\eta}_{i, t}\), and \(\boldsymbol{\varepsilon}_{i, t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{i, t}\) represents a vector of observed random variables, \(\boldsymbol{\eta}_{i, t}\) a vector of latent random variables, and \(\boldsymbol{\varepsilon}_{i, t}\) a vector of random measurement errors, at time \(t\) and individual \(i\). \(\boldsymbol{\nu}\) denotes a vector of intercepts, \(\boldsymbol{\Lambda}\) a matrix of factor loadings, and \(\boldsymbol{\Theta}\) the covariance matrix of \(\boldsymbol{\varepsilon}\).

An alternative representation of the measurement error is given by $$ \boldsymbol{\varepsilon}_{i, t} = \boldsymbol{\Theta}^{\frac{1}{2}} \mathbf{z}_{i, t}, \quad \mathrm{with} \quad \mathbf{z}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \mathbf{I} \right) $$ where \(\mathbf{z}_{i, t}\) is a vector of independent standard normal random variables and \( \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right) \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right)^{\prime} = \boldsymbol{\Theta} . \)

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \left( \boldsymbol{\iota} + \boldsymbol{\Phi} \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$ where \(\boldsymbol{\iota}\) is a term which is unobserved and constant over time, \(\boldsymbol{\Phi}\) is the drift matrix which represents the rate of change of the solution in the absence of any random fluctuations, \(\boldsymbol{\Sigma}\) is the matrix of volatility or randomness in the process, and \(\mathrm{d}\boldsymbol{W}\) is a Wiener process or Brownian motion, which represents random fluctuations.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(9),
  R = 100L # use a large value for R in actual research
)
#> [[1]]
#>         x       m       y
#> x -0.0760 -0.4862  0.5124
#> m  0.0567 -0.5421 -0.3605
#> y -0.6663  0.7630 -0.2764
#> 
#> [[2]]
#>         x       m       y
#> x -0.3763 -0.2982  0.4505
#> m  1.3634 -0.2192 -0.1485
#> y -0.6588  0.9190 -1.2884
#> 
#> [[3]]
#>         x       m       y
#> x -0.4850  0.1231 -0.0557
#> m  0.4696 -0.6395 -0.0242
#> y -0.4942  0.4707 -1.0609
#> 
#> [[4]]
#>         x       m       y
#> x -0.6006 -0.3019  0.2921
#> m  1.2003 -0.4252 -0.1033
#> y  0.1619  0.0476 -1.1901
#> 
#> [[5]]
#>         x       m       y
#> x -0.5262 -0.2212 -0.2565
#> m  0.9227 -0.7390  0.1162
#> y -0.2691  0.3604 -0.5651
#> 
#> [[6]]
#>         x       m       y
#> x -0.4711 -0.0115  0.2527
#> m  0.3314 -0.7492 -0.4056
#> y -0.6492  0.5207 -1.0941
#> 
#> [[7]]
#>         x       m       y
#> x -0.5214  0.1565  0.2549
#> m  0.9763 -0.7219  0.0511
#> y  0.1858  0.6129 -1.3985
#> 
#> [[8]]
#>         x       m       y
#> x -0.4391  0.1602  0.3140
#> m  0.9760 -0.4755 -0.0345
#> y  0.0321  0.1359 -0.8687
#> 
#> [[9]]
#>         x       m       y
#> x -0.0748 -0.1272  0.2834
#> m  0.4057 -0.3778  0.6357
#> y -0.9118  0.6011 -1.0473
#> 
#> [[10]]
#>         x       m       y
#> x -0.4348 -0.1729  0.0064
#> m  0.6073 -0.7849 -0.1063
#> y -0.4893  0.2763 -0.7855
#> 
#> [[11]]
#>         x       m       y
#> x -0.6792  0.0588  0.3546
#> m  0.9055 -0.5302  0.0559
#> y -0.0558  0.7199 -0.9038
#> 
#> [[12]]
#>         x       m       y
#> x -0.5706  0.5133 -0.2524
#> m  0.2214 -0.7430  0.7783
#> y -0.0700  0.4227 -0.5470
#> 
#> [[13]]
#>         x       m       y
#> x -0.0807 -0.3228 -0.3278
#> m  1.3262 -0.4005  0.3093
#> y -0.6630  0.6763 -0.9570
#> 
#> [[14]]
#>         x       m       y
#> x -0.9077 -0.3329  0.2023
#> m  0.8398 -0.3269 -0.0567
#> y -0.2031  0.4089 -0.8382
#> 
#> [[15]]
#>         x       m       y
#> x -0.6272 -0.0692  0.2282
#> m  1.0970 -0.7986 -0.2687
#> y -0.3245  0.6999 -0.9664
#> 
#> [[16]]
#>         x       m       y
#> x -0.4389  0.0381  0.4637
#> m  0.2230 -0.6555  0.0034
#> y -0.5628  0.8697 -0.4950
#> 
#> [[17]]
#>         x       m       y
#> x -0.0543 -0.4824 -0.1335
#> m  0.8400 -0.4073  0.3872
#> y -0.4829  0.5947 -0.7367
#> 
#> [[18]]
#>         x       m       y
#> x -0.2738 -0.0609 -0.0883
#> m  0.8278 -0.7235 -0.0455
#> y -0.7862  0.5349 -0.7620
#> 
#> [[19]]
#>         x       m       y
#> x -0.9074  0.2631 -0.0412
#> m  0.4722 -1.0775  0.2562
#> y -0.0558  0.8079 -0.9916
#> 
#> [[20]]
#>         x       m       y
#> x -0.5909 -0.2019  0.1427
#> m  1.0399 -0.7779  0.3103
#> y -0.8074  0.3883 -0.3136
#> 
#> [[21]]
#>         x       m       y
#> x -0.7848  0.0313 -0.4615
#> m  0.4934 -0.4426  0.2998
#> y -0.2495  1.0250 -1.1428
#> 
#> [[22]]
#>         x       m       y
#> x -0.8018 -0.5704  0.0329
#> m  1.0348 -0.9235  0.3845
#> y -0.7037  1.2415 -0.4322
#> 
#> [[23]]
#>         x       m       y
#> x -0.2281 -0.2938  0.0001
#> m  0.6231 -0.8344  0.0630
#> y -1.1451  0.9438 -0.6006
#> 
#> [[24]]
#>         x       m       y
#> x -0.9412 -0.1682 -0.1885
#> m  0.6984 -0.5485  0.1097
#> y -0.3340  1.1849 -0.7935
#> 
#> [[25]]
#>         x       m       y
#> x -0.8638  0.2775 -0.0527
#> m  0.8174 -0.7943 -0.2610
#> y -0.2942  0.7943 -1.1121
#> 
#> [[26]]
#>         x       m       y
#> x -0.2640 -0.8483  0.0682
#> m  0.5336 -0.1617 -0.0268
#> y -0.6184  0.4031 -0.8170
#> 
#> [[27]]
#>         x       m       y
#> x -0.1246 -0.4074  0.2188
#> m  0.8197 -0.5434 -0.0706
#> y -0.3232  0.6011 -0.5412
#> 
#> [[28]]
#>         x       m       y
#> x -0.2204  0.0710 -0.3837
#> m  0.7401 -0.7955 -0.0066
#> y -0.6826  1.1953 -0.8408
#> 
#> [[29]]
#>         x       m       y
#> x -0.3920 -0.2335 -0.0759
#> m  0.6293 -0.5354 -0.0789
#> y -0.3671  0.1221 -0.8238
#> 
#> [[30]]
#>         x       m       y
#> x -0.5097  0.0261  0.2407
#> m  0.4611 -0.8364  0.0237
#> y -0.9475  0.9784 -0.3313
#> 
#> [[31]]
#>         x       m       y
#> x -0.3285 -0.3038 -0.0189
#> m  0.8849 -1.0224  0.1026
#> y -0.5892  0.0759 -0.7218
#> 
#> [[32]]
#>         x       m       y
#> x -0.3532  0.4504  0.6227
#> m  0.5017 -0.9511  0.8448
#> y -0.4372  0.8194 -0.8388
#> 
#> [[33]]
#>         x       m       y
#> x -0.5509 -0.4350 -0.3038
#> m  1.3369 -0.4285  0.1684
#> y -0.5887  0.4109 -0.5725
#> 
#> [[34]]
#>         x       m       y
#> x -0.6701  0.0880  0.0296
#> m  0.9804 -0.0583 -0.2594
#> y -0.2403  0.6601 -0.3565
#> 
#> [[35]]
#>         x       m       y
#> x -0.5199  0.1511 -0.1155
#> m  0.7510 -0.8170 -0.0285
#> y -0.6244  0.3641 -0.2383
#> 
#> [[36]]
#>         x       m       y
#> x -0.5749 -0.3671  0.0427
#> m  1.0524 -0.2760 -0.1973
#> y -0.6568  0.7831 -0.9197
#> 
#> [[37]]
#>         x       m       y
#> x -0.8423 -0.1387  0.1696
#> m  0.4624 -0.8768  0.1500
#> y -0.3318  1.5993 -0.9462
#> 
#> [[38]]
#>         x       m       y
#> x -0.5150 -0.1889  0.0468
#> m  0.4370 -0.2799 -0.0201
#> y -0.2641  1.1021 -0.8907
#> 
#> [[39]]
#>         x       m       y
#> x -0.4983 -0.0373  0.4961
#> m  0.6042 -0.8376 -0.1911
#> y -0.6549  0.6904 -0.7765
#> 
#> [[40]]
#>         x       m       y
#> x -0.2599 -0.3627 -0.9200
#> m  0.4429 -0.3375  0.6920
#> y -0.0975  0.7651 -1.2381
#> 
#> [[41]]
#>         x       m       y
#> x -0.4103 -0.0509 -0.1326
#> m  0.6537 -0.4068 -0.5623
#> y  0.0255  0.7562 -0.8038
#> 
#> [[42]]
#>         x       m       y
#> x -0.6199 -0.0061 -0.1077
#> m  1.1081 -0.4645  0.1069
#> y -1.3638  0.3986 -0.9301
#> 
#> [[43]]
#>         x       m       y
#> x -0.5259  0.6106 -0.6260
#> m  0.9630 -0.9045 -0.6314
#> y -0.0071  0.4031 -0.8791
#> 
#> [[44]]
#>         x       m       y
#> x -0.0782 -0.2313  0.6672
#> m  0.8591 -0.5688  0.1983
#> y -0.5510  0.4824 -0.7528
#> 
#> [[45]]
#>         x       m       y
#> x -0.2759 -0.1524  0.0048
#> m  0.6326 -1.1501 -0.3458
#> y -0.3071  0.5839 -0.6811
#> 
#> [[46]]
#>         x       m       y
#> x -0.3473 -0.2020 -0.3020
#> m  0.7495 -0.7286 -0.3584
#> y -0.3480  0.5229 -1.0091
#> 
#> [[47]]
#>         x       m       y
#> x -1.1203 -0.3872  0.0003
#> m  0.7209 -0.4724  0.1819
#> y -1.0782  1.0150 -0.6808
#> 
#> [[48]]
#>         x       m       y
#> x -0.5090 -0.0410 -0.3107
#> m  1.0340 -0.1081 -0.2667
#> y -0.5768  0.9607 -1.0205
#> 
#> [[49]]
#>         x       m       y
#> x -0.3721 -0.0977 -0.1090
#> m  0.2274 -0.8759 -0.0955
#> y -0.8434  0.7656 -0.1588
#> 
#> [[50]]
#>         x       m       y
#> x -0.1471  0.0207 -0.3629
#> m  0.6750 -1.0159 -0.1092
#> y -0.4941  1.4109 -0.4628
#> 
#> [[51]]
#>         x       m       y
#> x -0.5352 -0.2625  0.1291
#> m  0.7398 -0.4409  0.0751
#> y -0.4159  0.8846 -0.7308
#> 
#> [[52]]
#>         x       m       y
#> x -0.4928 -0.2565  0.1287
#> m  0.2505 -0.4536 -0.0459
#> y -0.9598  1.0094 -0.7691
#> 
#> [[53]]
#>         x       m       y
#> x -0.5283 -0.4697 -0.1946
#> m  0.5602 -0.6929  0.3783
#> y -0.5180  0.9683 -0.4484
#> 
#> [[54]]
#>         x       m       y
#> x -0.4568 -0.2649 -0.6771
#> m  0.4294 -0.6605 -0.2647
#> y -0.1390  0.4209 -0.9345
#> 
#> [[55]]
#>         x       m       y
#> x -0.6912 -0.0731  0.0679
#> m  0.6340 -0.3666 -0.1454
#> y -0.5207  0.4137 -0.1958
#> 
#> [[56]]
#>         x       m       y
#> x -0.1600 -0.3663  0.4777
#> m  1.0687 -0.0267 -0.0422
#> y -0.8129  0.5192 -0.9898
#> 
#> [[57]]
#>         x       m       y
#> x -0.4898 -0.3337 -0.3006
#> m  1.3280 -0.5775  0.1096
#> y -0.5941  0.1938 -0.3480
#> 
#> [[58]]
#>         x       m       y
#> x -0.4901 -0.6516  0.2534
#> m  0.8244 -0.4806  0.1500
#> y -0.3919  0.5778 -0.7430
#> 
#> [[59]]
#>         x       m       y
#> x -0.6873  0.3057 -0.3883
#> m  1.0725 -0.5698  0.0606
#> y -0.9457  0.5310 -1.1525
#> 
#> [[60]]
#>         x       m       y
#> x -0.3720  0.3435 -0.4295
#> m  1.1643 -1.0275  0.0208
#> y -0.4396  0.7747 -1.0855
#> 
#> [[61]]
#>         x       m       y
#> x -0.2976  0.0067  0.0649
#> m  0.0507 -0.4751  0.1708
#> y -0.2387  0.5225 -1.0189
#> 
#> [[62]]
#>         x       m       y
#> x -0.2483 -0.1431 -0.2055
#> m  0.7113 -1.0393  0.3029
#> y -0.2097  0.8985 -0.7370
#> 
#> [[63]]
#>         x       m       y
#> x -0.6067  0.3246 -0.1533
#> m  0.9970 -1.0205  0.1420
#> y -0.2641  0.6972 -1.0631
#> 
#> [[64]]
#>         x       m       y
#> x -0.2586  0.0143  0.2905
#> m  0.6174 -0.8401 -0.3931
#> y -0.5609  0.1311 -0.8925
#> 
#> [[65]]
#>         x       m       y
#> x -0.6577  0.5154 -0.0694
#> m  1.0424 -0.4520 -0.6430
#> y -0.3562  0.7949 -0.7043
#> 
#> [[66]]
#>         x       m       y
#> x -0.7562  0.4940 -0.5617
#> m  0.1389 -1.0324  0.1706
#> y -0.3597  0.8123 -0.8420
#> 
#> [[67]]
#>         x       m       y
#> x -0.5527  0.4049 -0.2545
#> m  0.9771 -0.9074 -0.1453
#> y -0.3193  0.6812 -1.5210
#> 
#> [[68]]
#>         x       m       y
#> x -0.6285 -0.1003 -0.0815
#> m  0.3604 -0.0809  0.1642
#> y -0.2456  0.5074 -0.6107
#> 
#> [[69]]
#>         x       m       y
#> x -0.5013  0.4953  0.4625
#> m  0.9026 -0.5453  0.1820
#> y -0.9528  0.2263 -0.6401
#> 
#> [[70]]
#>         x       m       y
#> x -0.3892  0.0050  0.5327
#> m  0.6615 -0.2819 -0.3928
#> y -0.1055  0.7149 -1.3148
#> 
#> [[71]]
#>         x       m       y
#> x -0.4809 -0.1947  0.2503
#> m  0.0677 -0.7323 -0.3383
#> y -0.3704  0.8351 -1.0399
#> 
#> [[72]]
#>         x       m       y
#> x -0.4019  0.1198 -0.0706
#> m  0.9825 -0.2584  0.1087
#> y -0.0331  0.8138 -0.7420
#> 
#> [[73]]
#>         x       m       y
#> x -0.6176  0.1362 -0.0417
#> m  0.6394 -0.6405  0.0416
#> y -0.1130  0.4741 -0.2869
#> 
#> [[74]]
#>         x       m       y
#> x -0.2472 -0.0723  0.5290
#> m  0.9725 -0.3605  0.3614
#> y -0.8785  0.0914 -0.7131
#> 
#> [[75]]
#>         x       m       y
#> x -0.4021 -0.6945 -0.4206
#> m  0.5264 -0.8066 -0.1381
#> y -0.5404  0.8423 -0.2368
#> 
#> [[76]]
#>         x       m       y
#> x -0.1604 -0.0821 -0.0733
#> m  0.8413 -0.9684  0.2162
#> y -0.3881  0.5588 -0.9423
#> 
#> [[77]]
#>         x       m       y
#> x -0.5729 -0.0446  0.0622
#> m  0.4789 -0.8515  0.2045
#> y -0.7156  1.0018 -0.6161
#> 
#> [[78]]
#>         x       m       y
#> x -0.5267  0.0776  0.2530
#> m  1.0610 -0.7165 -0.0581
#> y -0.5207  0.6180 -0.8570
#> 
#> [[79]]
#>         x       m       y
#> x -0.4672  0.2470 -0.2379
#> m  0.5415 -0.5113 -0.0580
#> y -0.2024  0.6926 -0.5558
#> 
#> [[80]]
#>         x       m       y
#> x -0.2230 -0.0151 -0.5356
#> m  0.9608 -0.6024 -0.0230
#> y -0.6082  1.2926 -0.8014
#> 
#> [[81]]
#>         x       m       y
#> x -0.1020 -0.4329 -0.0785
#> m  1.0110 -0.2399 -0.2657
#> y -0.3593  0.5771 -0.7099
#> 
#> [[82]]
#>         x       m       y
#> x -0.4645 -0.2353  0.2163
#> m  0.4183 -0.6466  0.4196
#> y -0.5220  1.0451 -1.1392
#> 
#> [[83]]
#>         x       m       y
#> x -0.3555 -0.1099  0.0578
#> m  0.6104 -0.7135 -0.0670
#> y -0.2520  1.0587 -1.1014
#> 
#> [[84]]
#>         x       m       y
#> x -0.7825  0.2066  0.1108
#> m  0.8524 -0.2320 -0.3256
#> y -0.0374  0.4381 -0.5199
#> 
#> [[85]]
#>         x       m       y
#> x -0.4374 -0.7443 -0.0376
#> m  0.5663 -1.1042  0.3078
#> y -0.1892  0.6988 -1.1493
#> 
#> [[86]]
#>         x       m       y
#> x -0.4695 -0.2393 -0.1125
#> m  1.0630 -0.6656  0.2291
#> y -0.2762  0.6862 -0.8820
#> 
#> [[87]]
#>         x       m       y
#> x -0.1130 -0.1603 -0.4277
#> m  1.2055 -0.5050 -0.4344
#> y -0.4888  0.8244 -0.6906
#> 
#> [[88]]
#>         x       m       y
#> x -0.7055  0.0290 -0.1808
#> m  0.8513 -0.6802  0.4804
#> y -0.3514  0.4804 -0.6106
#> 
#> [[89]]
#>         x       m       y
#> x -0.1752  0.0185  0.2668
#> m  0.5738 -0.6456 -0.0268
#> y -0.7892  1.0631 -0.5795
#> 
#> [[90]]
#>         x       m       y
#> x -0.2628 -0.0399  0.0463
#> m  0.6589 -0.4640 -0.2417
#> y -0.7183  1.1576 -1.0098
#> 
#> [[91]]
#>         x       m       y
#> x -0.1569  0.1362 -0.1585
#> m  0.5004 -0.8936  0.0031
#> y -0.2710  0.2579 -0.6716
#> 
#> [[92]]
#>         x       m       y
#> x -0.1727  0.0829 -0.2946
#> m  0.5516 -0.8779 -0.1840
#> y -0.4574  0.6886 -1.2225
#> 
#> [[93]]
#>         x       m       y
#> x -0.2914 -0.2958  0.0513
#> m  1.0339 -0.3858 -0.0674
#> y -0.9905  0.7895 -1.2285
#> 
#> [[94]]
#>         x       m       y
#> x -0.3575 -0.0768  0.0838
#> m  0.5043 -0.1732 -0.2195
#> y -0.8807  0.9582 -0.5528
#> 
#> [[95]]
#>         x       m       y
#> x -0.5319 -0.1651 -0.0300
#> m  1.0110 -0.9016 -0.0997
#> y -0.3932  0.7935 -0.8232
#> 
#> [[96]]
#>         x       m       y
#> x -0.1050  0.0211 -0.1646
#> m  1.3785 -0.2052  0.0743
#> y  0.3112  0.1185 -0.7787
#> 
#> [[97]]
#>         x       m       y
#> x -0.6516  0.0080 -0.2044
#> m  1.1186 -0.9496  0.0270
#> y -0.2543  1.2535 -0.9854
#> 
#> [[98]]
#>         x       m       y
#> x -0.7151 -0.3095  0.2745
#> m  1.3058 -0.7629 -0.0194
#> y -0.8651  0.4425 -0.5278
#> 
#> [[99]]
#>         x       m       y
#> x -0.5241 -0.2879 -0.2188
#> m  0.4010 -0.0715 -0.0294
#> y -0.2946  0.5394 -0.8664
#> 
#> [[100]]
#>         x       m       y
#> x -0.3205  0.1786  0.0711
#> m  0.9488 -0.7668 -0.3865
#> y -0.2852  1.0044 -0.8033
#> 
phi <- matrix(
  data = c(
    -6, 5.5, 0, 0,
    1.25, -2.5, 5.9, -7.3,
    0, 0, -6, 2.5,
    5, 0, 0, -6
  ),
  nrow = 4
)
colnames(phi) <- rownames(phi) <- paste0("y", 1:4)
MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(16),
  R = 100L, # use a large value for R in actual research
  test_phi = FALSE
)
#> [[1]]
#>         y1      y2      y3      y4
#> y1 -6.0997  1.0927 -0.1029  5.4688
#> y2  5.6853 -3.0368 -0.0993 -0.1365
#> y3 -0.1116  6.0650 -6.4239 -0.0780
#> y4 -0.0734 -6.6092  2.7201 -5.5802
#> 
#> [[2]]
#>         y1      y2      y3      y4
#> y1 -5.8152  0.6049 -0.0480  5.0575
#> y2  5.5586 -2.5129 -0.8266  0.2388
#> y3 -0.1761  6.0698 -6.1346 -0.0960
#> y4 -0.4429 -7.2737  2.3384 -5.8010
#> 
#> [[3]]
#>         y1      y2      y3      y4
#> y1 -5.6721  1.1614 -0.1763  5.1728
#> y2  5.6622 -2.6888 -0.2719  0.6681
#> y3 -0.1906  5.9744 -5.5126  0.3930
#> y4  0.3494 -6.8521  3.1167 -6.0866
#> 
#> [[4]]
#>         y1      y2      y3      y4
#> y1 -6.2642  1.9351 -0.4506  4.7755
#> y2  5.2752 -1.9571  0.2394 -0.1624
#> y3 -0.5053  5.4119 -6.2278 -0.1412
#> y4 -0.4084 -7.4423  2.9313 -5.9375
#> 
#> [[5]]
#>         y1      y2      y3      y4
#> y1 -6.8150  1.1136  0.0518  4.9864
#> y2  5.6247 -2.2161 -0.3849 -0.6516
#> y3  0.3508  5.5740 -5.8861 -0.0673
#> y4  0.0122 -7.1370  2.4462 -6.4040
#> 
#> [[6]]
#>         y1      y2      y3      y4
#> y1 -6.0924  1.3937  0.0802  4.9241
#> y2  5.7707 -2.8492  0.2187 -0.0093
#> y3  0.6596  6.0491 -6.2397  0.1290
#> y4 -0.2285 -7.5046  2.8187 -6.0275
#> 
#> [[7]]
#>         y1      y2      y3      y4
#> y1 -6.0641  1.0594 -0.0333  5.3468
#> y2  5.6145 -2.4197 -0.2903 -0.1088
#> y3  0.0226  5.8328 -5.9751 -0.0610
#> y4  0.1511 -7.3443  2.1919 -6.4119
#> 
#> [[8]]
#>         y1      y2      y3      y4
#> y1 -6.2889  0.9410 -0.0319  5.0616
#> y2  5.5146 -2.0537 -0.6578 -0.3167
#> y3  0.6534  5.7478 -6.6727  0.0247
#> y4  0.2322 -7.0470  2.7351 -6.0858
#> 
#> [[9]]
#>         y1      y2      y3      y4
#> y1 -6.0648  1.0318  0.0554  5.1603
#> y2  5.5887 -2.4747 -0.0318  0.2056
#> y3 -0.2009  5.9805 -6.3370  0.4053
#> y4 -0.2170 -8.4937  2.2833 -5.7778
#> 
#> [[10]]
#>         y1      y2      y3      y4
#> y1 -5.9315  1.3002  0.1036  5.0537
#> y2  4.9584 -2.8780  0.2888  0.1916
#> y3 -0.1987  6.1014 -5.5830  0.1359
#> y4 -0.2521 -7.5178  1.8048 -5.7113
#> 
#> [[11]]
#>         y1      y2      y3      y4
#> y1 -6.2379  1.2063  0.2395  4.9173
#> y2  5.5899 -3.1934  0.0984 -0.3894
#> y3  0.3426  6.0841 -6.0915  0.3078
#> y4 -0.1746 -7.0587  2.4372 -6.0825
#> 
#> [[12]]
#>         y1      y2      y3      y4
#> y1 -5.9535  1.4764 -0.3767  5.4662
#> y2  5.8784 -2.7233 -0.1695  0.3307
#> y3  0.0189  5.5947 -6.4067  0.1737
#> y4 -0.2975 -6.9923  2.5144 -6.1763
#> 
#> [[13]]
#>         y1      y2      y3      y4
#> y1 -6.0287  0.8460 -0.4575  4.5554
#> y2  5.1888 -2.8896 -0.3384 -0.3867
#> y3 -0.1245  5.5808 -6.4727  0.0147
#> y4  0.1603 -7.3845  2.5957 -5.8991
#> 
#> [[14]]
#>         y1      y2      y3      y4
#> y1 -5.8052  0.6481 -0.1823  4.9781
#> y2  5.6495 -2.1146  0.1288  0.3818
#> y3  0.8894  6.6391 -5.9254  0.2503
#> y4 -0.3611 -7.3983  2.1828 -5.8728
#> 
#> [[15]]
#>         y1      y2      y3      y4
#> y1 -6.5820  1.3779 -0.7276  5.1864
#> y2  5.4380 -2.4188  0.1020 -0.5026
#> y3  0.0925  5.3323 -5.0892  0.1034
#> y4  0.2750 -7.2416  2.3345 -6.9807
#> 
#> [[16]]
#>         y1      y2      y3      y4
#> y1 -5.8570  1.7348  0.1636  4.5807
#> y2  5.4063 -2.5068  0.0296 -0.2532
#> y3 -0.3781  5.8302 -6.0762  0.1290
#> y4  0.0228 -7.1498  2.4330 -5.4899
#> 
#> [[17]]
#>         y1      y2      y3      y4
#> y1 -6.3301  1.4364  0.0517  4.8440
#> y2  5.8876 -2.8413  0.5898  0.2683
#> y3 -0.1285  6.1896 -6.1736 -0.3813
#> y4  0.1932 -8.1752  2.5418 -6.2199
#> 
#> [[18]]
#>         y1      y2      y3      y4
#> y1 -5.9028  1.3935  0.6286  4.8288
#> y2  5.2794 -2.4119  0.5223 -0.6316
#> y3  0.2789  5.6189 -6.1380  0.5329
#> y4 -0.1880 -7.2506  1.8582 -5.3934
#> 
#> [[19]]
#>         y1      y2      y3      y4
#> y1 -5.4374  1.2176  0.1008  5.1189
#> y2  5.6336 -2.0007 -0.4681  0.3487
#> y3 -0.3560  5.8857 -6.0880 -0.9181
#> y4  0.5947 -7.3283  2.8384 -6.2293
#> 
#> [[20]]
#>         y1      y2      y3      y4
#> y1 -5.8152  1.2938 -0.1121  5.2663
#> y2  5.4596 -2.8260  0.2437 -0.1255
#> y3 -0.2047  5.8548 -5.7602  0.0360
#> y4  0.4655 -7.4409  2.6666 -6.4757
#> 
#> [[21]]
#>         y1      y2      y3      y4
#> y1 -6.0840  1.4193 -0.1022  4.7914
#> y2  4.9051 -3.1290 -0.0309  0.4231
#> y3 -0.3881  5.6833 -6.0397  0.0697
#> y4  0.5195 -6.9034  2.4382 -5.5571
#> 
#> [[22]]
#>         y1      y2      y3      y4
#> y1 -5.4077  0.8402 -0.4076  5.0556
#> y2  5.3960 -2.6949 -0.3074 -0.2453
#> y3 -0.6824  6.1535 -5.8886 -0.4302
#> y4 -0.0690 -7.3680  2.5095 -5.6829
#> 
#> [[23]]
#>         y1      y2      y3      y4
#> y1 -5.6834  1.6799  0.2975  4.9592
#> y2  5.4337 -2.5756  0.2467 -0.0241
#> y3 -0.3776  5.7957 -6.2889  0.7140
#> y4 -0.2704 -6.6793  2.3860 -6.2320
#> 
#> [[24]]
#>         y1      y2      y3      y4
#> y1 -5.6965  1.4480  0.0157  4.9911
#> y2  5.2268 -2.7581  0.4088  0.6689
#> y3  0.2561  5.4159 -5.7118  0.0965
#> y4  0.2735 -6.9110  2.7261 -5.8564
#> 
#> [[25]]
#>         y1      y2      y3      y4
#> y1 -5.7067  0.7881  0.1962  5.3325
#> y2  5.4345 -2.1865  0.2824 -0.3664
#> y3 -0.2378  6.0765 -5.7499 -0.6397
#> y4  0.6324 -7.4791  2.6196 -6.5689
#> 
#> [[26]]
#>         y1      y2      y3      y4
#> y1 -5.9069  1.4980 -0.0699  4.9648
#> y2  5.8660 -2.5110  0.4341 -0.3900
#> y3 -0.0858  6.1336 -6.2642 -0.0311
#> y4  0.4447 -7.0371  2.6757 -5.8999
#> 
#> [[27]]
#>         y1      y2      y3      y4
#> y1 -5.8617  1.1201  0.3485  4.6246
#> y2  5.2831 -3.1511  0.1032  0.5487
#> y3  0.2317  6.0255 -6.3656  0.5854
#> y4 -0.0111 -7.3643  2.7533 -6.1541
#> 
#> [[28]]
#>         y1      y2      y3      y4
#> y1 -6.1108  1.3582  0.1565  5.0835
#> y2  5.3395 -2.6267 -0.1702  0.3747
#> y3  0.0979  6.1896 -6.1561 -0.2778
#> y4 -0.5711 -7.1928  2.4219 -5.7843
#> 
#> [[29]]
#>         y1      y2      y3      y4
#> y1 -5.6807  1.0587 -0.2181  4.4869
#> y2  4.9665 -2.1921 -0.2006  0.2824
#> y3  0.4508  5.6466 -6.1440 -0.0612
#> y4 -0.3115 -7.4841  2.5836 -5.6916
#> 
#> [[30]]
#>         y1      y2      y3      y4
#> y1 -6.2164  0.9615 -0.0582  5.0372
#> y2  5.3911 -2.4714  0.0840 -0.1159
#> y3 -0.4069  6.3736 -5.8655  0.5325
#> y4 -0.6563 -7.2230  2.7239 -6.2866
#> 
#> [[31]]
#>         y1      y2      y3      y4
#> y1 -6.1128  1.5769 -0.2951  4.6318
#> y2  5.8222 -2.8338 -0.6979 -0.2153
#> y3 -0.0793  5.6227 -5.6304  0.2376
#> y4 -0.3564 -7.3523  2.4516 -6.0424
#> 
#> [[32]]
#>         y1      y2      y3      y4
#> y1 -5.6092  1.1742  0.0538  4.9519
#> y2  5.2888 -2.6336  0.1553  0.1049
#> y3  0.2190  5.8874 -6.7839  0.0053
#> y4  0.9235 -7.3801  1.6709 -6.0198
#> 
#> [[33]]
#>         y1      y2      y3      y4
#> y1 -6.1799  1.2682 -0.6805  4.7693
#> y2  5.4932 -2.9208  0.1448  0.0689
#> y3 -0.1182  5.7104 -6.1867 -0.4115
#> y4  0.1085 -7.2829  2.4017 -5.9838
#> 
#> [[34]]
#>         y1      y2      y3      y4
#> y1 -5.5267  1.6497  0.2116  5.0133
#> y2  5.6084 -2.1901  0.1423  0.3552
#> y3  0.0041  5.8114 -6.4979  0.2069
#> y4 -0.2085 -7.4465  2.9355 -5.8620
#> 
#> [[35]]
#>         y1      y2      y3      y4
#> y1 -6.1651  1.7174 -0.2021  4.8136
#> y2  5.4490 -3.0233  0.1331  0.0589
#> y3  0.0419  6.2731 -5.8123  0.2847
#> y4  0.3476 -6.4457  3.0014 -6.0736
#> 
#> [[36]]
#>         y1      y2      y3      y4
#> y1 -6.2275  1.1616 -0.3508  5.0662
#> y2  4.9689 -2.0972 -0.1180  0.1902
#> y3 -0.1371  6.1843 -6.6684  0.2731
#> y4  0.0669 -7.0663  2.3468 -6.0330
#> 
#> [[37]]
#>         y1      y2      y3      y4
#> y1 -5.6033  1.2881  0.0003  4.5359
#> y2  5.5017 -2.4047 -0.1765  0.2554
#> y3  0.0473  5.6318 -5.8369  0.1554
#> y4 -0.4280 -7.4869  3.0594 -6.4491
#> 
#> [[38]]
#>         y1      y2      y3      y4
#> y1 -5.8253  1.3295  0.4899  5.0968
#> y2  5.4044 -2.9757  0.9210 -0.1226
#> y3 -0.0456  6.3288 -6.0151  0.6198
#> y4  0.0293 -7.3327  2.1425 -6.9133
#> 
#> [[39]]
#>         y1      y2      y3      y4
#> y1 -5.6473  1.1000 -0.0851  4.5797
#> y2  5.2361 -3.4854 -0.0701 -0.2389
#> y3 -0.3917  5.5814 -6.0723  0.2432
#> y4 -0.5658 -7.7778  2.1400 -6.5042
#> 
#> [[40]]
#>         y1      y2      y3      y4
#> y1 -5.5872  1.6594 -0.2911  5.1557
#> y2  5.6460 -2.6393 -0.0151 -0.2549
#> y3 -0.1432  6.2997 -5.2242  0.1154
#> y4 -0.6034 -7.8621  2.7011 -5.7661
#> 
#> [[41]]
#>         y1      y2      y3      y4
#> y1 -6.2063  1.2478 -0.0529  5.1313
#> y2  5.6878 -2.3809  0.0091  0.1898
#> y3 -0.2578  5.9843 -6.0915 -0.6805
#> y4  0.5546 -7.5342  2.6409 -5.4350
#> 
#> [[42]]
#>         y1      y2      y3      y4
#> y1 -6.2387  0.8233 -0.2057  5.0903
#> y2  5.9402 -2.5120 -0.0633 -0.1668
#> y3  0.1026  5.6023 -4.9505 -0.0025
#> y4  0.0799 -7.9813  1.9399 -6.5894
#> 
#> [[43]]
#>         y1      y2      y3      y4
#> y1 -6.4656  1.2757  0.6892  5.1458
#> y2  5.6410 -2.9090  0.2741  0.3812
#> y3  0.4236  6.3745 -6.2330 -0.0743
#> y4 -0.3654 -7.0417  2.8766 -6.4089
#> 
#> [[44]]
#>         y1      y2      y3      y4
#> y1 -5.4365  1.1521 -0.1312  5.0104
#> y2  5.8832 -2.9827  0.5871 -0.3701
#> y3 -0.0782  6.0257 -6.0998  0.1321
#> y4 -0.2361 -7.6680  3.2670 -6.5515
#> 
#> [[45]]
#>         y1      y2      y3      y4
#> y1 -5.9422  1.7718 -0.2249  5.2263
#> y2  5.3443 -2.2811  0.5396  0.0353
#> y3 -0.7622  6.2748 -5.9860 -0.3980
#> y4 -0.1266 -7.4529  2.7097 -5.6488
#> 
#> [[46]]
#>         y1      y2      y3      y4
#> y1 -5.5994  1.4771 -0.5030  4.8100
#> y2  5.7767 -2.0613  0.1368  0.2438
#> y3  0.2048  6.6772 -5.9552 -0.0133
#> y4 -0.2031 -6.7421  1.8768 -5.7137
#> 
#> [[47]]
#>         y1      y2      y3      y4
#> y1 -6.3290  1.7312  0.1506  5.0015
#> y2  5.5164 -2.7571  0.1539 -0.6438
#> y3  0.2056  5.9514 -5.5600  0.1617
#> y4  0.0587 -7.1376  2.0163 -5.8523
#> 
#> [[48]]
#>         y1      y2      y3      y4
#> y1 -6.0645  1.5163 -0.0990  5.0213
#> y2  5.8543 -2.1826 -0.0942 -0.3138
#> y3  0.0169  6.3865 -6.0874  0.1238
#> y4  0.5178 -7.3387  2.7016 -5.9439
#> 
#> [[49]]
#>         y1      y2      y3      y4
#> y1 -5.4281  0.9421  0.2490  4.9954
#> y2  5.0225 -2.2258  0.3041 -0.6366
#> y3  0.2454  5.6938 -6.0781  0.4004
#> y4 -0.5632 -7.6266  2.5159 -6.3946
#> 
#> [[50]]
#>         y1      y2      y3      y4
#> y1 -5.8623  1.5475 -0.1445  4.6926
#> y2  5.7678 -2.4154 -0.6402 -0.2043
#> y3 -0.0508  5.7345 -5.4943  0.0434
#> y4  0.5152 -6.9688  2.8029 -5.8728
#> 
#> [[51]]
#>         y1      y2      y3      y4
#> y1 -5.7053  0.7847  0.2968  4.8373
#> y2  6.0308 -2.5309  0.0258  0.5789
#> y3  0.3155  6.0711 -5.8783 -0.1103
#> y4 -0.2869 -7.3461  2.2289 -5.8124
#> 
#> [[52]]
#>         y1      y2      y3      y4
#> y1 -5.9596  1.7185  0.2252  5.0912
#> y2  5.0676 -2.3494  0.2235 -0.0243
#> y3 -0.4731  5.5698 -5.8907  0.0151
#> y4 -0.4517 -7.0866  2.5139 -5.7784
#> 
#> [[53]]
#>         y1      y2      y3      y4
#> y1 -5.9538  1.5540 -0.2166  4.6659
#> y2  5.3072 -2.5572  0.5932  0.2492
#> y3  0.1448  5.7408 -5.7262  0.2501
#> y4 -0.2943 -7.3943  2.3154 -6.1222
#> 
#> [[54]]
#>         y1      y2      y3      y4
#> y1 -6.1906  1.5693 -0.0493  5.0879
#> y2  5.3991 -3.2319  0.5825  0.1935
#> y3  0.1497  5.7183 -5.9713  0.2144
#> y4 -0.2377 -7.3921  2.6696 -6.2544
#> 
#> [[55]]
#>         y1      y2      y3      y4
#> y1 -5.7037  1.4463 -0.1787  4.9228
#> y2  5.3629 -2.8630 -0.2951  0.3896
#> y3  0.6043  5.5064 -5.6901 -0.4693
#> y4  0.0524 -7.3310  2.3554 -6.1661
#> 
#> [[56]]
#>         y1      y2      y3      y4
#> y1 -5.9499  0.8204  0.3580  4.8365
#> y2  5.4434 -2.3140 -0.3683  0.1815
#> y3 -0.2425  6.1816 -5.8524 -0.1818
#> y4  0.6021 -7.5627  2.8055 -6.3042
#> 
#> [[57]]
#>         y1      y2      y3      y4
#> y1 -5.4811  0.8284 -0.1755  5.1875
#> y2  5.2720 -3.0185 -0.2874  0.0744
#> y3  0.0485  5.8303 -5.9354 -0.0117
#> y4  0.0294 -7.2583  2.4088 -6.4980
#> 
#> [[58]]
#>         y1      y2      y3      y4
#> y1 -6.2979  0.8135 -0.3412  4.8067
#> y2  5.3626 -2.8106  0.2273  0.2806
#> y3  0.0792  5.4888 -6.1236 -0.2509
#> y4  0.4159 -7.3767  2.6538 -5.5200
#> 
#> [[59]]
#>         y1      y2      y3      y4
#> y1 -6.6244  1.4046 -0.2259  4.9115
#> y2  5.4229 -2.5284 -0.3666  0.1156
#> y3 -0.3609  5.8220 -6.0951 -0.1030
#> y4  0.1138 -6.7139  2.4940 -6.1402
#> 
#> [[60]]
#>         y1      y2      y3      y4
#> y1 -5.8463  1.2323 -0.1842  4.7883
#> y2  5.6793 -2.6646  0.4244 -0.0538
#> y3 -0.2186  6.3690 -6.3722  0.2561
#> y4 -0.2139 -7.7736  2.3217 -5.8240
#> 
#> [[61]]
#>         y1      y2      y3      y4
#> y1 -6.5559  0.7982 -0.2189  4.9218
#> y2  5.4756 -2.5471 -0.5988 -0.0666
#> y3  0.1045  5.6999 -6.2377 -0.0124
#> y4 -0.1745 -7.4226  3.0092 -6.2215
#> 
#> [[62]]
#>         y1      y2      y3      y4
#> y1 -5.3726  1.2277  0.1510  4.2108
#> y2  5.4980 -2.3666  0.4830  0.2065
#> y3 -0.1006  5.7155 -6.2552 -0.4508
#> y4 -0.6863 -7.2331  2.4594 -5.8112
#> 
#> [[63]]
#>         y1      y2      y3      y4
#> y1 -6.2017  1.3462  0.0826  5.0226
#> y2  5.1786 -1.9849 -0.0752 -0.3715
#> y3 -0.4124  6.2253 -6.3165 -0.7154
#> y4  0.4178 -7.1263  2.5544 -6.3468
#> 
#> [[64]]
#>         y1      y2      y3      y4
#> y1 -5.6526  1.6195  0.3325  4.9124
#> y2  5.3647 -2.7584  0.4914 -0.2989
#> y3  0.0898  5.5830 -5.7311 -0.0665
#> y4  0.6028 -7.0933  2.5407 -5.4274
#> 
#> [[65]]
#>         y1      y2      y3      y4
#> y1 -6.1217  0.7183 -0.1762  5.1117
#> y2  4.9675 -2.5841 -0.0772  0.3194
#> y3  0.4443  5.4497 -5.7449  0.3024
#> y4  0.2342 -6.5684  2.1264 -5.4135
#> 
#> [[66]]
#>         y1      y2      y3      y4
#> y1 -6.3206  1.3321  0.5565  5.3275
#> y2  5.8768 -2.7903  0.4065 -0.5802
#> y3  0.3235  5.9049 -6.1163 -0.6650
#> y4  0.2063 -6.7797  1.9795 -6.2447
#> 
#> [[67]]
#>         y1      y2      y3      y4
#> y1 -5.8986  1.3053 -0.2026  4.8342
#> y2  5.4708 -2.2583 -0.6361 -0.2348
#> y3 -0.1413  5.9735 -5.8545  0.2243
#> y4  0.4620 -6.6713  2.9248 -6.2491
#> 
#> [[68]]
#>         y1      y2      y3      y4
#> y1 -6.4409  1.3229 -0.0498  5.3828
#> y2  5.6003 -1.8849 -0.5953 -0.0817
#> y3 -0.3207  5.7942 -6.1938  0.2928
#> y4 -0.4500 -7.1085  2.5946 -6.1377
#> 
#> [[69]]
#>         y1      y2      y3      y4
#> y1 -6.0988  1.3684 -0.0272  5.0182
#> y2  5.5109 -2.9592 -0.2258  0.0944
#> y3 -0.0166  5.4399 -5.7348 -0.0208
#> y4  0.0667 -7.0733  2.0007 -6.0209
#> 
#> [[70]]
#>         y1      y2      y3      y4
#> y1 -5.6166  1.4904  0.5344  4.5116
#> y2  5.8408 -2.2396 -0.1459  0.2414
#> y3 -0.0621  6.0437 -6.4148  0.1029
#> y4 -0.1057 -7.2250  2.4703 -6.0267
#> 
#> [[71]]
#>         y1      y2      y3      y4
#> y1 -6.5304  1.3178 -0.0535  4.9499
#> y2  4.8130 -2.6807  0.2546  0.2766
#> y3 -0.4895  6.0657 -6.0537 -0.2924
#> y4  0.8743 -6.9249  2.2902 -5.8972
#> 
#> [[72]]
#>         y1      y2      y3      y4
#> y1 -6.1534  1.2487  0.4461  4.5050
#> y2  5.8576 -2.0318 -0.0697 -0.0641
#> y3  0.2685  5.7050 -5.7529 -0.4780
#> y4 -0.4503 -7.3260  2.4252 -6.3866
#> 
#> [[73]]
#>         y1      y2      y3      y4
#> y1 -6.2259  1.0925  0.2764  5.5451
#> y2  5.3550 -2.1596 -0.1744  0.1471
#> y3  0.1951  6.3393 -6.5934  0.4193
#> y4  0.2169 -7.2437  2.6231 -6.2507
#> 
#> [[74]]
#>         y1      y2      y3      y4
#> y1 -5.4144  1.1511  0.6656  5.2468
#> y2  5.3224 -2.4808 -0.1223  0.1994
#> y3  0.2452  5.7263 -6.2848  0.0911
#> y4 -0.3390 -7.9290  2.5044 -5.9216
#> 
#> [[75]]
#>         y1      y2      y3      y4
#> y1 -5.6012  1.8250  0.3497  5.1050
#> y2  5.9040 -2.2550 -0.0914 -0.6015
#> y3 -0.9142  5.5944 -6.2113 -0.0088
#> y4 -0.2755 -7.8592  2.5255 -5.6521
#> 
#> [[76]]
#>         y1      y2      y3      y4
#> y1 -5.9582  1.6150 -0.0596  5.4303
#> y2  5.9060 -2.9016  0.0978 -0.5010
#> y3  0.5238  6.0212 -5.7341 -0.1566
#> y4 -0.2522 -6.4831  2.6610 -6.2851
#> 
#> [[77]]
#>         y1      y2      y3      y4
#> y1 -5.4426  1.4279  0.0200  4.8141
#> y2  5.3213 -1.6559  0.5359 -0.0848
#> y3 -0.2739  5.6102 -6.4222 -0.3977
#> y4 -0.1151 -7.3838  1.9690 -6.0075
#> 
#> [[78]]
#>         y1      y2      y3      y4
#> y1 -5.7942  0.9761  0.2630  5.2993
#> y2  5.1951 -2.7612  0.1717 -0.2288
#> y3  0.4195  6.2718 -6.1608  0.0290
#> y4  0.1422 -6.8398  2.3432 -5.6912
#> 
#> [[79]]
#>         y1      y2      y3      y4
#> y1 -6.1877  1.0751 -0.1613  5.1155
#> y2  5.8869 -1.9172  0.5968  0.1230
#> y3  0.2801  6.0880 -6.0138 -0.3703
#> y4 -0.6950 -6.6348  2.3827 -5.3860
#> 
#> [[80]]
#>         y1      y2      y3      y4
#> y1 -5.7343  1.2459 -0.0424  5.1297
#> y2  5.6613 -2.3566  0.2391 -0.3112
#> y3 -0.2107  6.0896 -6.3066 -0.3796
#> y4 -0.5535 -7.3601  2.3918 -6.0191
#> 
#> [[81]]
#>         y1      y2      y3      y4
#> y1 -5.9654  1.0258 -0.0137  5.5688
#> y2  4.7611 -1.6726  0.0512  0.4613
#> y3 -0.4651  5.7228 -6.2287  0.1832
#> y4  0.2934 -7.2297  2.8684 -6.5139
#> 
#> [[82]]
#>         y1      y2      y3      y4
#> y1 -5.6882  1.3016 -0.4055  4.7428
#> y2  5.1391 -1.8846  0.0901 -0.3278
#> y3 -0.0223  5.0893 -5.8812  0.0034
#> y4  0.0503 -7.5459  2.0234 -6.2138
#> 
#> [[83]]
#>         y1      y2      y3      y4
#> y1 -5.8832  1.6363  0.0928  4.4724
#> y2  5.1371 -2.0875  0.2051  0.0894
#> y3 -0.5903  6.1065 -6.3766 -0.0246
#> y4  0.5818 -7.1165  2.6681 -6.4179
#> 
#> [[84]]
#>         y1      y2      y3      y4
#> y1 -5.9788  1.3044  0.5668  5.0275
#> y2  5.6490 -2.0393  0.0361 -0.3363
#> y3 -0.4147  5.9258 -5.5690 -0.1944
#> y4  0.7438 -7.4655  2.3636 -5.2432
#> 
#> [[85]]
#>         y1      y2      y3      y4
#> y1 -5.7419  1.2359  0.2319  5.3380
#> y2  5.8770 -2.8768  0.4138  0.3085
#> y3  0.0172  5.7582 -5.9257 -0.2019
#> y4  0.3323 -7.5563  2.6923 -6.0586
#> 
#> [[86]]
#>         y1      y2      y3      y4
#> y1 -5.6880  1.1592  0.1822  4.7133
#> y2  5.3728 -2.5775 -0.1229  0.1272
#> y3  0.1762  6.0386 -5.8511  0.2912
#> y4 -0.4010 -7.5123  2.7845 -6.5188
#> 
#> [[87]]
#>         y1      y2      y3      y4
#> y1 -6.0748  1.2978 -0.0702  4.7877
#> y2  5.6033 -2.5373  0.0522  0.4053
#> y3 -0.3451  5.1697 -6.3795 -0.0337
#> y4  0.1374 -7.0948  2.3840 -5.4827
#> 
#> [[88]]
#>         y1      y2      y3      y4
#> y1 -5.8496  1.2972 -0.2201  5.6090
#> y2  5.6908 -2.4301 -0.4447  0.2021
#> y3  0.1033  6.1973 -5.7252  0.6939
#> y4  0.3207 -7.7084  2.7509 -6.1409
#> 
#> [[89]]
#>         y1      y2      y3      y4
#> y1 -6.0742  1.3067 -0.1897  5.4992
#> y2  5.9449 -2.3881  0.0419 -0.0328
#> y3 -0.4819  5.7928 -6.1451 -0.0449
#> y4  0.1397 -6.9045  2.2978 -6.0387
#> 
#> [[90]]
#>         y1      y2      y3      y4
#> y1 -6.3360  0.7287  0.2466  5.1962
#> y2  5.4727 -2.3153  0.2253 -0.0957
#> y3 -0.2429  6.1721 -6.0877  0.1552
#> y4 -0.6499 -7.1973  2.0810 -6.5093
#> 
#> [[91]]
#>         y1      y2      y3      y4
#> y1 -5.7552  1.6801  0.1259  5.3994
#> y2  5.4883 -1.7731  0.2389  0.1920
#> y3  0.3119  5.9617 -5.8849  0.6475
#> y4  0.4811 -7.4543  2.1976 -6.3466
#> 
#> [[92]]
#>         y1      y2      y3      y4
#> y1 -5.9906  0.9727 -0.3041  4.8068
#> y2  5.1321 -3.0073  0.3761 -0.4380
#> y3 -0.5641  5.7772 -5.9532  0.1420
#> y4 -0.5267 -6.9904  2.3159 -6.2561
#> 
#> [[93]]
#>         y1      y2      y3      y4
#> y1 -6.1676  1.0423 -0.2423  4.9600
#> y2  5.9755 -2.8257 -0.1146 -0.0580
#> y3  0.6835  6.1543 -6.3659 -0.2062
#> y4 -0.0246 -7.4579  1.7724 -6.2233
#> 
#> [[94]]
#>         y1      y2      y3      y4
#> y1 -5.9353  0.9291  0.5458  4.9101
#> y2  5.4142 -2.1021 -0.1633 -0.2393
#> y3 -0.1490  5.6737 -6.0880  0.1265
#> y4  0.2064 -7.0823  2.4896 -5.9967
#> 
#> [[95]]
#>         y1      y2      y3      y4
#> y1 -6.2494  1.1335  0.2114  5.0960
#> y2  5.3727 -2.3160 -0.2422  0.2281
#> y3 -0.3663  6.1367 -5.7912 -0.2753
#> y4 -0.1140 -7.5241  2.3097 -6.1357
#> 
#> [[96]]
#>         y1      y2      y3      y4
#> y1 -6.0408  0.7212  0.4191  5.1015
#> y2  4.9623 -2.9197  0.1384  0.1524
#> y3 -0.3109  6.0256 -5.2264 -0.0877
#> y4 -0.6755 -7.5530  2.6742 -5.9693
#> 
#> [[97]]
#>         y1      y2      y3      y4
#> y1 -5.4653  0.9621  0.1435  4.7095
#> y2  5.6993 -2.2634  0.3774 -0.0918
#> y3  0.0319  5.4337 -5.9728  0.3008
#> y4  0.6743 -7.3527  2.9092 -5.4830
#> 
#> [[98]]
#>         y1      y2      y3      y4
#> y1 -5.8295  0.8375  0.2623  4.6541
#> y2  5.3865 -2.8807  0.3489  0.0648
#> y3  0.1500  6.4184 -6.0842 -0.4846
#> y4 -0.0981 -6.5892  2.4590 -6.2817
#> 
#> [[99]]
#>         y1      y2      y3      y4
#> y1 -6.0242  0.9970 -0.1571  5.1305
#> y2  5.6378 -2.3416 -0.3824  0.0046
#> y3  0.2931  5.7132 -6.3844 -0.3101
#> y4 -0.1788 -6.9748  2.4890 -5.7396
#> 
#> [[100]]
#>         y1      y2      y3      y4
#> y1 -5.6479  1.0140  0.3110  4.8002
#> y2  5.4031 -2.4595  0.2715  0.0495
#> y3 -0.5033  5.9112 -5.5255 -0.6251
#> y4  0.0132 -7.4014  2.8056 -5.3418
#>