Generate Random Drift Matrices and Process Noise Covariance Matrices Using the Monte Carlo Method

This function generates random drift matrices \(\boldsymbol{\Phi}\) and process noise covariabces matrices \(\boldsymbol{\Sigma}\) using the Monte Carlo method.

Usage

MCPhiSigma(
  phi,
  sigma,
  vcov_theta,
  R,
  test_phi = TRUE,
  ncores = NULL,
  seed = NULL
)

Arguments

phi

Numeric matrix. The drift matrix (\(\boldsymbol{\Phi}\)). phi should have row and column names pertaining to the variables in the system.

sigma

Numeric matrix. The process noise covariance matrix (\(\boldsymbol{\Sigma}\)).

vcov_theta

Numeric matrix. The sampling variance-covariance matrix of \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\) and \(\mathrm{vech} \left( \boldsymbol{\Sigma} \right)\)

R

Positive integer. Number of replications.

test_phi

Logical. If test_phi = TRUE, the function tests the stability of the generated drift matrix \(\boldsymbol{\Phi}\). If the test returns FALSE, the function generates a new drift matrix \(\boldsymbol{\Phi}\) and runs the test recursively until the test returns TRUE.

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of replications R is a large value.

seed

Random seed.

Value

Returns an object of class ctmedmc which is a list with the following elements:

call

Function call.

args

Function arguments.

fun

Function used ("MCPhiSigma").

output

A list simulated drift matrices.

Details

Monte Carlo Method

Let \(\boldsymbol{\theta}\) be a vector that combines \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\), that is, the elements of the \(\boldsymbol{\Phi}\) matrix in vector form sorted column-wise and \(\mathrm{vech} \left( \boldsymbol{\Sigma} \right)\), that is, the unique elements of the \(\boldsymbol{\Sigma}\) matrix in vector form sorted column-wise. Let \(\hat{\boldsymbol{\theta}}\) be a vector that combines \(\mathrm{vec} \left( \hat{\boldsymbol{\Phi}} \right)\) and \(\mathrm{vech} \left( \hat{\boldsymbol{\Sigma}} \right)\). Based on the asymptotic properties of maximum likelihood estimators, we can assume that estimators are normally distributed around the population parameters. $$ \hat{\boldsymbol{\theta}} \sim \mathcal{N} \left( \boldsymbol{\theta}, \mathbb{V} \left( \hat{\boldsymbol{\theta}} \right) \right) $$ Using this distributional assumption, a sampling distribution of \(\hat{\boldsymbol{\theta}}\) which we refer to as \(\hat{\boldsymbol{\theta}}^{\ast}\) can be generated by replacing the population parameters with sample estimates, that is, $$ \hat{\boldsymbol{\theta}}^{\ast} \sim \mathcal{N} \left( \hat{\boldsymbol{\theta}}, \hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right) \right) . $$

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
sigma <- matrix(
  data = c(
    0.24455556, 0.02201587, -0.05004762,
    0.02201587, 0.07067800, 0.01539456,
    -0.05004762, 0.01539456, 0.07553061
  ),
  nrow = 3
)
MCPhiSigma(
  phi = phi,
  sigma = sigma,
  vcov_theta = 0.1 * diag(15),
  R = 100L # use a large value for R in actual research
)
#> [[1]]
#> [[1]][[1]]
#>         x       m       y
#> x -0.0957  0.0263 -0.1507
#> m  0.8653 -0.9254  0.4226
#> y -0.3456  0.3478 -0.5309
#> 
#> [[1]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2084 -0.3647 -0.0479
#> [2,] -0.3647  0.6456  0.1214
#> [3,] -0.0479  0.1214  0.2013
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#>         x       m       y
#> x -0.5278 -0.3347 -0.2763
#> m  1.0161 -0.9849  0.2003
#> y -0.2007  0.6946 -0.6920
#> 
#> [[2]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2782 -0.0617 -0.3256
#> [2,] -0.0617  0.0901  0.2183
#> [3,] -0.3256  0.2183  0.6601
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#>         x       m       y
#> x -0.4018 -0.1907 -0.0677
#> m  0.7344 -1.3231 -0.3393
#> y  0.0472  0.5430 -0.3117
#> 
#> [[3]][[2]]
#>        [,1]    [,2]    [,3]
#> [1,] 0.4038  0.1140  0.0883
#> [2,] 0.1140  0.1321 -0.1766
#> [3,] 0.0883 -0.1766  0.4256
#> 
#> 
#> [[4]]
#> [[4]][[1]]
#>         x       m       y
#> x -0.5798  0.1835  0.2893
#> m  0.6420 -0.4850  0.2752
#> y -0.6472  0.4132 -0.3708
#> 
#> [[4]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.3617  0.1826 -0.1071
#> [2,]  0.1826  0.0922 -0.0521
#> [3,] -0.1071 -0.0521  0.2382
#> 
#> 
#> [[5]]
#> [[5]][[1]]
#>         x       m       y
#> x -0.5308 -0.2971  0.4638
#> m  1.0700 -0.6486 -0.6680
#> y -0.6323  0.4139 -0.9234
#> 
#> [[5]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.6437  0.5209 -0.2458
#> [2,]  0.5209  0.4242 -0.1904
#> [3,] -0.2458 -0.1904  0.1198
#> 
#> 
#> [[6]]
#> [[6]][[1]]
#>         x       m       y
#> x -0.5323  0.0733  0.3680
#> m  0.4414 -0.8219  0.6145
#> y -0.9988  0.7608 -0.5748
#> 
#> [[6]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.0384 -0.0341  0.0446
#> [2,] -0.0341  0.1952 -0.2344
#> [3,]  0.0446 -0.2344  0.2820
#> 
#> 
#> [[7]]
#> [[7]][[1]]
#>         x       m       y
#> x -0.4964 -0.3485 -0.5275
#> m  0.8803 -0.2652 -0.2808
#> y -0.4947  0.6248 -0.8066
#> 
#> [[7]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.0219  0.0470 -0.0699
#> [2,]  0.0470  0.1010 -0.1503
#> [3,] -0.0699 -0.1503  0.2237
#> 
#> 
#> [[8]]
#> [[8]][[1]]
#>         x       m       y
#> x -0.3729  0.1268 -0.0831
#> m  0.7244 -1.0035 -0.2035
#> y -0.2491  0.4263 -0.2417
#> 
#> [[8]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.7350 -0.1220 -0.5629
#> [2,] -0.1220  0.0203  0.0935
#> [3,] -0.5629  0.0935  0.4311
#> 
#> 
#> [[9]]
#> [[9]][[1]]
#>         x      m       y
#> x -0.7018 -0.287 -0.4060
#> m  0.4014 -0.926 -0.6191
#> y -0.4522  1.042 -0.9174
#> 
#> [[9]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.1100 0.0805 0.1586
#> [2,] 0.0805 0.1239 0.2240
#> [3,] 0.1586 0.2240 0.4082
#> 
#> 
#> [[10]]
#> [[10]][[1]]
#>         x       m       y
#> x -0.7130 -0.0718  0.2803
#> m  0.6654 -0.3968 -0.1850
#> y  0.1915  0.1108 -0.7506
#> 
#> [[10]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2025  0.2538 -0.0539
#> [2,]  0.2538  0.3182 -0.0676
#> [3,] -0.0539 -0.0676  0.0143
#> 
#> 
#> [[11]]
#> [[11]][[1]]
#>         x       m       y
#> x -0.3922 -0.0144 -0.1370
#> m  0.5602 -0.4864 -0.6544
#> y -0.7721  0.4842 -0.0734
#> 
#> [[11]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.6277 0.1112 0.4736
#> [2,] 0.1112 0.1715 0.0026
#> [3,] 0.4736 0.0026 0.4009
#> 
#> 
#> [[12]]
#> [[12]][[1]]
#>         x       m       y
#> x -0.6543  0.1989 -0.5933
#> m  0.7583 -0.6301 -0.1023
#> y -0.3791  1.0368 -0.8090
#> 
#> [[12]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.3449  0.0936 -0.1936
#> [2,]  0.0936  0.1728 -0.3292
#> [3,] -0.1936 -0.3292  0.6279
#> 
#> 
#> [[13]]
#> [[13]][[1]]
#>         x       m       y
#> x -0.6089  0.2607 -0.2650
#> m  0.2459 -0.5695  0.0963
#> y -0.8200  0.4991 -0.8746
#> 
#> [[13]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.0530 -0.1447  0.1748
#> [2,] -0.1447  0.3952 -0.4775
#> [3,]  0.1748 -0.4775  0.5768
#> 
#> 
#> [[14]]
#> [[14]][[1]]
#>         x       m       y
#> x -0.9407  0.4237  0.1673
#> m  1.0778 -0.5615 -0.1271
#> y -0.1173  0.1810 -0.3245
#> 
#> [[14]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4544 -0.0291 -0.2844
#> [2,] -0.0291  0.0261  0.0182
#> [3,] -0.2844  0.0182  0.1780
#> 
#> 
#> [[15]]
#> [[15]][[1]]
#>         x       m       y
#> x -0.7717 -0.2800  0.1728
#> m  0.4799 -0.4767 -0.3760
#> y -0.1823  1.2392 -1.0738
#> 
#> [[15]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.5964 0.4158 0.3092
#> [2,] 0.4158 0.3037 0.1452
#> [3,] 0.3092 0.1452 0.5176
#> 
#> 
#> [[16]]
#> [[16]][[1]]
#>         x       m       y
#> x -0.6313  0.1458  0.5649
#> m  1.0035 -1.0179 -0.0652
#> y -0.7079  0.5710 -0.9275
#> 
#> [[16]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.0950 -0.2015 -0.1425
#> [2,] -0.2015  0.4275  0.3022
#> [3,] -0.1425  0.3022  0.2137
#> 
#> 
#> [[17]]
#> [[17]][[1]]
#>         x       m       y
#> x -0.6305  0.0519 -0.3075
#> m  0.2891 -0.8919 -0.1079
#> y -0.5822  0.8868 -0.8725
#> 
#> [[17]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.6199 -0.4343 -0.4270
#> [2,] -0.4343  0.3042  0.2992
#> [3,] -0.4270  0.2992  0.2942
#> 
#> 
#> [[18]]
#> [[18]][[1]]
#>         x       m       y
#> x -0.4866 -0.1255  0.3668
#> m  0.3933 -0.4558  0.3564
#> y -0.6396  0.3644 -0.4484
#> 
#> [[18]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4905 -0.4313  0.2472
#> [2,] -0.4313  0.3793 -0.2173
#> [3,]  0.2472 -0.2173  0.1246
#> 
#> 
#> [[19]]
#> [[19]][[1]]
#>         x       m       y
#> x -0.2808  0.0457  0.0345
#> m  0.6865 -0.4855 -0.4860
#> y -0.8859  0.3376 -1.2683
#> 
#> [[19]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.4990 0.0736 0.3997
#> [2,] 0.0736 0.5316 0.1545
#> [3,] 0.3997 0.1545 0.3377
#> 
#> 
#> [[20]]
#> [[20]][[1]]
#>         x       m       y
#> x -0.7559 -0.1536  0.4511
#> m  0.9942 -0.9598  0.5648
#> y -0.4914  0.5642 -1.0631
#> 
#> [[20]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.5334 0.1472 0.1451
#> [2,] 0.1472 0.0407 0.0434
#> [3,] 0.1451 0.0434 0.1142
#> 
#> 
#> [[21]]
#> [[21]][[1]]
#>         x       m       y
#> x -0.7730 -0.1430 -0.5482
#> m  0.4954 -0.4709 -0.1641
#> y -0.3138  1.2306 -0.6674
#> 
#> [[21]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.1942 0.0945 0.1136
#> [2,] 0.0945 0.1199 0.0997
#> [3,] 0.1136 0.0997 0.0931
#> 
#> 
#> [[22]]
#> [[22]][[1]]
#>         x       m       y
#> x -0.4945 -0.6671 -0.3027
#> m  1.0400 -0.4402  0.5274
#> y -0.5167  0.8127 -1.0204
#> 
#> [[22]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.2571 0.2225 0.0012
#> [2,] 0.2225 0.2074 0.0620
#> [3,] 0.0012 0.0620 0.2486
#> 
#> 
#> [[23]]
#> [[23]][[1]]
#>         x       m       y
#> x -0.2055 -0.1266 -0.0666
#> m  0.9118 -0.0650  0.2347
#> y -0.9983  0.2717 -0.7472
#> 
#> [[23]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.2463 0.2005 0.1557
#> [2,] 0.2005 0.1994 0.0751
#> [3,] 0.1557 0.0751 0.1725
#> 
#> 
#> [[24]]
#> [[24]][[1]]
#>         x       m       y
#> x -0.5935 -0.3221  0.2496
#> m  0.3853 -0.8539  0.2454
#> y -0.4149  0.2561 -0.7975
#> 
#> [[24]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.2061 0.0352 0.0933
#> [2,] 0.0352 0.0060 0.0159
#> [3,] 0.0933 0.0159 0.0422
#> 
#> 
#> [[25]]
#> [[25]][[1]]
#>         x       m       y
#> x -0.3634 -0.0708 -0.3003
#> m  0.9753 -0.7200  1.1534
#> y -0.4025  0.9655 -0.8761
#> 
#> [[25]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.3494  0.3209 -0.1330
#> [2,]  0.3209  0.2948 -0.1221
#> [3,] -0.1330 -0.1221  0.0506
#> 
#> 
#> [[26]]
#> [[26]][[1]]
#>         x       m       y
#> x -0.8219 -0.2116 -0.0700
#> m  0.4884 -0.5324  0.4498
#> y -0.3297  0.5871 -0.7476
#> 
#> [[26]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.0223 -0.0503  0.0619
#> [2,] -0.0503  0.1531 -0.1182
#> [3,]  0.0619 -0.1182  0.1828
#> 
#> 
#> [[27]]
#> [[27]][[1]]
#>         x       m       y
#> x -0.3332 -0.0772 -0.0003
#> m  0.5535 -0.0950 -0.1349
#> y -0.4507  0.7802 -0.5585
#> 
#> [[27]][[2]]
#>         [,1]    [,2]   [,3]
#> [1,]  0.8069 -0.0618 0.1021
#> [2,] -0.0618  0.0854 0.1171
#> [3,]  0.1021  0.1171 0.2064
#> 
#> 
#> [[28]]
#> [[28]][[1]]
#>         x       m       y
#> x -0.4629 -0.1020 -0.3401
#> m  0.8775 -0.2102 -0.4043
#> y -0.6210  0.5264 -0.9892
#> 
#> [[28]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1603 -0.0459 -0.2213
#> [2,] -0.0459  0.0131  0.0633
#> [3,] -0.2213  0.0633  0.3055
#> 
#> 
#> [[29]]
#> [[29]][[1]]
#>         x       m       y
#> x -0.3671 -0.0507  0.2011
#> m  0.4971 -0.9741 -0.2671
#> y -0.3027  1.1691 -0.6000
#> 
#> [[29]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.6322  0.1248 -0.1469
#> [2,]  0.1248  0.0247 -0.0290
#> [3,] -0.1469 -0.0290  0.0342
#> 
#> 
#> [[30]]
#> [[30]][[1]]
#>         x       m       y
#> x -0.2002  0.1459 -0.2178
#> m  0.1059 -0.4572 -0.0207
#> y -0.2826  1.2068 -0.9261
#> 
#> [[30]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  1.0116 -0.0162 -0.1021
#> [2,] -0.0162  0.0003  0.0016
#> [3,] -0.1021  0.0016  0.0103
#> 
#> 
#> [[31]]
#> [[31]][[1]]
#>         x       m       y
#> x -0.5716 -0.2904 -0.9815
#> m  0.6002 -0.5104  0.3351
#> y -0.5145  0.9666 -0.5926
#> 
#> [[31]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.5471  0.1585 -0.4529
#> [2,]  0.1585  0.0459 -0.1312
#> [3,] -0.4529 -0.1312  0.3749
#> 
#> 
#> [[32]]
#> [[32]][[1]]
#>         x       m       y
#> x -0.7341 -0.0446 -0.5146
#> m  1.0279 -0.0788 -0.2571
#> y -0.6199  0.7966 -0.7640
#> 
#> [[32]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4086  0.2762 -0.4548
#> [2,]  0.2762  0.1867 -0.3074
#> [3,] -0.4548 -0.3074  0.5061
#> 
#> 
#> [[33]]
#> [[33]][[1]]
#>         x       m       y
#> x -0.0936 -0.1392  0.3097
#> m  1.0521 -0.4798 -0.4324
#> y -0.5027  0.6471 -0.7498
#> 
#> [[33]][[2]]
#>        [,1]    [,2]    [,3]
#> [1,] 0.5477  0.2456  0.1058
#> [2,] 0.2456  0.3927 -0.0594
#> [3,] 0.1058 -0.0594  0.0609
#> 
#> 
#> [[34]]
#> [[34]][[1]]
#>         x       m       y
#> x -0.0528 -0.0732 -0.4617
#> m  0.8469 -0.8229 -0.0929
#> y -0.2442  1.0004 -0.3120
#> 
#> [[34]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.5940 -0.1519  0.1558
#> [2,] -0.1519  0.0918 -0.1974
#> [3,]  0.1558 -0.1974  0.5098
#> 
#> 
#> [[35]]
#> [[35]][[1]]
#>         x       m       y
#> x -0.6377  0.0118 -0.2127
#> m  0.4611 -0.0850  0.0130
#> y -0.4858  0.9554 -0.8258
#> 
#> [[35]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1715  0.1234 -0.1656
#> [2,]  0.1234  0.4647 -0.5007
#> [3,] -0.1656 -0.5007  0.5472
#> 
#> 
#> [[36]]
#> [[36]][[1]]
#>         x       m       y
#> x -0.3556 -0.0205 -0.0255
#> m  1.1865 -0.1943 -0.0506
#> y -0.7327  0.6756 -1.0600
#> 
#> [[36]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.7062 -0.3264 -0.0720
#> [2,] -0.3264  0.3677 -0.2076
#> [3,] -0.0720 -0.2076  0.2748
#> 
#> 
#> [[37]]
#> [[37]][[1]]
#>         x       m       y
#> x -0.4298 -0.0444 -0.4327
#> m  0.8583 -0.2174  0.1232
#> y -0.5062  0.4359 -0.9625
#> 
#> [[37]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2513 -0.0211  0.1160
#> [2,] -0.0211  0.0018 -0.0098
#> [3,]  0.1160 -0.0098  0.0536
#> 
#> 
#> [[38]]
#> [[38]][[1]]
#>         x       m       y
#> x -0.7014  0.3487 -0.0053
#> m  0.7319 -1.2244  0.1098
#> y -0.5596  0.4242 -0.8731
#> 
#> [[38]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.0633 0.0811 0.0452
#> [2,] 0.0811 0.1040 0.0579
#> [3,] 0.0452 0.0579 0.0323
#> 
#> 
#> [[39]]
#> [[39]][[1]]
#>         x       m       y
#> x -0.5995 -0.4198 -0.5094
#> m  0.6576 -0.2874  0.3968
#> y -0.2405  0.7774 -0.6126
#> 
#> [[39]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.5414 0.1288 0.2492
#> [2,] 0.1288 0.3167 0.0554
#> [3,] 0.2492 0.0554 0.5313
#> 
#> 
#> [[40]]
#> [[40]][[1]]
#>         x       m       y
#> x -0.8816 -0.1537  0.1903
#> m  0.6856 -0.5012  0.0072
#> y -0.1474  1.2228 -1.0138
#> 
#> [[40]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2999 -0.0386 -0.0075
#> [2,] -0.0386  0.3005 -0.0819
#> [3,] -0.0075 -0.0819  0.0234
#> 
#> 
#> [[41]]
#> [[41]][[1]]
#>         x       m       y
#> x -0.3246 -0.2059 -0.0913
#> m  0.9513 -0.5399 -0.5873
#> y  0.0404  1.3811 -0.5884
#> 
#> [[41]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.7772 0.2049 0.1479
#> [2,] 0.2049 0.0884 0.0603
#> [3,] 0.1479 0.0603 0.0414
#> 
#> 
#> [[42]]
#> [[42]][[1]]
#>         x       m       y
#> x -0.4740  0.4396 -0.0832
#> m  0.6544 -0.7612  0.3754
#> y -0.2096  0.5643 -0.9836
#> 
#> [[42]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.4298 0.2421 0.1416
#> [2,] 0.2421 0.1725 0.2533
#> [3,] 0.1416 0.2533 0.8802
#> 
#> 
#> [[43]]
#> [[43]][[1]]
#>         x       m       y
#> x -0.1173 -0.0902  0.1875
#> m  1.1948 -1.0243 -0.2334
#> y -0.7911  1.4457 -0.8459
#> 
#> [[43]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4451  0.3015 -0.2017
#> [2,]  0.3015  0.2045 -0.1427
#> [3,] -0.2017 -0.1427  0.2367
#> 
#> 
#> [[44]]
#> [[44]][[1]]
#>         x       m       y
#> x -0.4710 -0.3980  0.3124
#> m  0.4508 -0.4456 -0.1643
#> y -0.3165  0.3697 -0.7265
#> 
#> [[44]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4850  0.4859 -0.0714
#> [2,]  0.4859  0.4867 -0.0716
#> [3,] -0.0714 -0.0716  0.0105
#> 
#> 
#> [[45]]
#> [[45]][[1]]
#>         x       m       y
#> x -0.4309 -0.1042 -0.0629
#> m  0.7116 -0.6306  0.1055
#> y -0.4704  0.3523 -0.9833
#> 
#> [[45]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.3105  0.0843 -0.3497
#> [2,]  0.0843  0.1367 -0.0448
#> [3,] -0.3497 -0.0448  0.4159
#> 
#> 
#> [[46]]
#> [[46]][[1]]
#>         x       m       y
#> x -0.0051 -0.2148 -0.0709
#> m  1.0329 -0.1408  0.0182
#> y -0.4525  0.5266 -0.7625
#> 
#> [[46]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2581  0.1793 -0.3177
#> [2,]  0.1793  0.1246 -0.2207
#> [3,] -0.3177 -0.2207  0.3910
#> 
#> 
#> [[47]]
#> [[47]][[1]]
#>         x       m       y
#> x -0.4616 -0.0725  0.0428
#> m  0.6098 -0.6282  0.5300
#> y -0.7799  0.4186 -0.9579
#> 
#> [[47]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1560 -0.0425  0.0423
#> [2,] -0.0425  0.3610 -0.2309
#> [3,]  0.0423 -0.2309  0.6432
#> 
#> 
#> [[48]]
#> [[48]][[1]]
#>         x       m       y
#> x -0.1502 -0.0957 -0.0033
#> m  0.3819 -0.4073 -0.2677
#> y  0.3205  1.1344 -0.9660
#> 
#> [[48]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.9879 -0.1351  0.1604
#> [2,] -0.1351  0.1350 -0.0800
#> [3,]  0.1604 -0.0800  0.0550
#> 
#> 
#> [[49]]
#> [[49]][[1]]
#>         x       m       y
#> x -0.3805  0.0326  0.2706
#> m  0.6879 -0.8394 -0.3294
#> y -0.3198  0.7781 -0.3086
#> 
#> [[49]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1539  0.2687 -0.1492
#> [2,]  0.2687  0.6435 -0.4119
#> [3,] -0.1492 -0.4119  0.2760
#> 
#> 
#> [[50]]
#> [[50]][[1]]
#>         x       m       y
#> x -0.2143  0.0397 -0.3903
#> m  0.4640 -0.6397  0.1573
#> y -0.3312  0.7080 -0.4495
#> 
#> [[50]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.3515 -0.2139 -0.1041
#> [2,] -0.2139  0.3212 -0.1573
#> [3,] -0.1041 -0.1573  0.2857
#> 
#> 
#> [[51]]
#> [[51]][[1]]
#>         x       m       y
#> x -0.2475 -0.4155  0.3497
#> m  0.6953 -0.4973  0.1498
#> y -0.3532  0.7291 -0.4281
#> 
#> [[51]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4008  0.1479 -0.1788
#> [2,]  0.1479  0.0546 -0.0660
#> [3,] -0.1788 -0.0660  0.0797
#> 
#> 
#> [[52]]
#> [[52]][[1]]
#>         x       m       y
#> x -0.1458  0.1409 -0.0531
#> m  0.9106 -0.7036 -0.0348
#> y -0.2370  0.8806 -0.7353
#> 
#> [[52]][[2]]
#>         [,1]   [,2]    [,3]
#> [1,]  0.3600 0.0878 -0.0557
#> [2,]  0.0878 0.1881  0.2124
#> [3,] -0.0557 0.2124  0.3151
#> 
#> 
#> [[53]]
#> [[53]][[1]]
#>         x       m       y
#> x -0.2222 -0.1310  0.2233
#> m  1.0857 -0.6299  0.5893
#> y -0.7770  0.5307 -1.0305
#> 
#> [[53]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.6087 -0.4546  0.1737
#> [2,] -0.4546  0.3573 -0.2092
#> [3,]  0.1737 -0.2092  0.4051
#> 
#> 
#> [[54]]
#> [[54]][[1]]
#>         x       m       y
#> x -0.8887  0.2168 -0.1064
#> m  0.5653 -0.6396 -0.0758
#> y -0.3042  0.6632 -0.6685
#> 
#> [[54]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1005 -0.1568 -0.1958
#> [2,] -0.1568  0.2448  0.3056
#> [3,] -0.1958  0.3056  0.3816
#> 
#> 
#> [[55]]
#> [[55]][[1]]
#>         x       m       y
#> x -0.9374  0.4206  0.3723
#> m  0.5856 -0.6921  0.2199
#> y -0.1164  0.9093 -0.9678
#> 
#> [[55]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4113 -0.2727  0.0610
#> [2,] -0.2727  0.2883 -0.2670
#> [3,]  0.0610 -0.2670  0.4866
#> 
#> 
#> [[56]]
#> [[56]][[1]]
#>         x       m       y
#> x -0.2503 -0.5146 -0.1520
#> m  1.1227 -0.6609  0.5791
#> y -0.3941  0.9865 -0.4976
#> 
#> [[56]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.7958 -0.5671 -0.3135
#> [2,] -0.5671  0.4042  0.2234
#> [3,] -0.3135  0.2234  0.1235
#> 
#> 
#> [[57]]
#> [[57]][[1]]
#>         x       m       y
#> x -0.6878 -0.0961 -0.3349
#> m  0.9391 -0.0706 -0.5326
#> y -0.6234  0.9463 -0.9343
#> 
#> [[57]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4845 -0.1940 -0.2539
#> [2,] -0.1940  0.7550  0.0008
#> [3,] -0.2539  0.0008  0.6943
#> 
#> 
#> [[58]]
#> [[58]][[1]]
#>         x       m       y
#> x -0.3363 -0.3818 -0.1050
#> m  1.0383 -0.5745 -0.3098
#> y -0.7873  0.7268 -0.6687
#> 
#> [[58]][[2]]
#>        [,1]    [,2]    [,3]
#> [1,] 0.6970  0.0902  0.1216
#> [2,] 0.0902  0.0818 -0.1256
#> [3,] 0.1216 -0.1256  0.3059
#> 
#> 
#> [[59]]
#> [[59]][[1]]
#>         x       m       y
#> x -0.6026 -0.4554  0.3668
#> m  0.9719 -0.9901 -0.1138
#> y -0.3521  1.1071 -0.6564
#> 
#> [[59]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.6601 -0.1795  0.2528
#> [2,] -0.1795  0.0515 -0.0812
#> [3,]  0.2528 -0.0812  0.1553
#> 
#> 
#> [[60]]
#> [[60]][[1]]
#>         x       m       y
#> x -0.3440  0.1419 -0.0562
#> m  1.0771 -0.7116 -0.0177
#> y -0.4075  0.2834 -1.0193
#> 
#> [[60]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2470 -0.2132 -0.1865
#> [2,] -0.2132  0.5927 -0.0580
#> [3,] -0.1865 -0.0580  0.2582
#> 
#> 
#> [[61]]
#> [[61]][[1]]
#>         x       m       y
#> x -0.3492  0.5789  0.0045
#> m  0.3345 -0.5409 -0.1746
#> y -0.2112  0.8523 -0.5644
#> 
#> [[61]][[2]]
#>         [,1]   [,2]    [,3]
#> [1,]  0.6503 0.1099 -0.3057
#> [2,]  0.1099 0.0724  0.0418
#> [3,] -0.3057 0.0418  0.3061
#> 
#> 
#> [[62]]
#> [[62]][[1]]
#>         x       m       y
#> x -0.1217 -0.1273 -0.2099
#> m  1.0016 -0.3975  0.1838
#> y -0.0019  1.1542 -0.9609
#> 
#> [[62]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.3483 0.1607 0.1890
#> [2,] 0.1607 0.2717 0.1950
#> [3,] 0.1890 0.1950 0.1614
#> 
#> 
#> [[63]]
#> [[63]][[1]]
#>         x       m       y
#> x -0.6779  0.1738 -0.0312
#> m  0.8548 -0.4255 -0.4178
#> y -0.2483  0.4674 -0.7161
#> 
#> [[63]][[2]]
#>         [,1]   [,2]    [,3]
#> [1,]  0.3584 0.1173 -0.0433
#> [2,]  0.1173 0.1364  0.2003
#> [3,] -0.0433 0.2003  0.7328
#> 
#> 
#> [[64]]
#> [[64]][[1]]
#>         x       m       y
#> x -0.3198 -0.3037  0.3106
#> m  0.8710 -0.7809 -0.0708
#> y -0.0486  0.2357 -1.0363
#> 
#> [[64]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.6845 -0.2377 -0.5361
#> [2,] -0.2377  0.2679  0.0118
#> [3,] -0.5361  0.0118  0.5839
#> 
#> 
#> [[65]]
#> [[65]][[1]]
#>         x       m       y
#> x -0.2864  0.3173 -0.1932
#> m  0.3938 -0.3615 -0.0835
#> y -0.6604  0.9228 -0.6961
#> 
#> [[65]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1552 -0.1702  0.1947
#> [2,] -0.1702  0.1876 -0.2055
#> [3,]  0.1947 -0.2055  0.3110
#> 
#> 
#> [[66]]
#> [[66]][[1]]
#>         x       m       y
#> x -0.5653  0.5082  0.1769
#> m  0.2996 -1.0707  0.0119
#> y -0.4249  0.4022 -1.1536
#> 
#> [[66]][[2]]
#>        [,1]    [,2]    [,3]
#> [1,] 0.6602  0.0318  0.2916
#> [2,] 0.0318  0.3417 -0.0354
#> [3,] 0.2916 -0.0354  0.1360
#> 
#> 
#> [[67]]
#> [[67]][[1]]
#>         x       m       y
#> x -0.8375 -0.3296  0.2027
#> m  1.2818 -0.8500  0.0494
#> y -0.2475  0.4389 -1.2989
#> 
#> [[67]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4990 -0.1841 -0.1650
#> [2,] -0.1841  0.0816  0.0937
#> [3,] -0.1650  0.0937  0.1334
#> 
#> 
#> [[68]]
#> [[68]][[1]]
#>         x       m       y
#> x -0.3728 -0.3627 -0.0973
#> m  0.8306 -0.7333 -0.1693
#> y -0.3713  1.4143 -0.4007
#> 
#> [[68]][[2]]
#>         [,1]   [,2]    [,3]
#> [1,]  0.8069 0.0429 -0.5716
#> [2,]  0.0429 0.3306  0.0608
#> [3,] -0.5716 0.0608  0.7985
#> 
#> 
#> [[69]]
#> [[69]][[1]]
#>         x       m       y
#> x -0.1259 -0.6083  0.3736
#> m  0.9839 -0.9559  0.4978
#> y -0.3993  0.6415 -0.4283
#> 
#> [[69]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.7396 -0.2474 -0.0459
#> [2,] -0.2474  0.3590 -0.2160
#> [3,] -0.0459 -0.2160  0.1966
#> 
#> 
#> [[70]]
#> [[70]][[1]]
#>         x       m       y
#> x -0.3332  0.1637  0.3328
#> m  0.6083 -0.6661 -0.0817
#> y -0.3382  0.6261 -0.3247
#> 
#> [[70]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.0999 -0.1531  0.1656
#> [2,] -0.1531  0.2345 -0.2537
#> [3,]  0.1656 -0.2537  0.2745
#> 
#> 
#> [[71]]
#> [[71]][[1]]
#>         x       m       y
#> x -0.3295 -0.0327  0.3620
#> m  0.4292 -0.2922 -0.1802
#> y -0.3511  0.5100 -0.7638
#> 
#> [[71]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4526  0.0713 -0.3305
#> [2,]  0.0713  0.0325 -0.0575
#> [3,] -0.3305 -0.0575  0.2428
#> 
#> 
#> [[72]]
#> [[72]][[1]]
#>         x       m       y
#> x -0.7649  0.0023 -0.0388
#> m  0.2166 -0.3383 -0.1438
#> y -0.2593  0.9316 -0.7254
#> 
#> [[72]][[2]]
#>      [,1] [,2] [,3]
#> [1,]    0    0    0
#> [2,]    0    0    0
#> [3,]    0    0    0
#> 
#> 
#> [[73]]
#> [[73]][[1]]
#>         x       m       y
#> x -0.3282 -0.1264 -0.1184
#> m  0.7138 -0.9852 -0.2561
#> y -0.8044  0.5748 -1.2341
#> 
#> [[73]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.6982 0.4062 0.0483
#> [2,] 0.4062 0.2404 0.0496
#> [3,] 0.0483 0.0496 0.1159
#> 
#> 
#> [[74]]
#> [[74]][[1]]
#>         x       m       y
#> x -0.0355 -0.1531 -0.0393
#> m  1.0087 -0.5874 -0.0048
#> y -0.5520  0.5136 -0.3720
#> 
#> [[74]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1988 -0.2241  0.3197
#> [2,] -0.2241  0.2527 -0.3605
#> [3,]  0.3197 -0.3605  0.5142
#> 
#> 
#> [[75]]
#> [[75]][[1]]
#>         x       m       y
#> x -0.4091  0.2248 -0.1308
#> m  1.1192 -0.2920 -0.0253
#> y -0.5631  1.3991 -0.6319
#> 
#> [[75]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.3777 0.0686 0.4435
#> [2,] 0.0686 0.0125 0.0806
#> [3,] 0.4435 0.0806 0.5207
#> 
#> 
#> [[76]]
#> [[76]][[1]]
#>         x       m       y
#> x -0.4452  0.1454 -0.2587
#> m  0.3094 -0.7091 -0.1579
#> y -0.2925  0.9600 -1.1439
#> 
#> [[76]][[2]]
#>         [,1]    [,2]   [,3]
#> [1,]  0.4494 -0.2002 0.0348
#> [2,] -0.2002  0.0929 0.0395
#> [3,]  0.0348  0.0395 0.8192
#> 
#> 
#> [[77]]
#> [[77]][[1]]
#>         x       m       y
#> x -0.4079  0.2799  0.1064
#> m  0.4409 -0.5127 -0.2555
#> y -0.1341  0.1615 -1.1697
#> 
#> [[77]][[2]]
#>        [,1]    [,2]    [,3]
#> [1,]  0.566 -0.3470 -0.0920
#> [2,] -0.347  0.4580 -0.1594
#> [3,] -0.092 -0.1594  0.2049
#> 
#> 
#> [[78]]
#> [[78]][[1]]
#>         x       m       y
#> x -0.4250  0.1499 -0.1253
#> m  0.7440 -0.5436 -0.2915
#> y -0.3072  0.2398 -0.7246
#> 
#> [[78]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1668  0.1920 -0.0725
#> [2,]  0.1920  0.2210 -0.0835
#> [3,] -0.0725 -0.0835  0.0315
#> 
#> 
#> [[79]]
#> [[79]][[1]]
#>         x       m       y
#> x -0.9654  0.3400 -0.2761
#> m  1.6497 -0.8338 -0.5047
#> y -0.3753  0.3957 -0.3499
#> 
#> [[79]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4109 -0.4424 -0.4244
#> [2,] -0.4424  0.6746  0.4574
#> [3,] -0.4244  0.4574  0.4383
#> 
#> 
#> [[80]]
#> [[80]][[1]]
#>         x       m       y
#> x -0.3026 -0.2047  0.0103
#> m  0.7055 -0.9353 -0.6513
#> y -0.2565  0.7110 -0.5678
#> 
#> [[80]][[2]]
#>         [,1]    [,2]   [,3]
#> [1,]  0.9495 -0.5654 0.2685
#> [2,] -0.5654  0.5518 0.0761
#> [3,]  0.2685  0.0761 0.3347
#> 
#> 
#> [[81]]
#> [[81]][[1]]
#>         x       m       y
#> x -0.4775 -0.6702 -0.3674
#> m  0.6551 -1.1234  0.4847
#> y -0.4049  0.6701 -0.7851
#> 
#> [[81]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1560 -0.1204 -0.3361
#> [2,] -0.1204  0.0929  0.2594
#> [3,] -0.3361  0.2594  0.7243
#> 
#> 
#> [[82]]
#> [[82]][[1]]
#>         x       m       y
#> x -0.8997  0.1674 -0.0816
#> m  0.0371 -0.7317 -0.0569
#> y -0.3321  0.9109 -0.6667
#> 
#> [[82]][[2]]
#>         [,1]   [,2]    [,3]
#> [1,]  0.3845 0.2797 -0.0336
#> [2,]  0.2797 0.3338  0.1989
#> [3,] -0.0336 0.1989  0.3856
#> 
#> 
#> [[83]]
#> [[83]][[1]]
#>         x       m       y
#> x -0.1421 -0.1995  0.1394
#> m  0.9687 -0.3008  0.1489
#> y -0.6430  1.1407 -0.8898
#> 
#> [[83]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.1318 -0.0655 -0.0658
#> [2,] -0.0655  0.8798  0.1607
#> [3,] -0.0658  0.1607  0.1233
#> 
#> 
#> [[84]]
#> [[84]][[1]]
#>         x       m       y
#> x -0.3852 -0.0315  0.2066
#> m  0.9687 -0.3978 -0.2098
#> y -1.1187  0.8381 -0.6674
#> 
#> [[84]][[2]]
#>         [,1]    [,2]   [,3]
#> [1,]  0.3246 -0.2849 0.0603
#> [2,] -0.2849  0.3160 0.0328
#> [3,]  0.0603  0.0328 0.1225
#> 
#> 
#> [[85]]
#> [[85]][[1]]
#>         x       m       y
#> x -1.0925  0.0420  0.4191
#> m  0.8308 -0.6494 -0.0108
#> y -0.2201  0.3032 -0.9491
#> 
#> [[85]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4257  0.0134 -0.3028
#> [2,]  0.0134  0.0004 -0.0095
#> [3,] -0.3028 -0.0095  0.2154
#> 
#> 
#> [[86]]
#> [[86]][[1]]
#>         x       m       y
#> x -0.5568 -0.1659 -0.4839
#> m  0.5665 -0.7612 -0.3270
#> y -0.6213  0.8655 -0.3879
#> 
#> [[86]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.0339 0.1125 0.0390
#> [2,] 0.1125 0.3731 0.1294
#> [3,] 0.0390 0.1294 0.0449
#> 
#> 
#> [[87]]
#> [[87]][[1]]
#>         x       m       y
#> x -0.5987 -0.4909  0.0054
#> m  0.1809 -0.8659 -0.2070
#> y -0.7260  0.8523 -0.5673
#> 
#> [[87]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2267 -0.1144 -0.3784
#> [2,] -0.1144  0.2820  0.2866
#> [3,] -0.3784  0.2866  0.6725
#> 
#> 
#> [[88]]
#> [[88]][[1]]
#>         x       m       y
#> x -0.7659 -0.1603  0.0522
#> m  0.4770 -0.3408 -0.3605
#> y -0.6887  0.9374 -0.6395
#> 
#> [[88]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.1592 0.1971 0.2191
#> [2,] 0.1971 0.4384 0.4851
#> [3,] 0.2191 0.4851 0.5368
#> 
#> 
#> [[89]]
#> [[89]][[1]]
#>         x       m       y
#> x -0.5778 -0.0748  0.0267
#> m  0.8781 -0.1231  0.3063
#> y -0.6204  0.5089 -0.8793
#> 
#> [[89]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.0008  0.0072 -0.0189
#> [2,]  0.0072  0.0689 -0.1809
#> [3,] -0.0189 -0.1809  0.4748
#> 
#> 
#> [[90]]
#> [[90]][[1]]
#>         x       m       y
#> x -0.1857  0.0988  0.2849
#> m  0.9520 -1.0198  0.0733
#> y -0.8156  0.6942 -1.1091
#> 
#> [[90]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.5997 -0.1292  0.1560
#> [2,] -0.1292  0.2586 -0.0034
#> [3,]  0.1560 -0.0034  0.5472
#> 
#> 
#> [[91]]
#> [[91]][[1]]
#>         x       m       y
#> x -0.3640 -0.1348  0.1781
#> m  1.3082 -0.8654 -0.7101
#> y -0.8689 -0.1464 -0.7441
#> 
#> [[91]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2796 -0.3091 -0.0428
#> [2,] -0.3091  0.3419  0.0474
#> [3,] -0.0428  0.0474  0.0066
#> 
#> 
#> [[92]]
#> [[92]][[1]]
#>         x       m       y
#> x -0.0762 -0.0787 -0.2930
#> m  0.7552 -0.3781 -0.0594
#> y -0.5971  0.5933 -0.8808
#> 
#> [[92]][[2]]
#>        [,1]    [,2]    [,3]
#> [1,] 0.5715  0.1102  0.5575
#> [2,] 0.1102  0.8148 -0.1461
#> [3,] 0.5575 -0.1461  0.6248
#> 
#> 
#> [[93]]
#> [[93]][[1]]
#>         x       m       y
#> x -1.3732 -0.3970  0.0553
#> m  1.3340 -0.3115  0.0461
#> y -0.1439  0.6359 -0.5621
#> 
#> [[93]][[2]]
#>         [,1]   [,2]    [,3]
#> [1,]  0.3154 0.3700 -0.0793
#> [2,]  0.3700 0.4711  0.0266
#> [3,] -0.0793 0.0266  0.4070
#> 
#> 
#> [[94]]
#> [[94]][[1]]
#>         x       m       y
#> x -0.3704 -0.4275  0.1444
#> m  1.0074 -0.4072  0.0641
#> y -0.1719  0.6905 -1.0189
#> 
#> [[94]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.3796 -0.0664 -0.3805
#> [2,] -0.0664  0.0116  0.0666
#> [3,] -0.3805  0.0666  0.3815
#> 
#> 
#> [[95]]
#> [[95]][[1]]
#>         x       m       y
#> x -0.3181 -0.0828 -0.1344
#> m  0.8139 -0.5644 -0.1238
#> y -0.3566  1.0848 -0.7158
#> 
#> [[95]][[2]]
#>        [,1]    [,2]    [,3]
#> [1,] 0.3378  0.0921  0.0676
#> [2,] 0.0921  0.2749 -0.1716
#> [3,] 0.0676 -0.1716  0.1582
#> 
#> 
#> [[96]]
#> [[96]][[1]]
#>         x       m       y
#> x -0.2664 -0.0638  0.0999
#> m  0.2656 -0.3639 -0.2983
#> y -0.8135  0.9058 -1.0199
#> 
#> [[96]][[2]]
#>         [,1]    [,2]   [,3]
#> [1,]  0.3741 -0.0441 0.0745
#> [2,] -0.0441  0.1206 0.1639
#> [3,]  0.0745  0.1639 0.2732
#> 
#> 
#> [[97]]
#> [[97]][[1]]
#>         x       m       y
#> x -0.6166 -0.1953 -0.0611
#> m  0.9496 -0.5474  0.1815
#> y -0.4707  0.6810 -0.2917
#> 
#> [[97]][[2]]
#>        [,1]   [,2]   [,3]
#> [1,] 0.5194 0.3316 0.0633
#> [2,] 0.3316 0.2116 0.0404
#> [3,] 0.0633 0.0404 0.0077
#> 
#> 
#> [[98]]
#> [[98]][[1]]
#>         x       m       y
#> x -0.0497 -0.0940  0.2368
#> m  0.1232 -0.8687  0.3928
#> y -1.2069  0.5177 -0.9762
#> 
#> [[98]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.2133 -0.2306  0.0941
#> [2,] -0.2306  0.2577 -0.1322
#> [3,]  0.0941 -0.1322  0.1518
#> 
#> 
#> [[99]]
#> [[99]][[1]]
#>         x       m       y
#> x -0.7706  0.1325 -0.1804
#> m  0.8584 -0.7141 -0.1271
#> y -0.4257  0.9780 -0.3795
#> 
#> [[99]][[2]]
#>         [,1]    [,2]    [,3]
#> [1,]  0.4109 -0.3524  0.0863
#> [2,] -0.3524  0.3022 -0.0740
#> [3,]  0.0863 -0.0740  0.0181
#> 
#> 
#> [[100]]
#> [[100]][[1]]
#>         x       m       y
#> x -0.2084  0.1132  0.0256
#> m  0.5607 -0.6442 -0.3224
#> y -0.1631  0.8814 -0.5928
#> 
#> [[100]][[2]]
#>         [,1]    [,2]   [,3]
#> [1,]  0.3654 -0.0459 0.0433
#> [2,] -0.0459  0.2185 0.0453
#> [3,]  0.0433  0.0453 0.0173
#> 
#>