Delta Method Confidence Intervals

Delta Method Confidence Intervals

Usage

# S3 method for ctmeddelta
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class ctmeddelta.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)
confint(delta)
#>     effect interval      2.5 %      97.5 %
#> 1    total        1 -0.1792687 -0.05570067
#> 2   direct        1 -0.3867332 -0.22369725
#> 3 indirect        1  0.1487954  0.22666574

# Range of time-intervals ---------------------------------------------------
delta <- DeltaMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m"
)
confint(delta)
#>      effect interval         2.5 %        97.5 %
#> 1     total        1 -1.792687e-01 -0.0557006699
#> 2    direct        1 -3.867332e-01 -0.2236972542
#> 3  indirect        1  1.487954e-01  0.2266657374
#> 4     total        2  1.763296e-02  0.1616663805
#> 5    direct        2 -4.494536e-01 -0.2599610560
#> 6  indirect        2  3.667970e-01  0.5219170390
#> 7     total        3  2.111651e-01  0.3599382553
#> 8    direct        3 -4.059407e-01 -0.2232276042
#> 9  indirect        3  4.993473e-01  0.7009243460
#> 10    total        4  3.220000e-01  0.4712114540
#> 11   direct        4 -3.359179e-01 -0.1683260275
#> 12 indirect        4  5.331570e-01  0.7642982496
#> 13    total        5  3.549707e-01  0.5072741429
#> 14   direct        5 -2.671754e-01 -0.1175583898
#> 15 indirect        5  5.000528e-01  0.7469258895
#> 16    total        6  3.365880e-01  0.4934882037
#> 17   direct        6 -2.082041e-01 -0.0776046942
#> 18 indirect        6  4.333483e-01  0.6825366709
#> 19    total        7  2.920002e-01  0.4513809716
#> 20   direct        7 -1.603908e-01 -0.0486840299
#> 21 indirect        7  3.561830e-01  0.5962729480
#> 22    total        8  2.384150e-01  0.3964095275
#> 23   direct        8 -1.226832e-01 -0.0288617274
#> 24 indirect        8  2.816820e-01  0.5046874189
#> 25    total        9  1.856325e-01  0.3386803468
#> 26   direct        9 -9.338705e-02 -0.0158469885
#> 27 indirect        9  2.159063e-01  0.4176404909
#> 28    total       10  1.384146e-01  0.2841022127
#> 29   direct       10 -7.082676e-02 -0.0076397386
#> 30 indirect       10  1.607298e-01  0.3402534812
#> 31    total       11  9.858764e-02  0.2355584765
#> 32   direct       11 -5.355458e-02 -0.0026908560
#> 33 indirect       11  1.159205e-01  0.2744710119
#> 34    total       12  6.642321e-02  0.1939575046
#> 35   direct       12 -4.038756e-02  0.0001236903
#> 36 indirect       12  8.040186e-02  0.2202427238
#> 37    total       13  4.141717e-02  0.1590793337
#> 38   direct       13 -3.038487e-02  0.0015853565
#> 39 indirect       13  5.288527e-02  0.1764107506
#> 40    total       14  2.269987e-02  0.1301793340
#> 41   direct       14 -2.280894e-02  0.0022201608
#> 42 indirect       14  3.211269e-02  0.1413552976
#> 43    total       15  9.258554e-03  0.1063599096
#> 44   direct       15 -1.708662e-02  0.0023724551
#> 45 indirect       15  1.691483e-02  0.1134178004
#> 46    total       16  6.982039e-05  0.0867619260
#> 47   direct       16 -1.277523e-02  0.0022615976
#> 48 indirect       16  6.218458e-03  0.0911269231
#> 49    total       17 -5.818234e-03  0.0706403916
#> 50   direct       17 -9.534463e-03  0.0020231646
#> 51 indirect       17 -9.444874e-04  0.0732779428
#> 52    total       18 -9.239112e-03  0.0573782840
#> 53   direct       18 -7.103759e-03  0.0017378590
#> 54 indirect       18 -5.421536e-03  0.0589266080
#> 55    total       19 -1.088784e-02  0.0464736613
#> 56   direct       19 -5.284347e-03  0.0014512617
#> 57 indirect       19 -7.927967e-03  0.0473468697
#> 58    total       20 -1.132016e-02  0.0375190345
#> 59   direct       20 -3.925085e-03  0.0011870382
#> 60 indirect       20 -9.044886e-03  0.0379818111