Monte Carlo Method Confidence Intervals

Monte Carlo Method Confidence Intervals

Usage

# S3 method for ctmedmc
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class ctmedmc.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.002704274, -0.001475275, 0.000949122,
    -0.001619422, 0.000885122, -0.000569404,
    0.00085493, -0.000465824, 0.000297815,
    -0.001475275, 0.004428442, -0.002642303,
    0.000980573, -0.00271817, 0.001618805,
    -0.000586921, 0.001478421, -0.000871547,
    0.000949122, -0.002642303, 0.006402668,
    -0.000697798, 0.001813471, -0.004043138,
    0.000463086, -0.001120949, 0.002271711,
    -0.001619422, 0.000980573, -0.000697798,
    0.002079286, -0.001152501, 0.000753,
    -0.001528701, 0.000820587, -0.000517524,
    0.000885122, -0.00271817, 0.001813471,
    -0.001152501, 0.00342605, -0.002075005,
    0.000899165, -0.002532849, 0.001475579,
    -0.000569404, 0.001618805, -0.004043138,
    0.000753, -0.002075005, 0.004984032,
    -0.000622255, 0.001634917, -0.003705661,
    0.00085493, -0.000586921, 0.000463086,
    -0.001528701, 0.000899165, -0.000622255,
    0.002060076, -0.001096684, 0.000686386,
    -0.000465824, 0.001478421, -0.001120949,
    0.000820587, -0.002532849, 0.001634917,
    -0.001096684, 0.003328692, -0.001926088,
    0.000297815, -0.000871547, 0.002271711,
    -0.000517524, 0.001475579, -0.003705661,
    0.000686386, -0.001926088, 0.004726235
  ),
  nrow = 9
)

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
confint(mc)
#>     effect interval      2.5 %      97.5 %
#> 1    total        1 -0.1597830 -0.03064721
#> 2   direct        1 -0.3340759 -0.18151643
#> 3 indirect        1  0.1312711  0.20074746

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
confint(mc)
#>      effect interval         2.5 %        97.5 %
#> 1     total        1 -0.1465395904 -3.933845e-02
#> 2    direct        1 -0.3385171506 -1.808655e-01
#> 3  indirect        1  0.1302737772  2.026325e-01
#> 4     total        2  0.0294366713  1.482262e-01
#> 5    direct        2 -0.4022948656 -2.184674e-01
#> 6  indirect        2  0.3215710008  4.717419e-01
#> 7     total        3  0.1957772671  3.111064e-01
#> 8    direct        3 -0.3714985373 -1.974495e-01
#> 9  indirect        3  0.4438600954  6.232289e-01
#> 10    total        4  0.2849301650  3.994857e-01
#> 11   direct        4 -0.3077271246 -1.595457e-01
#> 12 indirect        4  0.4793349550  6.695436e-01
#> 13    total        5  0.3110449794  4.286952e-01
#> 14   direct        5 -0.2528212871 -1.200537e-01
#> 15 indirect        5  0.4568724275  6.540173e-01
#> 16    total        6  0.2973707066  4.101927e-01
#> 17   direct        6 -0.2044431613 -8.660404e-02
#> 18 indirect        6  0.3965343917  5.981419e-01
#> 19    total        7  0.2549263496  3.696481e-01
#> 20   direct        7 -0.1617024700 -5.879492e-02
#> 21 indirect        7  0.3278925486  5.325988e-01
#> 22    total        8  0.2046352029  3.184604e-01
#> 23   direct        8 -0.1265626673 -4.048274e-02
#> 24 indirect        8  0.2573872520  4.503167e-01
#> 25    total        9  0.1542244858  2.643303e-01
#> 26   direct        9 -0.1016733099 -2.775551e-02
#> 27 indirect        9  0.1919390037  3.660036e-01
#> 28    total       10  0.1111736581  2.129763e-01
#> 29   direct       10 -0.0815241605 -1.874093e-02
#> 30 indirect       10  0.1402397269  2.878081e-01
#> 31    total       11  0.0754302553  1.719416e-01
#> 32   direct       11 -0.0653910259 -1.225771e-02
#> 33 indirect       11  0.0986850483  2.201912e-01
#> 34    total       12  0.0487271648  1.346305e-01
#> 35   direct       12 -0.0524223061 -7.718129e-03
#> 36 indirect       12  0.0660913722  1.685220e-01
#> 37    total       13  0.0292569750  1.039515e-01
#> 38   direct       13 -0.0420278003 -4.648616e-03
#> 39 indirect       13  0.0417067239  1.270337e-01
#> 40    total       14  0.0161298194  7.908357e-02
#> 41   direct       14 -0.0337088261 -2.704671e-03
#> 42 indirect       14  0.0250575822  9.932843e-02
#> 43    total       15  0.0077471873  6.201970e-02
#> 44   direct       15 -0.0270545966 -1.525102e-03
#> 45 indirect       15  0.0142780898  7.728964e-02
#> 46    total       16  0.0027248315  4.914621e-02
#> 47   direct       16 -0.0216352203 -8.575519e-04
#> 48 indirect       16  0.0077697606  5.785043e-02
#> 49    total       17 -0.0001189665  3.911016e-02
#> 50   direct       17 -0.0171058783 -4.904416e-04
#> 51 indirect       17  0.0038256683  4.599999e-02
#> 52    total       18 -0.0031125206  3.115184e-02
#> 53   direct       18 -0.0135280017 -2.804173e-04
#> 54 indirect       18  0.0015570376  3.455476e-02
#> 55    total       19 -0.0034847227  2.500385e-02
#> 56   direct       19 -0.0107016430 -1.585191e-04
#> 57 indirect       19 -0.0001163848  2.668322e-02
#> 58    total       20 -0.0036274364  2.007867e-02
#> 59   direct       20 -0.0084685928 -8.854448e-05
#> 60 indirect       20 -0.0005780885  2.105821e-02