Print Method for Object of Class ctmedmc

Print Method for Object of Class ctmedmc

Usage

# S3 method for ctmedmc
print(x, alpha = 0.05, digits = 4, ...)

Arguments

x

an object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
print(mc)
#> $`1`
#>          interval     est     se R    2.5%   97.5%
#> total           1 -0.1175 0.0565 5 -0.1962 -0.0613
#> direct          1 -0.3052 0.0565 5 -0.3655 -0.2360
#> indirect        1  0.1877 0.0095 5  0.1565  0.1808
#> 

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
print(mc)
#> $`1`
#>          interval     est     se R    2.5%   97.5%
#> total           1 -0.1175 0.0292 5 -0.1423 -0.0703
#> direct          1 -0.3052 0.0425 5 -0.3437 -0.2452
#> indirect        1  0.1877 0.0221 5  0.1459  0.2019
#> 
#> $`2`
#>          interval     est     se R    2.5%   97.5%
#> total           2  0.0896 0.0342 5  0.0602  0.1382
#> direct          2 -0.3547 0.0513 5 -0.4042 -0.2841
#> indirect        2  0.4444 0.0480 5  0.3620  0.4848
#> 
#> $`3`
#>          interval     est     se R    2.5%   97.5%
#> total           3  0.2856 0.0393 5  0.2305  0.3238
#> direct          3 -0.3146 0.0503 5 -0.3669 -0.2469
#> indirect        3  0.6001 0.0644 5  0.5045  0.6675
#> 
#> $`4`
#>          interval     est     se R    2.5%   97.5%
#> total           4  0.3966 0.0420 5  0.3353  0.4322
#> direct          4 -0.2521 0.0464 5 -0.3043 -0.1939
#> indirect        4  0.6487 0.0724 5  0.5568  0.7347
#> 
#> $`5`
#>          interval     est     se R    2.5%   97.5%
#> total           5  0.4311 0.0413 5  0.3754  0.4750
#> direct          5 -0.1924 0.0418 5 -0.2421 -0.1449
#> indirect        5  0.6235 0.0733 5  0.5439  0.7170
#> 
#> $`6`
#>          interval     est     se R    2.5%   97.5%
#> total           6  0.4150 0.0367 5  0.3708  0.4605
#> direct          6 -0.1429 0.0369 5 -0.1886 -0.1054
#> indirect        6  0.5579 0.0679 5  0.4941  0.6491
#> 
#> $`7`
#>          interval     est     se R    2.5%   97.5%
#> total           7  0.3717 0.0291 5  0.3413  0.4125
#> direct          7 -0.1045 0.0321 5 -0.1454 -0.0755
#> indirect        7  0.4762 0.0579 5  0.4293  0.5578
#> 
#> $`8`
#>          interval     est     se R    2.5%   97.5%
#> total           8  0.3174 0.0197 5  0.3011  0.3496
#> direct          8 -0.0758 0.0276 5 -0.1125 -0.0536
#> indirect        8  0.3932 0.0453 5  0.3629  0.4610
#> 
#> $`9`
#>          interval     est     se R    2.5%   97.5%
#> total           9  0.2622 0.0109 5  0.2586  0.2844
#> direct          9 -0.0546 0.0234 5 -0.0877 -0.0378
#> indirect        9  0.3168 0.0322 5  0.2990  0.3695
#> 
#> $`10`
#>          interval     est     se R    2.5%   97.5%
#> total          10  0.2113 0.0079 5  0.2059  0.2251
#> direct         10 -0.0392 0.0196 5 -0.0683 -0.0266
#> indirect       10  0.2505 0.0206 5  0.2387  0.2885
#> 
#> $`11`
#>          interval     est     se R    2.5%   97.5%
#> total          11  0.1671 0.0127 5  0.1527  0.1842
#> direct         11 -0.0281 0.0162 5 -0.0532 -0.0186
#> indirect       11  0.1952 0.0127 5  0.1881  0.2205
#> 
#> $`12`
#>          interval     est     se R    2.5%   97.5%
#> total          12  0.1302 0.0180 5  0.1084  0.1540
#> direct         12 -0.0201 0.0134 5 -0.0413 -0.0130
#> indirect       12  0.1503 0.0112 5  0.1450  0.1695
#> 
#> $`13`
#>          interval     est     se R    2.5%   97.5%
#> total          13  0.1002 0.0218 5  0.0731  0.1284
#> direct         13 -0.0144 0.0109 5 -0.0321 -0.0091
#> indirect       13  0.1146 0.0141 5  0.1041  0.1387
#> 
#> $`14`
#>          interval     est     se R    2.5%   97.5%
#> total          14  0.0764 0.0240 5  0.0465  0.1070
#> direct         14 -0.0103 0.0088 5 -0.0250 -0.0064
#> indirect       14  0.0867 0.0170 5  0.0710  0.1140
#> 
#> $`15`
#>          interval     est     se R    2.5%   97.5%
#> total          15  0.0578 0.0246 5  0.0271  0.0892
#> direct         15 -0.0074 0.0071 5 -0.0194 -0.0044
#> indirect       15  0.0652 0.0187 5  0.0465  0.0940
#> 
#> $`16`
#>          interval     est     se R    2.5%   97.5%
#> total          16  0.0434 0.0241 5  0.0137  0.0743
#> direct         16 -0.0053 0.0057 5 -0.0151 -0.0031
#> indirect       16  0.0487 0.0193 5  0.0288  0.0776
#> 
#> $`17`
#>          interval     est     se R    2.5%   97.5%
#> total          17  0.0324 0.0227 5  0.0049  0.0620
#> direct         17 -0.0038 0.0045 5 -0.0117 -0.0021
#> indirect       17  0.0362 0.0188 5  0.0166  0.0642
#> 
#> $`18`
#>          interval     est     se R    2.5%   97.5%
#> total          18  0.0241 0.0208 5 -0.0005  0.0517
#> direct         18 -0.0027 0.0036 5 -0.0091 -0.0014
#> indirect       18  0.0268 0.0177 5  0.0086  0.0533
#> 
#> $`19`
#>          interval     est     se R    2.5%   97.5%
#> total          19  0.0178 0.0186 5 -0.0035  0.0432
#> direct         19 -0.0019 0.0029 5 -0.0071 -0.0010
#> indirect       19  0.0197 0.0162 5  0.0036  0.0442
#> 
#> $`20`
#>          interval     est     se R    2.5%   97.5%
#> total          20  0.0131 0.0164 5 -0.0048  0.0361
#> direct         20 -0.0014 0.0023 5 -0.0055 -0.0006
#> indirect       20  0.0145 0.0145 5  0.0007  0.0368
#>