Print Method for Object of Class ctmedmcphi

Print Method for Object of Class ctmedmcphi

Usage

# S3 method for class 'ctmedmcphi'
print(x, digits = 4, ...)

Arguments

x

an object of class ctmedmcphi.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Prints a list of drift matrices.

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
mc <- MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(9),
  R = 100L # use a large value for R in actual research
)
print(mc)
#> [[1]]
#>         x       m       y
#> x -0.3295  0.2220 -0.2679
#> m  0.7555 -0.8565  0.0091
#> y -0.4690  1.5626 -0.7618
#> 
#> [[2]]
#>         x       m       y
#> x -0.3830  0.0943 -0.4144
#> m  0.9453 -0.4066 -0.3812
#> y -0.1887  0.7553 -0.8437
#> 
#> [[3]]
#>         x       m      y
#> x -0.1077 -0.0344  0.001
#> m  0.4363 -0.7873 -0.033
#> y -0.9239  0.9293 -0.606
#> 
#> [[4]]
#>         x       m       y
#> x -0.7750  0.1073 -0.1698
#> m  1.2028 -0.5626 -0.0798
#> y  0.0148  0.6496 -0.5988
#> 
#> [[5]]
#>         x       m       y
#> x -1.0155  0.2169 -0.4148
#> m  0.8304 -0.3326 -0.1625
#> y -0.0059  0.3293 -0.4056
#> 
#> [[6]]
#>         x       m       y
#> x -0.5023 -0.3002  0.2227
#> m  0.5239 -0.0857  0.3440
#> y -0.0757  0.5680 -0.9511
#> 
#> [[7]]
#>         x       m       y
#> x -0.1279 -0.4010 -0.1586
#> m  0.7976 -0.4652 -0.6030
#> y -0.3153  0.3125 -0.7378
#> 
#> [[8]]
#>         x       m       y
#> x -0.6284  0.2739 -0.1972
#> m  0.4140 -0.7338  0.1835
#> y -0.1817  0.6000 -0.6670
#> 
#> [[9]]
#>         x       m       y
#> x -0.3203 -0.4097 -0.1738
#> m  1.0749 -0.4948  0.2990
#> y -0.4625  0.9025 -0.8753
#> 
#> [[10]]
#>         x       m       y
#> x -0.4946 -0.6680  0.8914
#> m  0.4559 -0.7414 -0.3237
#> y  0.0138  0.7617 -0.7599
#> 
#> [[11]]
#>         x       m       y
#> x -0.3574 -0.3296 -0.3109
#> m  0.9960 -1.0598  0.0318
#> y -0.6253  0.8023 -0.3250
#> 
#> [[12]]
#>         x       m       y
#> x -0.2388  0.1046  0.1173
#> m  0.5637 -0.5110 -0.0937
#> y -0.4943  0.3738 -0.8324
#> 
#> [[13]]
#>         x       m       y
#> x -0.4017 -0.1042 -0.1136
#> m  0.4225 -1.0385 -0.7173
#> y -0.2042  0.4482 -0.7084
#> 
#> [[14]]
#>         x       m       y
#> x -0.5168 -0.1229  0.2009
#> m  0.4649 -0.5269  0.1268
#> y -0.4530  0.6824 -1.1855
#> 
#> [[15]]
#>         x       m       y
#> x -0.4401  0.1808 -0.1407
#> m  0.5675 -0.7721 -0.0552
#> y  0.0013 -0.1627 -0.8033
#> 
#> [[16]]
#>         x      m       y
#> x -0.7018 -0.287 -0.4060
#> m  0.4014 -0.926 -0.6191
#> y -0.4522  1.042 -0.9174
#> 
#> [[17]]
#>         x       m       y
#> x -0.2428 -0.1850  0.3692
#> m  0.1528 -0.5686 -0.1734
#> y -0.1697  0.4695 -0.5461
#> 
#> [[18]]
#>         x       m       y
#> x -0.8570 -0.0544 -0.0911
#> m  1.4080 -0.4738 -1.0594
#> y  0.0874  0.8196 -0.4478
#> 
#> [[19]]
#>         x       m       y
#> x -0.4047 -0.2895  0.1201
#> m  0.7085 -0.7802  0.3837
#> y  0.3360  0.2906 -0.3345
#> 
#> [[20]]
#>         x       m       y
#> x -0.1396 -0.0238 -0.1357
#> m  0.3961 -0.4810  0.0038
#> y  0.1203  0.4805 -0.1099
#> 
#> [[21]]
#>         x       m       y
#> x -0.5919  0.0691 -0.0411
#> m  0.7065 -0.8078 -0.0403
#> y -0.7677  0.9824 -0.9226
#> 
#> [[22]]
#>         x       m       y
#> x -0.3922 -0.0144 -0.1370
#> m  0.5602 -0.4864 -0.6544
#> y -0.7721  0.4842 -0.0734
#> 
#> [[23]]
#>         x       m       y
#> x -0.0552 -0.2630 -0.2973
#> m  1.6914 -0.6528 -0.0127
#> y -0.4832  0.2784 -0.6221
#> 
#> [[24]]
#>         x       m       y
#> x -0.4761 -0.1023  0.0572
#> m  1.0788 -0.6270 -0.1512
#> y -1.0433  0.8290 -1.1084
#> 
#> [[25]]
#>         x       m       y
#> x -0.5386  0.3221  0.0621
#> m  0.4560 -0.7578  0.2615
#> y -0.5445 -0.2587 -0.8716
#> 
#> [[26]]
#>         x       m       y
#> x -0.8226 -0.0375  0.1275
#> m  0.9053 -0.3728 -0.0042
#> y -0.1520  0.6408 -0.3339
#> 
#> [[27]]
#>         x       m       y
#> x -0.3714 -0.2501  0.0273
#> m  0.3441 -0.5902  0.4339
#> y -0.1173  0.6587 -0.6179
#> 
#> [[28]]
#>         x       m       y
#> x -0.2288  0.1370  0.0740
#> m  0.6924 -0.8384 -0.6592
#> y -0.6282  0.4929 -0.9160
#> 
#> [[29]]
#>         x       m       y
#> x -0.9407  0.4237  0.1673
#> m  1.0778 -0.5615 -0.1271
#> y -0.1173  0.1810 -0.3245
#> 
#> [[30]]
#>         x       m       y
#> x -0.4081  0.0028  0.0232
#> m  0.4974 -0.4713 -0.4130
#> y -0.4946  0.2607 -0.5754
#> 
#> [[31]]
#>         x       m       y
#> x -0.1392 -0.4151  0.1329
#> m  0.4121 -0.3814 -0.0712
#> y -0.8051  0.4695 -0.5703
#> 
#> [[32]]
#>         x       m       y
#> x -1.0668 -0.3412 -0.0488
#> m  1.5288 -0.5792 -0.6432
#> y -0.0236  0.6782 -0.9510
#> 
#> [[33]]
#>         x       m       y
#> x -0.6313  0.1458  0.5649
#> m  1.0035 -1.0179 -0.0652
#> y -0.7079  0.5710 -0.9275
#> 
#> [[34]]
#>         x       m       y
#> x -0.5981  0.5447  0.2370
#> m  0.5518 -0.8220  0.1444
#> y -0.2837  0.7441 -0.8026
#> 
#> [[35]]
#>         x       m       y
#> x -0.3760 -0.6159 -0.1619
#> m  0.7455 -0.4486  0.3184
#> y -0.5399  0.8208 -0.9259
#> 
#> [[36]]
#>         x       m       y
#> x -0.8211  0.3333  0.5403
#> m  0.4818 -0.6742  0.1147
#> y -0.4575  1.0438 -0.6069
#> 
#> [[37]]
#>         x       m       y
#> x -0.0438 -0.1913 -0.2189
#> m  1.0121 -1.0879  0.4560
#> y -0.7420  0.9661 -0.4182
#> 
#> [[38]]
#>         x       m       y
#> x -0.6305  0.0519 -0.3075
#> m  0.2891 -0.8919 -0.1079
#> y -0.5822  0.8868 -0.8725
#> 
#> [[39]]
#>         x       m       y
#> x -0.9969  0.6508 -0.2677
#> m  0.4777 -0.5442 -0.0148
#> y -0.8395  0.8875 -1.0651
#> 
#> [[40]]
#>         x       m       y
#> x -0.4153 -0.3777  0.0552
#> m  0.6768 -0.7006 -0.3646
#> y -0.5796  0.6035 -0.3262
#> 
#> [[41]]
#>         x       m       y
#> x -0.4109 -0.2373 -0.0544
#> m  0.5736 -0.9596 -0.2123
#> y -0.3284  1.0132 -0.3790
#> 
#> [[42]]
#>         x       m       y
#> x -0.1467 -0.3364 -0.4359
#> m  0.2769 -0.4348  0.0457
#> y -0.5458  0.6445 -0.6675
#> 
#> [[43]]
#>         x       m       y
#> x -0.3225  0.0097  0.4471
#> m  0.2850 -0.5174  0.2141
#> y -1.0253  1.4949 -0.8394
#> 
#> [[44]]
#>         x       m       y
#> x -0.7559 -0.1536  0.4511
#> m  0.9942 -0.9598  0.5648
#> y -0.4914  0.5642 -1.0631
#> 
#> [[45]]
#>         x       m       y
#> x -0.1769  0.0374 -0.4160
#> m  0.9637 -0.4728 -0.2756
#> y -0.6885  0.3754 -0.5568
#> 
#> [[46]]
#>         x       m       y
#> x -0.3169 -0.1641  0.0323
#> m  1.2726 -0.4854  0.2306
#> y -0.9982  0.6590 -0.7262
#> 
#> [[47]]
#>         x       m       y
#> x -0.5678  0.2690  0.0708
#> m  0.4662 -0.5777  0.0837
#> y -0.5875  0.0619 -0.9957
#> 
#> [[48]]
#>         x       m       y
#> x -0.9053 -0.4573 -0.0542
#> m  0.6444 -0.5776 -0.2122
#> y -0.0040  0.9637 -0.3440
#> 
#> [[49]]
#>         x       m       y
#> x -0.3641  0.2027  0.0351
#> m  0.7360 -0.7475 -0.3221
#> y -0.4191  0.3433 -1.0359
#> 
#> [[50]]
#>         x       m       y
#> x -0.1074 -0.1336 -0.3040
#> m  1.0164 -0.4057 -0.1125
#> y -0.5545  1.0478 -1.0712
#> 
#> [[51]]
#>         x       m       y
#> x -0.3634 -0.0708 -0.3003
#> m  0.9753 -0.7200  1.1534
#> y -0.4025  0.9655 -0.8761
#> 
#> [[52]]
#>         x       m       y
#> x -0.4116  0.1819  0.0896
#> m  0.2703 -0.4336  0.1018
#> y -0.5597  0.6048 -0.6692
#> 
#> [[53]]
#>         x       m       y
#> x -0.9718 -0.1154 -0.1710
#> m  1.2792 -0.6169 -0.1020
#> y -0.5817  0.8355 -0.3922
#> 
#> [[54]]
#>         x       m       y
#> x -0.3671 -0.0507  0.2011
#> m  0.4971 -0.9741 -0.2671
#> y -0.3027  1.1691 -0.6000
#> 
#> [[55]]
#>         x       m       y
#> x -0.9081 -0.2732 -0.1377
#> m  0.7826 -0.2509  0.0539
#> y -0.4285  0.7691 -0.6318
#> 
#> [[56]]
#>         x       m       y
#> x -0.6330 -0.1352  0.4032
#> m  1.3981 -0.4071 -0.4848
#> y -0.0628  0.8709 -0.6484
#> 
#> [[57]]
#>         x       m       y
#> x -0.5716 -0.2904 -0.9815
#> m  0.6002 -0.5104  0.3351
#> y -0.5145  0.9666 -0.5926
#> 
#> [[58]]
#>         x       m       y
#> x -0.0853 -0.1908  0.5859
#> m -0.0739 -0.7610 -0.2784
#> y -1.0674  0.6428 -0.8445
#> 
#> [[59]]
#>         x       m       y
#> x -0.3777  0.0656  0.1128
#> m  0.8989 -0.5156 -0.4871
#> y -0.4501  0.3930 -0.4726
#> 
#> [[60]]
#>         x       m       y
#> x -1.1444  0.2666 -0.0877
#> m  0.2862 -0.3202 -0.2117
#> y -0.5150  1.3012 -0.5467
#> 
#> [[61]]
#>         x       m       y
#> x -0.5269  0.0676 -0.0710
#> m  0.7264 -1.0256  0.0231
#> y -0.0178  0.4719 -0.4232
#> 
#> [[62]]
#>         x       m       y
#> x -0.2998 -0.1015  0.1827
#> m  0.4915 -0.6801  0.3549
#> y -0.0945  0.9665 -0.9195
#> 
#> [[63]]
#>         x       m       y
#> x -0.0936 -0.1392  0.3097
#> m  1.0521 -0.4798 -0.4324
#> y -0.5027  0.6471 -0.7498
#> 
#> [[64]]
#>         x       m       y
#> x -0.6689 -0.0929 -0.2490
#> m  1.0424 -0.1300  0.1806
#> y -0.9117  1.0659 -1.1202
#> 
#> [[65]]
#>         x       m       y
#> x -0.4898 -0.1196  0.4541
#> m  0.6970 -0.1351 -0.3419
#> y -0.3528  0.1951 -0.7034
#> 
#> [[66]]
#>         x       m       y
#> x -0.2360 -0.0164 -0.3757
#> m  0.9830 -0.6066 -0.2857
#> y -0.1591  0.7775 -1.0288
#> 
#> [[67]]
#>         x       m       y
#> x -0.6339  0.0342  0.0401
#> m  0.8270 -1.0489  0.4998
#> y  0.0652  0.5969 -0.5418
#> 
#> [[68]]
#>         x       m       y
#> x -0.5422 -0.0616  0.0014
#> m  0.6718 -0.3817  0.4155
#> y -0.7427  0.3234 -0.9757
#> 
#> [[69]]
#>         x       m       y
#> x -0.0403 -0.0506 -0.5266
#> m  0.7176 -0.8780 -0.1823
#> y -0.4755  1.0920 -0.7177
#> 
#> [[70]]
#>         x       m       y
#> x -0.4184  0.0873  0.2936
#> m  0.5819 -0.5672 -0.2931
#> y -0.5228  0.6846 -1.1257
#> 
#> [[71]]
#>         x       m       y
#> x -0.6265  0.2090 -0.1042
#> m  0.7528 -0.9778  0.0057
#> y -0.5539  0.7630 -0.0298
#> 
#> [[72]]
#>         x       m       y
#> x -0.3142 -0.4175 -0.3436
#> m  0.6047 -0.6731  0.4674
#> y -0.0897  0.5923 -1.0471
#> 
#> [[73]]
#>         x       m       y
#> x -0.7625 -0.1079 -0.1096
#> m  0.5141 -0.8554  0.3487
#> y -0.6520  0.6899 -1.4064
#> 
#> [[74]]
#>         x       m       y
#> x -0.3623 -0.2808 -0.0804
#> m  0.8808 -0.4030  0.1781
#> y -0.6301  0.8759 -1.0519
#> 
#> [[75]]
#>         x       m       y
#> x -0.5995 -0.4198 -0.5094
#> m  0.6576 -0.2874  0.3968
#> y -0.2405  0.7774 -0.6126
#> 
#> [[76]]
#>         x       m       y
#> x -0.2502  0.0400 -0.2393
#> m  1.0703 -0.0552 -0.2317
#> y -0.2040  0.5293 -0.4795
#> 
#> [[77]]
#>         x       m       y
#> x -0.3583  0.4105 -0.3260
#> m  0.4468 -0.7572 -0.4285
#> y  0.0135  0.4630 -0.4630
#> 
#> [[78]]
#>         x       m       y
#> x -0.6778  0.0109 -0.5704
#> m  0.8244 -0.3219 -0.1490
#> y -0.5195  0.4863 -1.0459
#> 
#> [[79]]
#>         x       m       y
#> x -0.1261 -0.2689  0.1719
#> m  1.2347 -0.8141 -0.5353
#> y -0.5857  1.1799 -0.7561
#> 
#> [[80]]
#>         x       m       y
#> x -0.2520 -0.4989  0.4904
#> m  0.5242 -0.4786 -0.2059
#> y -0.3187  0.9093 -0.7219
#> 
#> [[81]]
#>         x       m       y
#> x -0.4483  0.5324 -0.0998
#> m  0.1837 -0.3331  0.2339
#> y -0.3454  0.9350 -1.0312
#> 
#> [[82]]
#>         x       m       y
#> x -0.4740  0.4396 -0.0832
#> m  0.6544 -0.7612  0.3754
#> y -0.2096  0.5643 -0.9836
#> 
#> [[83]]
#>         x       m       y
#> x -0.8703 -0.2334  0.4392
#> m  1.4877 -0.6639 -0.1450
#> y -0.2625  0.8243 -0.8017
#> 
#> [[84]]
#>         x       m       y
#> x -0.1962 -0.3202  0.0654
#> m  1.3910 -0.3775 -0.3593
#> y -0.5640  0.3310 -0.3806
#> 
#> [[85]]
#>         x       m       y
#> x -0.3905 -0.2777 -0.3170
#> m  0.3367 -0.6976  0.0301
#> y  0.6498  0.8609 -0.6813
#> 
#> [[86]]
#>         x       m       y
#> x -0.6248 -0.4520  0.2134
#> m  0.8028 -0.2845  0.0879
#> y -0.9383  0.9861 -0.6502
#> 
#> [[87]]
#>         x       m       y
#> x -0.1593 -0.5304 -0.0042
#> m  0.9188 -0.6926 -0.0764
#> y -0.2094  0.5743 -0.4895
#> 
#> [[88]]
#>         x       m       y
#> x -0.4309 -0.1042 -0.0629
#> m  0.7116 -0.6306  0.1055
#> y -0.4704  0.3523 -0.9833
#> 
#> [[89]]
#>         x       m       y
#> x -0.1073  0.0902  0.3519
#> m  0.0455 -0.5125  0.2619
#> y -0.4501  0.8671 -0.6955
#> 
#> [[90]]
#>         x       m       y
#> x -0.1855 -0.1612 -0.1172
#> m  0.9788 -0.8409 -0.3104
#> y -0.5546  0.6565 -0.6502
#> 
#> [[91]]
#>         x       m       y
#> x -0.2742 -0.0278 -0.0435
#> m  0.3944 -0.2534 -0.2422
#> y  0.6082  0.6999 -0.5411
#> 
#> [[92]]
#>         x       m       y
#> x -0.8169  0.2507  0.1256
#> m  0.2168 -0.8083  0.4825
#> y  0.0757  0.5949 -0.5242
#> 
#> [[93]]
#>         x       m       y
#> x -0.3167  0.0613 -0.1022
#> m  0.6812 -0.6474 -0.5121
#> y -0.3802  0.8593 -0.4606
#> 
#> [[94]]
#>         x       m       y
#> x -0.4927  0.0970  0.1964
#> m  0.6259 -0.5075  0.2843
#> y -0.2871  0.1613 -1.0010
#> 
#> [[95]]
#>         x       m       y
#> x -0.2908 -0.0831 -0.3284
#> m  0.7124 -0.3808  0.0491
#> y -0.4735  0.7616 -0.4224
#> 
#> [[96]]
#>         x       m       y
#> x -0.2382 -0.0210  0.2435
#> m  0.8107 -0.9013  0.0443
#> y -0.5787  0.8863 -1.0013
#> 
#> [[97]]
#>         x       m       y
#> x -0.3714 -0.2303 -0.3200
#> m  1.2856 -0.9282  0.2868
#> y -0.1415  1.1378 -0.8186
#> 
#> [[98]]
#>         x       m       y
#> x -0.2475 -0.4155  0.3497
#> m  0.6953 -0.4973  0.1498
#> y -0.3532  0.7291 -0.4281
#> 
#> [[99]]
#>         x       m       y
#> x -0.2293 -0.2961  0.2112
#> m  0.6084 -0.7606  0.1396
#> y -0.7304  0.3160 -0.4800
#> 
#> [[100]]
#>         x       m       y
#> x -0.5496 -0.0348  0.0700
#> m  0.9226 -0.5533 -0.0088
#> y -0.5031  0.8429 -0.5873
#>