Summary Method for an Object of Class ctmedmc

Summary Method for an Object of Class ctmedmc

Usage

# S3 method for ctmedmc
summary(object, alpha = 0.05, ...)

Arguments

object

Object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

...

additional arguments.

Value

Returns a matrix of effects, time-interval, estimates, standard errors, test statistics, p-values, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.002704274, -0.001475275, 0.000949122,
    -0.001619422, 0.000885122, -0.000569404,
    0.00085493, -0.000465824, 0.000297815,
    -0.001475275, 0.004428442, -0.002642303,
    0.000980573, -0.00271817, 0.001618805,
    -0.000586921, 0.001478421, -0.000871547,
    0.000949122, -0.002642303, 0.006402668,
    -0.000697798, 0.001813471, -0.004043138,
    0.000463086, -0.001120949, 0.002271711,
    -0.001619422, 0.000980573, -0.000697798,
    0.002079286, -0.001152501, 0.000753,
    -0.001528701, 0.000820587, -0.000517524,
    0.000885122, -0.00271817, 0.001813471,
    -0.001152501, 0.00342605, -0.002075005,
    0.000899165, -0.002532849, 0.001475579,
    -0.000569404, 0.001618805, -0.004043138,
    0.000753, -0.002075005, 0.004984032,
    -0.000622255, 0.001634917, -0.003705661,
    0.00085493, -0.000586921, 0.000463086,
    -0.001528701, 0.000899165, -0.000622255,
    0.002060076, -0.001096684, 0.000686386,
    -0.000465824, 0.001478421, -0.001120949,
    0.000820587, -0.002532849, 0.001634917,
    -0.001096684, 0.003328692, -0.001926088,
    0.000297815, -0.000871547, 0.002271711,
    -0.000517524, 0.001475579, -0.003705661,
    0.000686386, -0.001926088, 0.004726235
  ),
  nrow = 9
)

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
summary(mc)
#>     effect interval        est         se   R       2.5%       97.5%
#> 1    total        1 -0.1000384 0.02881942 100 -0.1528160 -0.04792766
#> 2   direct        1 -0.2674539 0.03708595 100 -0.3357502 -0.20032939
#> 3 indirect        1  0.1674155 0.01628182 100  0.1338138  0.19675017

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
summary(mc)
#>      effect interval          est          se   R          2.5%         97.5%
#> 1     total        1 -0.100038374 0.030775880 100 -0.1576329476 -5.098466e-02
#> 2    direct        1 -0.267453887 0.041111717 100 -0.3508789856 -1.981548e-01
#> 3  indirect        1  0.167415512 0.018275991 100  0.1313279055  2.053333e-01
#> 4     total        2  0.079900802 0.033637708 100  0.0215897876  1.497538e-01
#> 5    direct        2 -0.320903501 0.048587001 100 -0.4202005022 -2.345465e-01
#> 6  indirect        2  0.400804303 0.038753667 100  0.3385178210  4.858798e-01
#> 7     total        3  0.250813826 0.034723545 100  0.1914236495  3.238471e-01
#> 8    direct        3 -0.291442601 0.046514746 100 -0.3871490885 -2.168632e-01
#> 9  indirect        3  0.542256427 0.051915046 100  0.4646294961  6.507494e-01
#> 10    total        4  0.344927909 0.036832913 100  0.2798358875  4.159074e-01
#> 11   direct        4 -0.237390017 0.041835206 100 -0.3159647315 -1.695679e-01
#> 12 indirect        4  0.582317927 0.059756314 100  0.4834742877  6.913648e-01
#> 13    total        5  0.369253788 0.039231754 100  0.3044293250  4.454510e-01
#> 14   direct        5 -0.182844717 0.036512431 100 -0.2569074034 -1.220830e-01
#> 15 indirect        5  0.552098505 0.063200778 100  0.4545398584  6.726042e-01
#> 16    total        6  0.348237468 0.040504206 100  0.2840325105  4.313025e-01
#> 17   direct        6 -0.136313791 0.031230360 100 -0.2041899980 -8.609139e-02
#> 18 indirect        6  0.484551259 0.062736835 100  0.3895837289  6.114523e-01
#> 19    total        7  0.304148647 0.039995769 100  0.2428112673  3.934672e-01
#> 20   direct        7 -0.099571301 0.026303660 100 -0.1558650855 -5.863706e-02
#> 21 indirect        7  0.403719949 0.059147083 100  0.3181306740  5.278625e-01
#> 22    total        8  0.252388378 0.037885068 100  0.1901097048  3.387198e-01
#> 23   direct        8 -0.071768895 0.021883565 100 -0.1213128785 -3.871628e-02
#> 24 indirect        8  0.324157274 0.053475030 100  0.2468692800  4.400461e-01
#> 25    total        9  0.201984090 0.034759718 100  0.1419181844  2.743782e-01
#> 26   direct        9 -0.051267872 0.018024319 100 -0.0939576421 -2.524641e-02
#> 27 indirect        9  0.253251962 0.046792047 100  0.1824472735  3.582678e-01
#> 28    total       10  0.157383000 0.031241815 100  0.1025023870  2.194358e-01
#> 29   direct       10 -0.036398910 0.014721530 100 -0.0703163871 -1.631111e-02
#> 30 indirect       10  0.193781910 0.039996932 100  0.1322034747  2.840093e-01
#> 31    total       11  0.120165209 0.027780485 100  0.0717473167  1.754067e-01
#> 32   direct       11 -0.025732526 0.011938203 100 -0.0541728177 -1.046461e-02
#> 33 indirect       11  0.145897735 0.033707045 100  0.0947661229  2.200662e-01
#> 34    total       12  0.090312306 0.024598453 100  0.0497568105  1.393677e-01
#> 35   direct       12 -0.018137719 0.009621540 100 -0.0420202843 -6.677660e-03
#> 36 indirect       12  0.108450025 0.028241577 100  0.0653399591  1.684519e-01
#> 37    total       13  0.067034481 0.021741078 100  0.0323427564  1.104075e-01
#> 38   direct       13 -0.012757688 0.007713125 100 -0.0330332196 -4.243372e-03
#> 39 indirect       13  0.079792169 0.023673321 100  0.0443599494  1.297860e-01
#> 40    total       14  0.049261672 0.019163592 100  0.0185328268  8.681266e-02
#> 41   direct       14 -0.008960190 0.006154817 100 -0.0259636681 -2.687681e-03
#> 42 indirect       14  0.058221862 0.019916441 100  0.0294203190  1.002295e-01
#> 43    total       15  0.035909063 0.016804069 100  0.0089787530  6.970498e-02
#> 44   direct       15 -0.006286452 0.004892032 100 -0.0204142081 -1.646603e-03
#> 45 indirect       15  0.042195515 0.016817050 100  0.0179850993  7.777738e-02
#> 46    total       16  0.026003123 0.014620414 100  0.0021654045  5.651657e-02
#> 47   direct       16 -0.004407267 0.003875434 100 -0.0160519152 -1.000834e-03
#> 48 indirect       16  0.030410389 0.014220685 100  0.0094962584  5.957436e-02
#> 49    total       17  0.018727568 0.012597051 100 -0.0016014524  4.584330e-02
#> 50   direct       17 -0.003088175 0.003061638 100 -0.0126216634 -6.034406e-04
#> 51 indirect       17  0.021815742 0.012006339 100  0.0040853463  4.754951e-02
#> 52    total       18  0.013426970 0.010736739 100 -0.0038233722  3.701436e-02
#> 53   direct       18 -0.002163065 0.002413312 100 -0.0099245088 -3.612497e-04
#> 54 indirect       18  0.015590036 0.010092847 100  0.0010517119  3.829615e-02
#> 55    total       19  0.009590557 0.009049526 100 -0.0050647507  2.980711e-02
#> 56   direct       19 -0.001514677 0.001898890 100 -0.0078038386 -1.957748e-04
#> 57 indirect       19  0.011105233 0.008430582 100 -0.0007928019  3.090021e-02
#> 58    total       20  0.006828830 0.007544432 100 -0.0052526963  2.364546e-02
#> 59   direct       20 -0.001060441 0.001492070 100 -0.0061364435 -9.629861e-05
#> 60 indirect       20  0.007889270 0.006989508 100 -0.0016982308  2.493069e-02