Generate Random Drift Matrices Using the Monte Carlo Method

This function generates random drift matrices \(\boldsymbol{\Phi}\) using the Monte Carlo method.

Usage

MCPhi(phi, vcov_phi_vec, R, test_phi = TRUE, ncores = NULL, seed = NULL)

Arguments

phi

Numeric matrix. The drift matrix (\(\boldsymbol{\Phi}\)). phi should have row and column names pertaining to the variables in the system.

vcov_phi_vec

Numeric matrix. The sampling variance-covariance matrix of \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\).

R

Positive integer. Number of replications.

test_phi

Logical. Check the generated \(\boldsymbol{\Phi}\) and generate different values if the test fails. The test includes the following:

test that the largest eigen value of \(\boldsymbol{\Phi}\) is less than one, and

test that the diagonal values of \(\boldsymbol{\Phi}\) are between 0 to negative inifinity.

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of replications R is a large value.

seed

Random seed.

Value

Returns a list of simulated drift matrices.

Details

Monte Carlo Method

Let \(\boldsymbol{\theta}\) be \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\), that is, the elements of the \(\boldsymbol{\Phi}\) matrix in vector form sorted column-wise. Let \(\hat{\boldsymbol{\theta}}\) be \(\mathrm{vec} \left( \hat{\boldsymbol{\Phi}} \right)\). Based on the asymptotic properties of maximum likelihood estimators, we can assume that estimators are normally distributed around the population parameters. $$ \hat{\boldsymbol{\theta}} \sim \mathcal{N} \left( \boldsymbol{\theta}, \mathbb{V} \left( \hat{\boldsymbol{\theta}} \right) \right) $$ Using this distributional assumption, a sampling distribution of \(\hat{\boldsymbol{\theta}}\) which we refer to as \(\hat{\boldsymbol{\theta}}^{\ast}\) can be generated by replacing the population parameters with sample estimates, that is, $$ \hat{\boldsymbol{\theta}}^{\ast} \sim \mathcal{N} \left( \hat{\boldsymbol{\theta}}, \hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right) \right) . $$

Linear Stochastic Differential Equation Model

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t}, \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\mathbf{y}_{i, t}\), \(\boldsymbol{\eta}_{i, t}\), and \(\boldsymbol{\varepsilon}_{i, t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{i, t}\) represents a vector of observed random variables, \(\boldsymbol{\eta}_{i, t}\) a vector of latent random variables, and \(\boldsymbol{\varepsilon}_{i, t}\) a vector of random measurement errors, at time \(t\) and individual \(i\). \(\boldsymbol{\nu}\) denotes a vector of intercepts, \(\boldsymbol{\Lambda}\) a matrix of factor loadings, and \(\boldsymbol{\Theta}\) the covariance matrix of \(\boldsymbol{\varepsilon}\).

An alternative representation of the measurement error is given by $$ \boldsymbol{\varepsilon}_{i, t} = \boldsymbol{\Theta}^{\frac{1}{2}} \mathbf{z}_{i, t}, \quad \mathrm{with} \quad \mathbf{z}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \mathbf{I} \right) $$ where \(\mathbf{z}_{i, t}\) is a vector of independent standard normal random variables and \( \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right) \left( \boldsymbol{\Theta}^{\frac{1}{2}} \right)^{\prime} = \boldsymbol{\Theta} . \)

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \left( \boldsymbol{\iota} + \boldsymbol{\Phi} \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$ where \(\boldsymbol{\iota}\) is a term which is unobserved and constant over time, \(\boldsymbol{\Phi}\) is the drift matrix which represents the rate of change of the solution in the absence of any random fluctuations, \(\boldsymbol{\Sigma}\) is the matrix of volatility or randomness in the process, and \(\mathrm{d}\boldsymbol{W}\) is a Wiener process or Brownian motion, which represents random fluctuations.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(9),
  R = 5
)
#> [[1]]
#>         x       m       y
#> x -0.1328 -0.6306 -0.0228
#> m  0.4928 -0.3738  0.2645
#> y -0.6151  0.2317 -0.4186
#> 
#> [[2]]
#>         x       m       y
#> x -0.6116 -0.1985  0.1571
#> m  1.1785 -0.5423 -0.2832
#> y -0.3714  0.2083 -0.5960
#> 
#> [[3]]
#>         x       m       y
#> x -0.4768  0.4630 -0.0769
#> m  0.9833 -0.3125 -0.3364
#> y -0.2895  0.8842 -0.8909
#> 
#> [[4]]
#>         x       m       y
#> x -0.0155  0.2122 -0.1556
#> m  1.2966 -0.0058  0.0401
#> y -0.3275  0.3899 -0.1681
#> 
#> [[5]]
#>         x       m       y
#> x -0.2665  0.2610 -0.0311
#> m -0.0711 -0.5450 -0.0695
#> y -0.0197  0.8798 -0.5721
#> 
phi <- matrix(
  data = c(
    -6, 5.5, 0, 0,
    1.25, -2.5, 5.9, -7.3,
    0, 0, -6, 2.5,
    5, 0, 0, -6
  ),
  nrow = 4
)
colnames(phi) <- rownames(phi) <- paste0("y", 1:4)
MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(16),
  R = 5,
  test_phi = FALSE
)
#> [[1]]
#>         y1      y2      y3      y4
#> y1 -5.7689  0.9941  0.0473  5.0511
#> y2  5.6979 -2.5707  0.1206  0.5016
#> y3  0.5322  5.6124 -6.3756 -0.6574
#> y4  0.7650 -7.1326  3.0002 -5.4661
#> 
#> [[2]]
#>         y1      y2      y3      y4
#> y1 -5.9173  1.7696  0.0596  5.3462
#> y2  5.6147 -2.7492  0.4017 -0.0841
#> y3  0.3568  6.0925 -6.1862  0.3037
#> y4  0.1117 -7.6432  2.3806 -5.7350
#> 
#> [[3]]
#>         y1      y2      y3      y4
#> y1 -5.9954  1.3963  0.1346  5.3604
#> y2  5.6429 -2.6685  0.3964 -0.4384
#> y3 -0.1424  5.7668 -6.2435 -0.2612
#> y4 -0.3370 -7.6575  2.6220 -5.5183
#> 
#> [[4]]
#>         y1      y2      y3      y4
#> y1 -5.9330  0.8672  0.3714  5.2092
#> y2  5.6444 -2.4235  0.1046  0.1446
#> y3 -0.3110  6.1756 -5.5716 -0.1052
#> y4  0.7281 -7.2346  2.3430 -6.2242
#> 
#> [[5]]
#>         y1      y2      y3      y4
#> y1 -6.0204  1.2935  0.1096  5.3630
#> y2  5.5154 -2.4907  0.2006  0.0411
#> y3 -0.0454  5.5296 -6.2657  0.3409
#> y4 -0.1691 -7.2591  3.1269 -5.9858
#>