Generate Random Drift Matrices Using the Monte Carlo Method

This function generates random drift matrices \(\boldsymbol{\Phi}\) using the Monte Carlo method.

Usage

MCPhi(phi, vcov_phi_vec, R, test_phi = TRUE, ncores = NULL, seed = NULL)

Arguments

phi

Numeric matrix. The drift matrix (\(\boldsymbol{\Phi}\)). phi should have row and column names pertaining to the variables in the system.

vcov_phi_vec

Numeric matrix. The sampling variance-covariance matrix of \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\).

R

Positive integer. Number of replications.

test_phi

Logical. If test_phi = TRUE, the function tests the stability of the generated drift matrix \(\boldsymbol{\Phi}\). If the test returns FALSE, the function generates a new drift matrix \(\boldsymbol{\Phi}\) and runs the test recursively until the test returns TRUE.

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of replications R is a large value.

seed

Random seed.

Value

Returns an object of class ctmedmc which is a list with the following elements:

call

Function call.

args

Function arguments.

fun

Function used ("MCPhi").

output

A list simulated drift matrices.

Details

Monte Carlo Method

Let \(\boldsymbol{\theta}\) be \(\mathrm{vec} \left( \boldsymbol{\Phi} \right)\), that is, the elements of the \(\boldsymbol{\Phi}\) matrix in vector form sorted column-wise. Let \(\hat{\boldsymbol{\theta}}\) be \(\mathrm{vec} \left( \hat{\boldsymbol{\Phi}} \right)\). Based on the asymptotic properties of maximum likelihood estimators, we can assume that estimators are normally distributed around the population parameters. $$ \hat{\boldsymbol{\theta}} \sim \mathcal{N} \left( \boldsymbol{\theta}, \mathbb{V} \left( \hat{\boldsymbol{\theta}} \right) \right) $$ Using this distributional assumption, a sampling distribution of \(\hat{\boldsymbol{\theta}}\) which we refer to as \(\hat{\boldsymbol{\theta}}^{\ast}\) can be generated by replacing the population parameters with sample estimates, that is, $$ \hat{\boldsymbol{\theta}}^{\ast} \sim \mathcal{N} \left( \hat{\boldsymbol{\theta}}, \hat{\mathbb{V}} \left( \hat{\boldsymbol{\theta}} \right) \right) . $$

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(9),
  R = 100L # use a large value for R in actual research
)
#> [[1]]
#>         x       m       y
#> x -0.3295  0.2220 -0.2679
#> m  0.7555 -0.8565  0.0091
#> y -0.4690  1.5626 -0.7618
#> 
#> [[2]]
#>         x       m       y
#> x -0.3830  0.0943 -0.4144
#> m  0.9453 -0.4066 -0.3812
#> y -0.1887  0.7553 -0.8437
#> 
#> [[3]]
#>         x       m      y
#> x -0.1077 -0.0344  0.001
#> m  0.4363 -0.7873 -0.033
#> y -0.9239  0.9293 -0.606
#> 
#> [[4]]
#>         x       m       y
#> x -0.7750  0.1073 -0.1698
#> m  1.2028 -0.5626 -0.0798
#> y  0.0148  0.6496 -0.5988
#> 
#> [[5]]
#>         x       m       y
#> x -1.0155  0.2169 -0.4148
#> m  0.8304 -0.3326 -0.1625
#> y -0.0059  0.3293 -0.4056
#> 
#> [[6]]
#>         x       m       y
#> x -0.5023 -0.3002  0.2227
#> m  0.5239 -0.0857  0.3440
#> y -0.0757  0.5680 -0.9511
#> 
#> [[7]]
#>         x       m       y
#> x -0.1279 -0.4010 -0.1586
#> m  0.7976 -0.4652 -0.6030
#> y -0.3153  0.3125 -0.7378
#> 
#> [[8]]
#>         x       m       y
#> x -0.6284  0.2739 -0.1972
#> m  0.4140 -0.7338  0.1835
#> y -0.1817  0.6000 -0.6670
#> 
#> [[9]]
#>         x       m       y
#> x -0.3203 -0.4097 -0.1738
#> m  1.0749 -0.4948  0.2990
#> y -0.4625  0.9025 -0.8753
#> 
#> [[10]]
#>         x       m       y
#> x -0.4946 -0.6680  0.8914
#> m  0.4559 -0.7414 -0.3237
#> y  0.0138  0.7617 -0.7599
#> 
#> [[11]]
#>         x       m       y
#> x -0.3574 -0.3296 -0.3109
#> m  0.9960 -1.0598  0.0318
#> y -0.6253  0.8023 -0.3250
#> 
#> [[12]]
#>         x       m       y
#> x -0.2388  0.1046  0.1173
#> m  0.5637 -0.5110 -0.0937
#> y -0.4943  0.3738 -0.8324
#> 
#> [[13]]
#>         x       m       y
#> x -0.4017 -0.1042 -0.1136
#> m  0.4225 -1.0385 -0.7173
#> y -0.2042  0.4482 -0.7084
#> 
#> [[14]]
#>         x       m       y
#> x -0.5168 -0.1229  0.2009
#> m  0.4649 -0.5269  0.1268
#> y -0.4530  0.6824 -1.1855
#> 
#> [[15]]
#>         x       m       y
#> x -0.4401  0.1808 -0.1407
#> m  0.5675 -0.7721 -0.0552
#> y  0.0013 -0.1627 -0.8033
#> 
#> [[16]]
#>         x      m       y
#> x -0.7018 -0.287 -0.4060
#> m  0.4014 -0.926 -0.6191
#> y -0.4522  1.042 -0.9174
#> 
#> [[17]]
#>         x       m       y
#> x -0.2428 -0.1850  0.3692
#> m  0.1528 -0.5686 -0.1734
#> y -0.1697  0.4695 -0.5461
#> 
#> [[18]]
#>         x       m       y
#> x -0.8570 -0.0544 -0.0911
#> m  1.4080 -0.4738 -1.0594
#> y  0.0874  0.8196 -0.4478
#> 
#> [[19]]
#>         x       m       y
#> x -0.4047 -0.2895  0.1201
#> m  0.7085 -0.7802  0.3837
#> y  0.3360  0.2906 -0.3345
#> 
#> [[20]]
#>         x       m       y
#> x -0.1396 -0.0238 -0.1357
#> m  0.3961 -0.4810  0.0038
#> y  0.1203  0.4805 -0.1099
#> 
#> [[21]]
#>         x       m       y
#> x -0.5919  0.0691 -0.0411
#> m  0.7065 -0.8078 -0.0403
#> y -0.7677  0.9824 -0.9226
#> 
#> [[22]]
#>         x       m       y
#> x -0.3922 -0.0144 -0.1370
#> m  0.5602 -0.4864 -0.6544
#> y -0.7721  0.4842 -0.0734
#> 
#> [[23]]
#>         x       m       y
#> x -0.0552 -0.2630 -0.2973
#> m  1.6914 -0.6528 -0.0127
#> y -0.4832  0.2784 -0.6221
#> 
#> [[24]]
#>         x       m       y
#> x -0.4761 -0.1023  0.0572
#> m  1.0788 -0.6270 -0.1512
#> y -1.0433  0.8290 -1.1084
#> 
#> [[25]]
#>         x       m       y
#> x -0.5386  0.3221  0.0621
#> m  0.4560 -0.7578  0.2615
#> y -0.5445 -0.2587 -0.8716
#> 
#> [[26]]
#>         x       m       y
#> x -0.8226 -0.0375  0.1275
#> m  0.9053 -0.3728 -0.0042
#> y -0.1520  0.6408 -0.3339
#> 
#> [[27]]
#>         x       m       y
#> x -0.3714 -0.2501  0.0273
#> m  0.3441 -0.5902  0.4339
#> y -0.1173  0.6587 -0.6179
#> 
#> [[28]]
#>         x       m       y
#> x -0.2288  0.1370  0.0740
#> m  0.6924 -0.8384 -0.6592
#> y -0.6282  0.4929 -0.9160
#> 
#> [[29]]
#>         x       m       y
#> x -0.9407  0.4237  0.1673
#> m  1.0778 -0.5615 -0.1271
#> y -0.1173  0.1810 -0.3245
#> 
#> [[30]]
#>         x       m       y
#> x -0.4081  0.0028  0.0232
#> m  0.4974 -0.4713 -0.4130
#> y -0.4946  0.2607 -0.5754
#> 
#> [[31]]
#>         x       m       y
#> x -0.1392 -0.4151  0.1329
#> m  0.4121 -0.3814 -0.0712
#> y -0.8051  0.4695 -0.5703
#> 
#> [[32]]
#>         x       m       y
#> x -1.0668 -0.3412 -0.0488
#> m  1.5288 -0.5792 -0.6432
#> y -0.0236  0.6782 -0.9510
#> 
#> [[33]]
#>         x       m       y
#> x -0.6313  0.1458  0.5649
#> m  1.0035 -1.0179 -0.0652
#> y -0.7079  0.5710 -0.9275
#> 
#> [[34]]
#>         x       m       y
#> x -0.5981  0.5447  0.2370
#> m  0.5518 -0.8220  0.1444
#> y -0.2837  0.7441 -0.8026
#> 
#> [[35]]
#>         x       m       y
#> x -0.3760 -0.6159 -0.1619
#> m  0.7455 -0.4486  0.3184
#> y -0.5399  0.8208 -0.9259
#> 
#> [[36]]
#>         x       m       y
#> x -0.8211  0.3333  0.5403
#> m  0.4818 -0.6742  0.1147
#> y -0.4575  1.0438 -0.6069
#> 
#> [[37]]
#>         x       m       y
#> x -0.0438 -0.1913 -0.2189
#> m  1.0121 -1.0879  0.4560
#> y -0.7420  0.9661 -0.4182
#> 
#> [[38]]
#>         x       m       y
#> x -0.6305  0.0519 -0.3075
#> m  0.2891 -0.8919 -0.1079
#> y -0.5822  0.8868 -0.8725
#> 
#> [[39]]
#>         x       m       y
#> x -0.9969  0.6508 -0.2677
#> m  0.4777 -0.5442 -0.0148
#> y -0.8395  0.8875 -1.0651
#> 
#> [[40]]
#>         x       m       y
#> x -0.4153 -0.3777  0.0552
#> m  0.6768 -0.7006 -0.3646
#> y -0.5796  0.6035 -0.3262
#> 
#> [[41]]
#>         x       m       y
#> x -0.4109 -0.2373 -0.0544
#> m  0.5736 -0.9596 -0.2123
#> y -0.3284  1.0132 -0.3790
#> 
#> [[42]]
#>         x       m       y
#> x -0.1467 -0.3364 -0.4359
#> m  0.2769 -0.4348  0.0457
#> y -0.5458  0.6445 -0.6675
#> 
#> [[43]]
#>         x       m       y
#> x -0.3225  0.0097  0.4471
#> m  0.2850 -0.5174  0.2141
#> y -1.0253  1.4949 -0.8394
#> 
#> [[44]]
#>         x       m       y
#> x -0.7559 -0.1536  0.4511
#> m  0.9942 -0.9598  0.5648
#> y -0.4914  0.5642 -1.0631
#> 
#> [[45]]
#>         x       m       y
#> x -0.1769  0.0374 -0.4160
#> m  0.9637 -0.4728 -0.2756
#> y -0.6885  0.3754 -0.5568
#> 
#> [[46]]
#>         x       m       y
#> x -0.3169 -0.1641  0.0323
#> m  1.2726 -0.4854  0.2306
#> y -0.9982  0.6590 -0.7262
#> 
#> [[47]]
#>         x       m       y
#> x -0.5678  0.2690  0.0708
#> m  0.4662 -0.5777  0.0837
#> y -0.5875  0.0619 -0.9957
#> 
#> [[48]]
#>         x       m       y
#> x -0.9053 -0.4573 -0.0542
#> m  0.6444 -0.5776 -0.2122
#> y -0.0040  0.9637 -0.3440
#> 
#> [[49]]
#>         x       m       y
#> x -0.3641  0.2027  0.0351
#> m  0.7360 -0.7475 -0.3221
#> y -0.4191  0.3433 -1.0359
#> 
#> [[50]]
#>         x       m       y
#> x -0.1074 -0.1336 -0.3040
#> m  1.0164 -0.4057 -0.1125
#> y -0.5545  1.0478 -1.0712
#> 
#> [[51]]
#>         x       m       y
#> x -0.3634 -0.0708 -0.3003
#> m  0.9753 -0.7200  1.1534
#> y -0.4025  0.9655 -0.8761
#> 
#> [[52]]
#>         x       m       y
#> x -0.4116  0.1819  0.0896
#> m  0.2703 -0.4336  0.1018
#> y -0.5597  0.6048 -0.6692
#> 
#> [[53]]
#>         x       m       y
#> x -0.9718 -0.1154 -0.1710
#> m  1.2792 -0.6169 -0.1020
#> y -0.5817  0.8355 -0.3922
#> 
#> [[54]]
#>         x       m       y
#> x -0.3671 -0.0507  0.2011
#> m  0.4971 -0.9741 -0.2671
#> y -0.3027  1.1691 -0.6000
#> 
#> [[55]]
#>         x       m       y
#> x -0.9081 -0.2732 -0.1377
#> m  0.7826 -0.2509  0.0539
#> y -0.4285  0.7691 -0.6318
#> 
#> [[56]]
#>         x       m       y
#> x -0.6330 -0.1352  0.4032
#> m  1.3981 -0.4071 -0.4848
#> y -0.0628  0.8709 -0.6484
#> 
#> [[57]]
#>         x       m       y
#> x -0.5716 -0.2904 -0.9815
#> m  0.6002 -0.5104  0.3351
#> y -0.5145  0.9666 -0.5926
#> 
#> [[58]]
#>         x       m       y
#> x -0.0853 -0.1908  0.5859
#> m -0.0739 -0.7610 -0.2784
#> y -1.0674  0.6428 -0.8445
#> 
#> [[59]]
#>         x       m       y
#> x -0.3777  0.0656  0.1128
#> m  0.8989 -0.5156 -0.4871
#> y -0.4501  0.3930 -0.4726
#> 
#> [[60]]
#>         x       m       y
#> x -1.1444  0.2666 -0.0877
#> m  0.2862 -0.3202 -0.2117
#> y -0.5150  1.3012 -0.5467
#> 
#> [[61]]
#>         x       m       y
#> x -0.5269  0.0676 -0.0710
#> m  0.7264 -1.0256  0.0231
#> y -0.0178  0.4719 -0.4232
#> 
#> [[62]]
#>         x       m       y
#> x -0.2998 -0.1015  0.1827
#> m  0.4915 -0.6801  0.3549
#> y -0.0945  0.9665 -0.9195
#> 
#> [[63]]
#>         x       m       y
#> x -0.0936 -0.1392  0.3097
#> m  1.0521 -0.4798 -0.4324
#> y -0.5027  0.6471 -0.7498
#> 
#> [[64]]
#>         x       m       y
#> x -0.6689 -0.0929 -0.2490
#> m  1.0424 -0.1300  0.1806
#> y -0.9117  1.0659 -1.1202
#> 
#> [[65]]
#>         x       m       y
#> x -0.4898 -0.1196  0.4541
#> m  0.6970 -0.1351 -0.3419
#> y -0.3528  0.1951 -0.7034
#> 
#> [[66]]
#>         x       m       y
#> x -0.2360 -0.0164 -0.3757
#> m  0.9830 -0.6066 -0.2857
#> y -0.1591  0.7775 -1.0288
#> 
#> [[67]]
#>         x       m       y
#> x -0.6339  0.0342  0.0401
#> m  0.8270 -1.0489  0.4998
#> y  0.0652  0.5969 -0.5418
#> 
#> [[68]]
#>         x       m       y
#> x -0.5422 -0.0616  0.0014
#> m  0.6718 -0.3817  0.4155
#> y -0.7427  0.3234 -0.9757
#> 
#> [[69]]
#>         x       m       y
#> x -0.0403 -0.0506 -0.5266
#> m  0.7176 -0.8780 -0.1823
#> y -0.4755  1.0920 -0.7177
#> 
#> [[70]]
#>         x       m       y
#> x -0.4184  0.0873  0.2936
#> m  0.5819 -0.5672 -0.2931
#> y -0.5228  0.6846 -1.1257
#> 
#> [[71]]
#>         x       m       y
#> x -0.6265  0.2090 -0.1042
#> m  0.7528 -0.9778  0.0057
#> y -0.5539  0.7630 -0.0298
#> 
#> [[72]]
#>         x       m       y
#> x -0.3142 -0.4175 -0.3436
#> m  0.6047 -0.6731  0.4674
#> y -0.0897  0.5923 -1.0471
#> 
#> [[73]]
#>         x       m       y
#> x -0.7625 -0.1079 -0.1096
#> m  0.5141 -0.8554  0.3487
#> y -0.6520  0.6899 -1.4064
#> 
#> [[74]]
#>         x       m       y
#> x -0.3623 -0.2808 -0.0804
#> m  0.8808 -0.4030  0.1781
#> y -0.6301  0.8759 -1.0519
#> 
#> [[75]]
#>         x       m       y
#> x -0.5995 -0.4198 -0.5094
#> m  0.6576 -0.2874  0.3968
#> y -0.2405  0.7774 -0.6126
#> 
#> [[76]]
#>         x       m       y
#> x -0.2502  0.0400 -0.2393
#> m  1.0703 -0.0552 -0.2317
#> y -0.2040  0.5293 -0.4795
#> 
#> [[77]]
#>         x       m       y
#> x -0.3583  0.4105 -0.3260
#> m  0.4468 -0.7572 -0.4285
#> y  0.0135  0.4630 -0.4630
#> 
#> [[78]]
#>         x       m       y
#> x -0.6778  0.0109 -0.5704
#> m  0.8244 -0.3219 -0.1490
#> y -0.5195  0.4863 -1.0459
#> 
#> [[79]]
#>         x       m       y
#> x -0.1261 -0.2689  0.1719
#> m  1.2347 -0.8141 -0.5353
#> y -0.5857  1.1799 -0.7561
#> 
#> [[80]]
#>         x       m       y
#> x -0.2520 -0.4989  0.4904
#> m  0.5242 -0.4786 -0.2059
#> y -0.3187  0.9093 -0.7219
#> 
#> [[81]]
#>         x       m       y
#> x -0.4483  0.5324 -0.0998
#> m  0.1837 -0.3331  0.2339
#> y -0.3454  0.9350 -1.0312
#> 
#> [[82]]
#>         x       m       y
#> x -0.4740  0.4396 -0.0832
#> m  0.6544 -0.7612  0.3754
#> y -0.2096  0.5643 -0.9836
#> 
#> [[83]]
#>         x       m       y
#> x -0.8703 -0.2334  0.4392
#> m  1.4877 -0.6639 -0.1450
#> y -0.2625  0.8243 -0.8017
#> 
#> [[84]]
#>         x       m       y
#> x -0.1962 -0.3202  0.0654
#> m  1.3910 -0.3775 -0.3593
#> y -0.5640  0.3310 -0.3806
#> 
#> [[85]]
#>         x       m       y
#> x -0.3905 -0.2777 -0.3170
#> m  0.3367 -0.6976  0.0301
#> y  0.6498  0.8609 -0.6813
#> 
#> [[86]]
#>         x       m       y
#> x -0.6248 -0.4520  0.2134
#> m  0.8028 -0.2845  0.0879
#> y -0.9383  0.9861 -0.6502
#> 
#> [[87]]
#>         x       m       y
#> x -0.1593 -0.5304 -0.0042
#> m  0.9188 -0.6926 -0.0764
#> y -0.2094  0.5743 -0.4895
#> 
#> [[88]]
#>         x       m       y
#> x -0.4309 -0.1042 -0.0629
#> m  0.7116 -0.6306  0.1055
#> y -0.4704  0.3523 -0.9833
#> 
#> [[89]]
#>         x       m       y
#> x -0.1073  0.0902  0.3519
#> m  0.0455 -0.5125  0.2619
#> y -0.4501  0.8671 -0.6955
#> 
#> [[90]]
#>         x       m       y
#> x -0.1855 -0.1612 -0.1172
#> m  0.9788 -0.8409 -0.3104
#> y -0.5546  0.6565 -0.6502
#> 
#> [[91]]
#>         x       m       y
#> x -0.2742 -0.0278 -0.0435
#> m  0.3944 -0.2534 -0.2422
#> y  0.6082  0.6999 -0.5411
#> 
#> [[92]]
#>         x       m       y
#> x -0.8169  0.2507  0.1256
#> m  0.2168 -0.8083  0.4825
#> y  0.0757  0.5949 -0.5242
#> 
#> [[93]]
#>         x       m       y
#> x -0.3167  0.0613 -0.1022
#> m  0.6812 -0.6474 -0.5121
#> y -0.3802  0.8593 -0.4606
#> 
#> [[94]]
#>         x       m       y
#> x -0.4927  0.0970  0.1964
#> m  0.6259 -0.5075  0.2843
#> y -0.2871  0.1613 -1.0010
#> 
#> [[95]]
#>         x       m       y
#> x -0.2908 -0.0831 -0.3284
#> m  0.7124 -0.3808  0.0491
#> y -0.4735  0.7616 -0.4224
#> 
#> [[96]]
#>         x       m       y
#> x -0.2382 -0.0210  0.2435
#> m  0.8107 -0.9013  0.0443
#> y -0.5787  0.8863 -1.0013
#> 
#> [[97]]
#>         x       m       y
#> x -0.3714 -0.2303 -0.3200
#> m  1.2856 -0.9282  0.2868
#> y -0.1415  1.1378 -0.8186
#> 
#> [[98]]
#>         x       m       y
#> x -0.2475 -0.4155  0.3497
#> m  0.6953 -0.4973  0.1498
#> y -0.3532  0.7291 -0.4281
#> 
#> [[99]]
#>         x       m       y
#> x -0.2293 -0.2961  0.2112
#> m  0.6084 -0.7606  0.1396
#> y -0.7304  0.3160 -0.4800
#> 
#> [[100]]
#>         x       m       y
#> x -0.5496 -0.0348  0.0700
#> m  0.9226 -0.5533 -0.0088
#> y -0.5031  0.8429 -0.5873
#> 
phi <- matrix(
  data = c(
    -6, 5.5, 0, 0,
    1.25, -2.5, 5.9, -7.3,
    0, 0, -6, 2.5,
    5, 0, 0, -6
  ),
  nrow = 4
)
colnames(phi) <- rownames(phi) <- paste0("y", 1:4)
MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(16),
  R = 100L, # use a large value for R in actual research
  test_phi = FALSE
)
#> [[1]]
#>         y1      y2      y3      y4
#> y1 -5.7668  0.9230  0.2233  4.2634
#> y2  5.6636 -2.6310  0.5893  0.1656
#> y3  0.1348  5.7811 -6.3375 -0.0939
#> y4  0.3147 -7.4983  2.6865 -6.3097
#> 
#> [[2]]
#>         y1      y2      y3      y4
#> y1 -5.6895  1.3958 -0.1064  4.9329
#> y2  5.6879 -2.2832 -0.0758 -0.3114
#> y3 -0.5317  5.7714 -5.9755 -0.0333
#> y4 -0.2057 -7.3658  2.2075 -5.4863
#> 
#> [[3]]
#>         y1      y2      y3      y4
#> y1 -5.9580  1.5836  0.3723  4.6407
#> y2  5.5283 -2.0794  0.2199  0.0805
#> y3 -0.5804  5.7189 -6.2748  0.1087
#> y4 -0.1854 -7.1197  2.6285 -6.3340
#> 
#> [[4]]
#>         y1      y2      y3      y4
#> y1 -5.6135  1.3059 -0.1520  3.7929
#> y2  6.2636 -3.0146  0.5791 -0.4636
#> y3  0.1067  5.7501 -5.8046 -0.4455
#> y4  0.3517 -7.0425  2.5319 -5.9514
#> 
#> [[5]]
#>         y1      y2      y3      y4
#> y1 -6.1717  1.0766 -0.3349  4.7839
#> y2  5.3780 -2.5961 -0.5326 -0.2039
#> y3 -0.3308  6.3404 -6.2413  0.6843
#> y4  0.1681 -7.0827  2.7400 -6.0146
#> 
#> [[6]]
#>         y1      y2      y3      y4
#> y1 -5.3812  1.2558  0.6623  5.2363
#> y2  5.0359 -3.1022  0.0187 -0.0110
#> y3  0.4751  6.0529 -6.0927 -0.3454
#> y4  0.3612 -7.0734  2.4150 -6.4608
#> 
#> [[7]]
#>         y1      y2      y3      y4
#> y1 -6.0145  0.9127 -0.1050  5.0986
#> y2  5.7418 -2.8818 -0.3098  0.1868
#> y3  0.0207  5.8365 -5.9757 -0.1299
#> y4  0.2673 -7.3022  2.9458 -6.2110
#> 
#> [[8]]
#>         y1      y2      y3      y4
#> y1 -5.8044  1.0844  0.1378  4.9405
#> y2  5.5657 -2.3141 -0.4056 -0.3457
#> y3  0.2661  5.9458 -5.6681 -0.1354
#> y4 -0.5021 -7.0220  2.6626 -5.6208
#> 
#> [[9]]
#>         y1      y2      y3      y4
#> y1 -6.5522  1.3479  0.3668  4.7705
#> y2  5.4461 -2.9554 -0.1138  0.2969
#> y3 -0.2456  5.4209 -5.9634 -0.1664
#> y4  0.2009 -6.9219  2.9102 -6.1279
#> 
#> [[10]]
#>         y1      y2      y3      y4
#> y1 -5.9269  1.3184  0.0610  4.7207
#> y2  5.7802 -2.7238 -0.0001 -0.1507
#> y3  0.6121  5.6427 -6.2988 -0.3708
#> y4 -0.2834 -7.9467  2.7768 -5.6733
#> 
#> [[11]]
#>         y1      y2      y3      y4
#> y1 -6.0108  1.6800 -0.2495  4.5300
#> y2  5.3815 -2.8865 -0.1977  0.1310
#> y3  0.3771  5.5380 -5.5998  0.0828
#> y4 -0.2887 -6.5693  2.5902 -6.1295
#> 
#> [[12]]
#>         y1      y2      y3      y4
#> y1 -5.8948  2.1404  0.5304  4.8120
#> y2  5.2657 -2.7172  0.2243  0.2726
#> y3  0.3392  6.2523 -6.4579  0.5665
#> y4 -0.0974 -7.1006  3.0208 -5.9235
#> 
#> [[13]]
#>         y1      y2      y3      y4
#> y1 -6.3194  1.2196  0.2219  4.5833
#> y2  5.6737 -2.2696  0.5674 -0.1532
#> y3 -0.6263  5.6929 -6.1867  0.5924
#> y4  0.5117 -6.6057  2.8267 -5.7002
#> 
#> [[14]]
#>         y1      y2      y3      y4
#> y1 -5.4823  1.1359 -0.5189  4.7700
#> y2  6.0158 -2.4379  0.1708 -0.3368
#> y3  0.2634  6.1350 -6.1545 -0.1257
#> y4  0.2983 -7.1599  2.4814 -5.8683
#> 
#> [[15]]
#>         y1      y2      y3      y4
#> y1 -6.6469  1.6101  0.0410  4.7742
#> y2  5.3425 -2.5490  0.2587 -0.0487
#> y3 -0.0882  6.2703 -5.8678  0.1974
#> y4  0.3573 -7.0197  2.0987 -5.7518
#> 
#> [[16]]
#>         y1      y2      y3      y4
#> y1 -6.1222  1.2925 -0.0562  4.6711
#> y2  5.6139 -2.3581 -0.0177 -0.3112
#> y3  0.0130  5.6994 -6.3263  0.4808
#> y4  0.3061 -7.7456  2.2897 -6.1503
#> 
#> [[17]]
#>         y1      y2      y3      y4
#> y1 -5.9608  1.4888  0.0045  5.1459
#> y2  5.5950 -1.9211 -0.1746 -0.2890
#> y3  0.0078  5.8701 -5.8714 -0.1302
#> y4 -0.4365 -7.1767  2.8803 -5.8976
#> 
#> [[18]]
#>         y1      y2      y3      y4
#> y1 -5.8132  1.1155 -0.2077  4.7644
#> y2  5.5631 -2.8843 -0.0421 -0.3663
#> y3  0.5158  5.6960 -6.5660 -0.7380
#> y4 -0.0856 -7.5594  2.2929 -6.0541
#> 
#> [[19]]
#>         y1      y2      y3      y4
#> y1 -5.8361  1.6981 -0.2099  5.1071
#> y2  5.8118 -2.6273  0.1838  0.2969
#> y3  0.2353  6.0135 -6.2679  0.1417
#> y4  0.2306 -6.8748  2.5869 -5.7116
#> 
#> [[20]]
#>         y1      y2      y3      y4
#> y1 -6.1133  1.2747  0.2049  4.9287
#> y2  5.3474 -2.4081  0.1822 -0.0050
#> y3  0.0505  6.2017 -5.5979  0.1538
#> y4  0.6316 -7.7229  2.5705 -5.7057
#> 
#> [[21]]
#>         y1      y2      y3      y4
#> y1 -6.3539  1.4517 -0.0312  5.0953
#> y2  5.1892 -2.3262 -0.4178  0.0068
#> y3 -0.3209  5.9855 -6.0231  0.0657
#> y4  0.0838 -7.5616  2.6139 -5.8151
#> 
#> [[22]]
#>         y1      y2      y3      y4
#> y1 -5.3427  1.6514  0.3106  4.5388
#> y2  5.6862 -2.8037 -0.0708 -0.7003
#> y3  0.0372  5.6301 -6.3433  0.0265
#> y4  0.1000 -7.7933  2.7021 -6.1851
#> 
#> [[23]]
#>         y1      y2      y3      y4
#> y1 -5.6847  1.0275  0.5207  4.9294
#> y2  5.9071 -2.4629 -0.2393  0.4546
#> y3  0.5209  5.9664 -5.9180 -0.1018
#> y4  0.1192 -8.2784  2.4460 -5.6730
#> 
#> [[24]]
#>         y1      y2      y3      y4
#> y1 -6.0209  1.0396 -0.1932  4.5331
#> y2  5.1288 -2.1827 -0.0835  0.2775
#> y3  0.0706  6.0495 -6.0031 -0.1035
#> y4 -0.3772 -7.1062  2.0684 -6.1947
#> 
#> [[25]]
#>         y1      y2      y3      y4
#> y1 -5.7677  1.2751  0.1769  5.0236
#> y2  5.6191 -1.9918  0.0119  0.4368
#> y3 -0.2083  5.3403 -6.4606  0.2666
#> y4 -0.4714 -7.6268  2.8729 -6.0815
#> 
#> [[26]]
#>         y1      y2      y3      y4
#> y1 -6.1515  1.4525  0.2027  4.6271
#> y2  5.4472 -2.8296  0.0494 -0.0455
#> y3 -0.4805  5.5610 -6.6059 -0.7099
#> y4  0.5108 -7.5901  2.7159 -5.6216
#> 
#> [[27]]
#>         y1      y2      y3      y4
#> y1 -6.0670  1.3287 -0.0973  5.0208
#> y2  6.2865 -2.8627 -0.1693 -0.5216
#> y3 -0.0158  5.6777 -5.7077  0.2599
#> y4  0.0596 -6.6147  3.0623 -5.9546
#> 
#> [[28]]
#>         y1      y2      y3      y4
#> y1 -5.2771  1.3007  0.3736  4.5974
#> y2  5.7214 -3.1083  0.4978 -0.1548
#> y3  0.2311  5.4551 -5.7353 -0.0374
#> y4  0.2129 -7.3875  2.9406 -6.6204
#> 
#> [[29]]
#>         y1      y2      y3      y4
#> y1 -6.3433  0.8689  0.0856  5.2539
#> y2  5.4734 -2.3393  0.2726  0.2675
#> y3 -0.0526  5.7604 -6.0303 -0.0223
#> y4 -0.2449 -6.5926  2.7770 -6.1566
#> 
#> [[30]]
#>         y1      y2      y3      y4
#> y1 -5.7536  1.3618  0.3328  4.9249
#> y2  5.3979 -2.3363 -0.0817  0.4274
#> y3  0.0238  5.7449 -5.6317 -0.1724
#> y4 -0.1627 -7.4029  2.1575 -6.6403
#> 
#> [[31]]
#>         y1      y2      y3      y4
#> y1 -6.2718  1.3489  0.3620  4.9804
#> y2  6.0443 -2.5327 -0.1802 -0.5506
#> y3  0.0275  6.1188 -6.0708 -0.0636
#> y4 -0.3418 -7.5190  2.5217 -6.1727
#> 
#> [[32]]
#>         y1      y2      y3      y4
#> y1 -6.2242  1.4407 -0.0388  4.9602
#> y2  5.8011 -2.4977 -0.1438  0.0753
#> y3 -0.4079  6.0727 -6.0324 -0.3399
#> y4 -0.5544 -7.0974  2.2145 -5.9584
#> 
#> [[33]]
#>         y1      y2      y3      y4
#> y1 -6.3039  0.8956 -0.1184  5.5026
#> y2  5.4044 -2.6264 -0.2561  0.0758
#> y3  0.0288  5.4258 -6.5411 -0.0384
#> y4 -0.0572 -7.4542  2.8864 -5.9262
#> 
#> [[34]]
#>         y1      y2      y3      y4
#> y1 -5.9672  1.1480 -0.0393  4.6688
#> y2  5.8418 -2.6531 -0.0048  0.4096
#> y3  0.3215  5.8236 -5.6790 -0.2734
#> y4  0.2377 -7.5154  2.2944 -6.6422
#> 
#> [[35]]
#>         y1      y2      y3      y4
#> y1 -5.7728  1.5450 -0.0139  5.5630
#> y2  5.7372 -2.1377  0.2431 -0.2941
#> y3 -0.2750  6.4422 -6.5744 -0.3441
#> y4 -0.3499 -7.4563  2.3695 -5.8145
#> 
#> [[36]]
#>         y1      y2      y3      y4
#> y1 -6.1589  1.1369 -0.1308  5.0856
#> y2  5.7150 -2.2752 -0.0253  0.8980
#> y3 -0.0521  6.1190 -5.9389 -0.0729
#> y4  0.3482 -6.6299  2.1511 -5.9658
#> 
#> [[37]]
#>         y1      y2      y3      y4
#> y1 -5.8945  0.9853  0.6941  4.9666
#> y2  5.1616 -3.0345  0.0107 -0.0917
#> y3  0.0343  5.8754 -5.5661  0.3381
#> y4  0.4594 -6.6046  2.0610 -5.9833
#> 
#> [[38]]
#>         y1      y2      y3      y4
#> y1 -6.0137  1.4075 -0.2587  4.7113
#> y2  5.1521 -2.3546 -0.1579  0.0894
#> y3 -0.0882  5.7019 -6.4509 -0.1253
#> y4 -0.4616 -7.0690  2.6749 -5.9659
#> 
#> [[39]]
#>         y1      y2      y3      y4
#> y1 -5.2570  1.5659  0.1064  4.6079
#> y2  5.4045 -2.2201 -0.2555 -0.0682
#> y3 -0.0509  5.8983 -6.4767  0.3584
#> y4 -0.3301 -7.8675  2.8030 -6.2076
#> 
#> [[40]]
#>         y1      y2      y3      y4
#> y1 -5.9080  1.3928 -0.1253  5.3514
#> y2  5.3735 -2.3501 -0.2915  0.0394
#> y3 -0.0680  5.8674 -6.0316 -0.0848
#> y4 -0.0270 -7.7892  2.2404 -6.3040
#> 
#> [[41]]
#>         y1      y2      y3      y4
#> y1 -6.4446  1.3247 -0.2761  4.5362
#> y2  5.7358 -2.1600 -0.5047 -0.6031
#> y3 -0.6084  5.5772 -5.6569  0.6039
#> y4  0.8787 -7.6333  2.4308 -5.5574
#> 
#> [[42]]
#>         y1      y2      y3      y4
#> y1 -5.8593  1.2839 -0.0032  4.7804
#> y2  5.8267 -2.5814 -0.0681  0.4127
#> y3  0.6440  5.9321 -5.8832 -0.1229
#> y4  0.1774 -6.8543  3.1020 -6.5602
#> 
#> [[43]]
#>         y1      y2      y3      y4
#> y1 -6.0673  0.8908  0.2522  5.3146
#> y2  5.5606 -2.4939  0.3249  0.3578
#> y3 -0.1531  5.7704 -5.2237  0.3980
#> y4  0.4033 -7.0085  2.2838 -6.7198
#> 
#> [[44]]
#>         y1      y2      y3      y4
#> y1 -5.8124  1.4435  0.0103  4.2477
#> y2  5.6099 -2.7047 -0.6513  0.4689
#> y3  0.0544  5.4757 -5.8748  0.2879
#> y4 -0.0655 -7.3180  3.0642 -5.7631
#> 
#> [[45]]
#>         y1      y2      y3      y4
#> y1 -5.9015  1.1435 -0.1690  4.9738
#> y2  4.9948 -2.8262  0.2239  0.0506
#> y3 -0.3837  6.4569 -6.3387  0.3022
#> y4  0.1928 -7.3322  2.5901 -6.3325
#> 
#> [[46]]
#>         y1      y2      y3      y4
#> y1 -6.0980  1.2951 -0.3674  4.6936
#> y2  5.5459 -3.1702  0.4847 -0.3380
#> y3 -0.1205  5.2876 -6.0921 -0.2235
#> y4 -0.1159 -7.3589  2.1731 -5.7440
#> 
#> [[47]]
#>         y1      y2      y3      y4
#> y1 -5.3795  0.6644  0.2127  4.3846
#> y2  5.3361 -2.5810  0.0289  0.3402
#> y3 -0.2481  5.8775 -4.8334  0.2223
#> y4 -0.4489 -7.3971  1.9999 -5.7160
#> 
#> [[48]]
#>         y1      y2      y3      y4
#> y1 -6.5926  1.3679 -0.0816  5.4116
#> y2  5.7363 -2.3326 -0.0569 -0.0734
#> y3 -0.5427  5.6793 -5.9737  0.0655
#> y4 -0.7339 -7.1181  2.5202 -5.7012
#> 
#> [[49]]
#>         y1      y2      y3      y4
#> y1 -5.7572  1.9445 -0.1988  4.9842
#> y2  5.6896 -2.7707  0.1606  0.0311
#> y3  0.3578  5.9488 -6.1304 -0.7058
#> y4 -0.0977 -7.0141  2.5897 -5.4540
#> 
#> [[50]]
#>         y1      y2      y3      y4
#> y1 -6.2746  1.0570  0.1394  4.9125
#> y2  5.6567 -2.6995  0.1489 -0.0157
#> y3  0.2149  6.1102 -6.1968  0.8091
#> y4  0.1977 -6.8883  2.3872 -5.8547
#> 
#> [[51]]
#>         y1      y2      y3      y4
#> y1 -5.9523  1.4773  0.4845  5.3901
#> y2  5.4737 -2.3223  0.4638 -0.0051
#> y3  0.1663  5.7208 -5.7078 -0.2873
#> y4 -0.2619 -7.5732  2.6116 -5.6516
#> 
#> [[52]]
#>         y1      y2      y3      y4
#> y1 -5.9569  1.1484  0.1413  4.9625
#> y2  5.4961 -1.9196 -0.2147  0.0860
#> y3  0.0434  5.1711 -6.0091 -0.4300
#> y4  0.0989 -6.9923  2.4675 -6.2801
#> 
#> [[53]]
#>         y1      y2      y3      y4
#> y1 -5.9964  1.5948  0.1424  4.8919
#> y2  4.9500 -2.3143 -0.2004  0.0746
#> y3  0.0413  6.1239 -6.1581  0.5520
#> y4 -0.1606 -6.6052  3.0729 -5.6252
#> 
#> [[54]]
#>         y1      y2      y3      y4
#> y1 -6.0056  0.5813  0.2066  4.1852
#> y2  5.6146 -2.5315 -0.2098  0.5014
#> y3 -0.0282  6.0132 -5.9744 -0.2524
#> y4  0.1977 -7.1909  2.0617 -5.5994
#> 
#> [[55]]
#>         y1      y2      y3      y4
#> y1 -6.2481  1.2533  0.0911  4.8759
#> y2  5.3307 -2.3027 -0.1617  0.5334
#> y3  0.1470  5.6411 -5.5382 -0.0052
#> y4  0.2635 -7.2861  2.4187 -6.3106
#> 
#> [[56]]
#>         y1      y2      y3      y4
#> y1 -5.9427  1.2370 -0.4303  4.2867
#> y2  5.4646 -2.7914  0.7996 -0.1051
#> y3  0.1021  6.0495 -5.6497 -0.1754
#> y4  0.2555 -7.3638  2.6484 -6.1278
#> 
#> [[57]]
#>         y1      y2      y3      y4
#> y1 -5.8707  1.4799  0.4191  5.0782
#> y2  5.6182 -2.4580 -0.0108 -0.2805
#> y3 -0.7355  5.7616 -6.2561 -0.4730
#> y4  0.0598 -7.7258  2.6587 -5.9206
#> 
#> [[58]]
#>         y1      y2      y3      y4
#> y1 -5.8946  1.0787 -0.4839  5.0753
#> y2  5.6657 -2.6659 -0.3270  0.1845
#> y3 -0.1998  5.6498 -5.6949  0.2662
#> y4 -0.2045 -7.1635  2.2299 -5.7683
#> 
#> [[59]]
#>         y1      y2      y3      y4
#> y1 -6.4531  0.9740  0.0054  4.9008
#> y2  5.8377 -2.9909 -0.2070 -0.4158
#> y3 -0.2417  5.5451 -5.8743  0.2004
#> y4 -0.5901 -7.1767  2.3551 -5.7033
#> 
#> [[60]]
#>         y1      y2      y3      y4
#> y1 -5.4631  1.0113  0.0522  5.1682
#> y2  4.9342 -2.6603 -0.3605  0.2755
#> y3 -0.4089  6.0702 -5.9465 -0.1638
#> y4 -0.2940 -7.0916  2.4145 -5.0471
#> 
#> [[61]]
#>         y1      y2      y3      y4
#> y1 -5.9779  1.0796  0.0267  4.8381
#> y2  5.4393 -2.5748  0.3063 -0.0395
#> y3 -0.2208  6.2879 -6.1863 -0.2941
#> y4  0.1071 -7.5201  1.8880 -6.3136
#> 
#> [[62]]
#>         y1      y2      y3      y4
#> y1 -5.6482  1.3129 -0.1826  4.8776
#> y2  6.0977 -1.9955 -0.5990  0.3349
#> y3 -0.3539  6.5941 -6.0968 -0.2040
#> y4  0.3579 -7.3144  2.8496 -5.5229
#> 
#> [[63]]
#>         y1      y2      y3      y4
#> y1 -5.6285  0.8844  0.2849  4.8488
#> y2  5.2297 -2.4012  0.0733  0.2061
#> y3  0.1713  5.3912 -6.4161  0.1879
#> y4  0.1810 -7.3348  2.8551 -6.0188
#> 
#> [[64]]
#>         y1      y2      y3      y4
#> y1 -5.5283  0.8311  0.1781  4.4061
#> y2  4.9419 -2.6348 -0.7101 -0.2532
#> y3 -0.0070  5.5456 -6.0511  0.0560
#> y4  0.5372 -8.1754  2.2046 -6.1299
#> 
#> [[65]]
#>         y1      y2      y3      y4
#> y1 -6.6001  1.6824 -0.0380  5.2231
#> y2  5.4676 -2.1973  0.5336  0.0204
#> y3  0.6301  5.1240 -5.9167  0.0347
#> y4 -0.0134 -7.2489  2.4992 -5.5505
#> 
#> [[66]]
#>         y1      y2      y3      y4
#> y1 -5.7521  0.7159 -0.2507  4.9922
#> y2  5.8027 -1.9508  0.5002  0.0441
#> y3  0.0835  6.1362 -5.8951 -0.1631
#> y4 -0.6097 -7.3279  2.7955 -6.1510
#> 
#> [[67]]
#>         y1      y2      y3      y4
#> y1 -6.6907  1.1029 -0.2930  5.1709
#> y2  4.9118 -2.5787 -0.0594  0.8663
#> y3  0.2808  6.0329 -6.1878  0.7187
#> y4 -0.0158 -7.4357  2.5586 -6.2412
#> 
#> [[68]]
#>         y1      y2      y3      y4
#> y1 -5.7003  1.5561  0.0553  5.6792
#> y2  5.4253 -2.8970  0.0461 -0.1316
#> y3 -1.0162  6.0995 -5.8691  0.1358
#> y4  0.5630 -7.3931  2.1565 -5.9070
#> 
#> [[69]]
#>         y1      y2      y3      y4
#> y1 -5.6939  1.5281  0.1444  4.8940
#> y2  5.6035 -2.9275  0.0641 -0.6012
#> y3 -0.0134  6.0038 -6.3259 -0.1658
#> y4  0.2364 -7.3385  2.3606 -5.9478
#> 
#> [[70]]
#>         y1      y2      y3      y4
#> y1 -5.9643  1.3434 -0.1344  5.0722
#> y2  5.1277 -2.5828 -0.1238  0.1204
#> y3  0.0389  5.8466 -6.0228  0.2002
#> y4  0.0429 -6.9442  2.5921 -6.1923
#> 
#> [[71]]
#>         y1      y2      y3      y4
#> y1 -5.9243  0.8865  0.0999  4.8679
#> y2  5.2796 -2.5638 -0.2983  0.1686
#> y3  0.0906  6.0471 -6.3269 -0.2134
#> y4 -0.5054 -7.1232  2.6130 -5.6755
#> 
#> [[72]]
#>         y1      y2      y3      y4
#> y1 -5.9200  1.2293 -0.0611  5.4255
#> y2  5.9229 -2.6953  0.1815  0.0032
#> y3 -0.2596  5.8636 -5.5987 -0.0868
#> y4  0.1786 -7.3480  2.7143 -5.7337
#> 
#> [[73]]
#>         y1      y2      y3      y4
#> y1 -6.4274  0.4931  0.2368  4.5700
#> y2  5.4101 -2.5940  0.3928  0.0951
#> y3  0.3073  5.5423 -6.2832  0.0019
#> y4 -0.6478 -7.5113  2.2986 -6.1988
#> 
#> [[74]]
#>         y1      y2      y3      y4
#> y1 -5.9378  1.2743 -0.1804  4.5888
#> y2  5.4514 -2.3675 -0.1271  0.2928
#> y3 -0.4136  5.6969 -5.6865  0.2059
#> y4  0.0874 -7.0510  2.6019 -6.0188
#> 
#> [[75]]
#>         y1      y2      y3      y4
#> y1 -6.5136  1.5369  0.0256  4.9086
#> y2  5.5642 -2.3868 -0.3224  0.1919
#> y3  0.1486  5.7668 -5.8998  0.1100
#> y4 -0.2103 -7.1476  2.6062 -5.8118
#> 
#> [[76]]
#>         y1      y2      y3      y4
#> y1 -6.7212  0.6250  0.0904  5.0172
#> y2  5.5723 -2.5360  0.1952  0.3232
#> y3 -0.3055  5.5932 -5.8743 -0.4162
#> y4  0.0984 -6.7820  2.4282 -5.8759
#> 
#> [[77]]
#>         y1      y2      y3      y4
#> y1 -6.2910  1.2902 -0.2498  5.4488
#> y2  5.1756 -2.4222  0.1451  0.1266
#> y3 -0.2061  5.6235 -6.2612  0.0346
#> y4 -0.3544 -6.9230  2.5766 -6.5837
#> 
#> [[78]]
#>         y1      y2      y3      y4
#> y1 -6.0322  0.9669 -0.3262  4.9366
#> y2  5.2301 -2.4229 -0.2520 -0.0096
#> y3  0.0000  5.7709 -6.3113 -0.2493
#> y4  0.0267 -7.0330  2.7660 -5.3736
#> 
#> [[79]]
#>         y1      y2      y3      y4
#> y1 -5.6171  1.3802  0.3513  5.4468
#> y2  5.5958 -2.1119 -0.3488  0.0908
#> y3  0.2067  5.6944 -5.7429  0.1289
#> y4 -0.1186 -7.4404  2.6279 -6.0663
#> 
#> [[80]]
#>         y1      y2      y3      y4
#> y1 -5.8217  1.6960 -0.1393  4.8420
#> y2  5.1737 -2.6817 -0.1878 -0.4379
#> y3 -0.0756  5.6943 -5.8635  0.0060
#> y4 -0.0536 -7.0538  3.0298 -5.6860
#> 
#> [[81]]
#>         y1      y2      y3      y4
#> y1 -5.9432  1.9662  0.1804  4.8325
#> y2  5.4171 -2.0434 -0.2288  0.0017
#> y3 -0.2531  6.1313 -5.6265  0.1767
#> y4  0.4081 -7.4475  2.6137 -5.4631
#> 
#> [[82]]
#>         y1      y2      y3      y4
#> y1 -6.1795  1.2062 -0.7283  4.5316
#> y2  6.0215 -2.4095 -0.5060  0.2728
#> y3 -0.0582  5.9522 -5.6993  0.7224
#> y4  0.1006 -7.1613  2.6657 -5.4758
#> 
#> [[83]]
#>         y1      y2      y3      y4
#> y1 -5.8417  1.2243  0.6788  5.1682
#> y2  5.4315 -2.4744 -0.2423  0.0133
#> y3 -0.3892  6.0071 -5.9319  0.2591
#> y4 -0.4600 -6.9390  2.4152 -5.9745
#> 
#> [[84]]
#>         y1      y2      y3      y4
#> y1 -6.2975  1.2838  0.2249  4.9901
#> y2  5.5440 -2.2441 -0.3645  0.2449
#> y3  0.1365  5.8487 -5.8212 -0.2674
#> y4 -0.0015 -7.1486  2.5200 -5.7578
#> 
#> [[85]]
#>         y1      y2      y3      y4
#> y1 -5.6978  1.4122  0.1484  4.5891
#> y2  5.4030 -2.3363  0.2806 -0.3610
#> y3 -0.1147  6.0320 -6.1666 -0.0848
#> y4 -0.1262 -7.8710  2.7927 -6.3336
#> 
#> [[86]]
#>         y1      y2      y3      y4
#> y1 -5.8476  0.9337  0.4785  5.2861
#> y2  5.3785 -3.1134  0.0101 -0.4904
#> y3 -0.6114  5.9998 -5.9727 -0.4166
#> y4  0.1387 -7.1715  2.9339 -5.7481
#> 
#> [[87]]
#>         y1      y2      y3      y4
#> y1 -5.8597  1.8850 -0.4586  4.9502
#> y2  5.4700 -2.7928  0.2341 -0.4140
#> y3  0.3105  5.9415 -6.0601 -0.5847
#> y4 -0.1970 -7.9899  2.2574 -6.3043
#> 
#> [[88]]
#>         y1      y2      y3      y4
#> y1 -6.3161  1.2327  0.0271  5.3631
#> y2  5.6828 -2.7258 -0.2832 -0.0979
#> y3 -0.2076  6.2184 -5.7706  0.4866
#> y4  0.3190 -7.3725  2.5438 -6.2305
#> 
#> [[89]]
#>         y1      y2      y3      y4
#> y1 -6.0877  1.2029  0.0267  5.0569
#> y2  5.2536 -2.4098  0.2891 -0.3400
#> y3 -0.0267  5.9468 -5.7560 -0.1364
#> y4 -0.4393 -7.5338  2.6983 -5.8723
#> 
#> [[90]]
#>         y1      y2      y3      y4
#> y1 -6.0322  1.4432 -0.0431  4.9221
#> y2  5.4464 -3.0324 -0.0247  0.0904
#> y3  0.3685  6.4214 -6.1455  0.2421
#> y4 -0.6743 -7.0642  2.5569 -5.7589
#> 
#> [[91]]
#>         y1      y2      y3      y4
#> y1 -6.0988  0.9930 -0.3469  4.7260
#> y2  5.3874 -3.0435  0.1365  0.5614
#> y3  0.1051  6.4381 -6.1544 -0.0754
#> y4  0.1931 -6.9494  1.8696 -6.0270
#> 
#> [[92]]
#>         y1      y2      y3      y4
#> y1 -5.5037  1.4817 -0.1673  4.8613
#> y2  5.2454 -1.6787  0.3387  0.1379
#> y3 -0.1108  5.8373 -6.0303 -0.0616
#> y4  0.2594 -7.3990  2.2196 -5.9596
#> 
#> [[93]]
#>         y1      y2      y3      y4
#> y1 -6.1597  0.9582  0.2737  5.4425
#> y2  5.6030 -2.1250 -0.4149  0.7565
#> y3 -0.3107  6.0609 -6.5682 -0.5186
#> y4  0.3355 -7.2279  2.7876 -5.9105
#> 
#> [[94]]
#>         y1      y2      y3      y4
#> y1 -5.8317  1.7254  0.3621  5.2930
#> y2  5.3161 -3.1388  0.0509  0.3506
#> y3 -0.1596  5.9242 -6.1536  0.3438
#> y4  0.0367 -6.7521  2.0423 -5.5868
#> 
#> [[95]]
#>         y1      y2      y3      y4
#> y1 -6.3414  1.2384  0.2634  5.4223
#> y2  5.2255 -2.3949 -0.5253  0.3888
#> y3 -0.1683  6.2406 -5.7544  0.2706
#> y4 -0.6913 -7.2530  2.0293 -5.5734
#> 
#> [[96]]
#>         y1      y2      y3      y4
#> y1 -5.5362  1.4831  0.1228  4.8351
#> y2  5.2138 -2.0728  0.5201 -0.0191
#> y3  0.0749  5.9327 -6.1675  0.2317
#> y4 -0.0376 -7.4086  2.4437 -5.9435
#> 
#> [[97]]
#>         y1      y2      y3      y4
#> y1 -5.5899  1.5127 -0.1531  5.1957
#> y2  5.2358 -1.6921  0.1887  0.1824
#> y3  0.3952  5.7920 -6.2844 -0.1212
#> y4 -0.3621 -6.9405  2.5557 -5.7922
#> 
#> [[98]]
#>         y1      y2      y3      y4
#> y1 -6.2095  1.4307 -0.0805  4.8610
#> y2  4.8708 -2.7822  0.1615 -0.0977
#> y3  0.1431  5.2626 -5.4246  0.3084
#> y4 -0.2285 -7.5144  2.0456 -5.5210
#> 
#> [[99]]
#>         y1      y2      y3      y4
#> y1 -5.7127  1.1469  0.1237  4.6626
#> y2  4.8916 -2.1677 -0.3059  0.2355
#> y3  0.0269  5.7291 -6.4156 -0.1528
#> y4  0.1273 -7.3281  1.8487 -5.8212
#> 
#> [[100]]
#>         y1      y2      y3      y4
#> y1 -5.4156  1.6693 -0.5459  5.0552
#> y2  5.5648 -2.6804 -0.7217 -0.1677
#> y3 -0.1679  5.6439 -5.7622 -0.8209
#> y4  0.0468 -7.1060  2.5218 -5.8344
#>