Posterior Distribution of Total, Direct, and Indirect Effects of X on Y Through M Over a Specific Time-Interval or a Range of Time-Intervals

This function generates a posterior distribution of the total, direct and indirect effects of the independent variable \(X\) on the dependent variable \(Y\) through mediator variables \(\mathbf{m}\) over a specific time-interval \(\Delta t\) or a range of time-intervals using the posterior distribution of the first-order stochastic differential equation model drift matrix \(\boldsymbol{\Phi}\).

Usage

PosteriorMed(phi, delta_t, from, to, med, ncores = NULL)

Arguments

phi

List of numeric matrices. Each element of the list is a sample from the posterior distribution of the drift matrix (\(\boldsymbol{\Phi}\)). Each matrix should have row and column names pertaining to the variables in the system.

delta_t

Numeric. Time interval (\(\Delta t\)).

from

Character string. Name of the independent variable \(X\) in phi.

to

Character string. Name of the dependent variable \(Y\) in phi.

med

Character vector. Name/s of the mediator variable/s in phi.

ncores

Positive integer. Number of cores to use. If ncores = NULL, use a single core. Consider using multiple cores when number of replications R is a large value.

Value

Returns an object of class ctmedmc which is a list with the following elements:

call

Function call.

args

Function arguments.

fun

Function used (PosteriorMed).

output

A list with length of length(delta_t).

Each element in the output list has the following elements:

est

Mean of the posterior distribution of the total, direct, and indirect effects.

thetahatstar

Posterior distribution of the total, direct, and indirect effects.

Details

See Total(), Direct(), and Indirect() for more details.

References

Bollen, K. A. (1987). Total, direct, and indirect effects in structural equation models. Sociological Methodology, 17, 37. doi:10.2307/271028

Deboeck, P. R., & Preacher, K. J. (2015). No need to be discrete: A method for continuous time mediation analysis. Structural Equation Modeling: A Multidisciplinary Journal, 23 (1), 61–75. doi:10.1080/10705511.2014.973960

Ryan, O., & Hamaker, E. L. (2021). Time to intervene: A continuous-time approach to network analysis and centrality. Psychometrika, 87 (1), 214–252. doi:10.1007/s11336-021-09767-0

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.002704274, -0.001475275, 0.000949122,
    -0.001619422, 0.000885122, -0.000569404,
    0.00085493, -0.000465824, 0.000297815,
    -0.001475275, 0.004428442, -0.002642303,
    0.000980573, -0.00271817, 0.001618805,
    -0.000586921, 0.001478421, -0.000871547,
    0.000949122, -0.002642303, 0.006402668,
    -0.000697798, 0.001813471, -0.004043138,
    0.000463086, -0.001120949, 0.002271711,
    -0.001619422, 0.000980573, -0.000697798,
    0.002079286, -0.001152501, 0.000753,
    -0.001528701, 0.000820587, -0.000517524,
    0.000885122, -0.00271817, 0.001813471,
    -0.001152501, 0.00342605, -0.002075005,
    0.000899165, -0.002532849, 0.001475579,
    -0.000569404, 0.001618805, -0.004043138,
    0.000753, -0.002075005, 0.004984032,
    -0.000622255, 0.001634917, -0.003705661,
    0.00085493, -0.000586921, 0.000463086,
    -0.001528701, 0.000899165, -0.000622255,
    0.002060076, -0.001096684, 0.000686386,
    -0.000465824, 0.001478421, -0.001120949,
    0.000820587, -0.002532849, 0.001634917,
    -0.001096684, 0.003328692, -0.001926088,
    0.000297815, -0.000871547, 0.002271711,
    -0.000517524, 0.001475579, -0.003705661,
    0.000686386, -0.001926088, 0.004726235
  ),
  nrow = 9
)

phi <- MCPhi(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  R = 1000L
)$output

# Specific time-interval ----------------------------------------------------
PosteriorMed(
  phi = phi,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m"
)
#> 
#> Total, Direct, and Indirect Effects
#> 
#> $`1`
#>          interval     est     se    R    2.5%   97.5%
#> total           1 -0.1003 0.0331 1000 -0.1629 -0.0357
#> direct          1 -0.2671 0.0437 1000 -0.3527 -0.1827
#> indirect        1  0.1668 0.0192 1000  0.1314  0.2042
#> 

# Range of time-intervals ---------------------------------------------------
posterior <- PosteriorMed(
  phi = phi,
  delta_t = 1:5,
  from = "x",
  to = "y",
  med = "m"
)

# Methods -------------------------------------------------------------------
# PosteriorMed has a number of methods including
# print, summary, confint, and plot
print(posterior)
#> 
#> Total, Direct, and Indirect Effects
#> 
#> $`1`
#>          interval     est     se    R    2.5%   97.5%
#> total           1 -0.1003 0.0331 1000 -0.1629 -0.0357
#> direct          1 -0.2671 0.0437 1000 -0.3527 -0.1827
#> indirect        1  0.1668 0.0192 1000  0.1314  0.2042
#> 
#> $`2`
#>          interval     est     se    R    2.5%   97.5%
#> total           2  0.0784 0.0372 1000 -0.0008  0.1483
#> direct          2 -0.3208 0.0515 1000 -0.4184 -0.2231
#> indirect        2  0.3992 0.0386 1000  0.3304  0.4754
#> 
#> $`3`
#>          interval     est     se    R    2.5%   97.5%
#> total           3  0.2487 0.0370 1000  0.1747  0.3159
#> direct          3 -0.2923 0.0488 1000 -0.3896 -0.2008
#> indirect        3  0.5409 0.0494 1000  0.4517  0.6395
#> 
#> $`4`
#>          interval     est     se    R    2.5%   97.5%
#> total           4  0.3432 0.0357 1000  0.2710  0.4142
#> direct          4 -0.2392 0.0435 1000 -0.3304 -0.1583
#> indirect        4  0.5823 0.0552 1000  0.4832  0.6994
#> 
#> $`5`
#>          interval     est     se    R    2.5%   97.5%
#> total           5  0.3684 0.0350 1000  0.2993  0.4399
#> direct          5 -0.1854 0.0378 1000 -0.2671 -0.1152
#> indirect        5  0.5538 0.0574 1000  0.4477  0.6784
#> 
summary(posterior)
#>      effect interval         est         se    R          2.5%       97.5%
#> 1     total        1 -0.10028765 0.03305188 1000 -0.1628532567 -0.03569231
#> 2    direct        1 -0.26708215 0.04374925 1000 -0.3526653024 -0.18272081
#> 3  indirect        1  0.16679450 0.01922145 1000  0.1313789256  0.20419138
#> 4     total        2  0.07840908 0.03722557 1000 -0.0008439241  0.14828608
#> 5    direct        2 -0.32082233 0.05148946 1000 -0.4184056128 -0.22307445
#> 6  indirect        2  0.39923141 0.03863736 1000  0.3304258647  0.47536953
#> 7     total        3  0.24868527 0.03696357 1000  0.1747274560  0.31593635
#> 8    direct        3 -0.29226178 0.04881295 1000 -0.3895777054 -0.20080932
#> 9  indirect        3  0.54094705 0.04938299 1000  0.4517322049  0.63946479
#> 10    total        4  0.34315675 0.03573105 1000  0.2709737602  0.41424872
#> 11   direct        4 -0.23918198 0.04350754 1000 -0.3303887570 -0.15831588
#> 12 indirect        4  0.58233872 0.05518154 1000  0.4831636588  0.69936274
#> 13    total        5  0.36838544 0.03499415 1000  0.2992505123  0.43992607
#> 14   direct        5 -0.18537411 0.03777463 1000 -0.2670724148 -0.11523301
#> 15 indirect        5  0.55375955 0.05738803 1000  0.4477157042  0.67844044
confint(posterior, level = 0.95)
#>      effect interval         2.5 %      97.5 %
#> 1     total        1 -0.1628532567 -0.03569231
#> 2    direct        1 -0.3526653024 -0.18272081
#> 3  indirect        1  0.1313789256  0.20419138
#> 4     total        2 -0.0008439241  0.14828608
#> 5    direct        2 -0.4184056128 -0.22307445
#> 6  indirect        2  0.3304258647  0.47536953
#> 7     total        3  0.1747274560  0.31593635
#> 8    direct        3 -0.3895777054 -0.20080932
#> 9  indirect        3  0.4517322049  0.63946479
#> 10    total        4  0.2709737602  0.41424872
#> 11   direct        4 -0.3303887570 -0.15831588
#> 12 indirect        4  0.4831636588  0.69936274
#> 13    total        5  0.2992505123  0.43992607
#> 14   direct        5 -0.2670724148 -0.11523301
#> 15 indirect        5  0.4477157042  0.67844044
plot(posterior)