Monte Carlo Method Confidence Intervals

Monte Carlo Method Confidence Intervals

Usage

# S3 method for ctmedmc
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class ctmedmc.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
confint(mc)
#>     effect interval      2.5 %      97.5 %
#> 1    total        1 -0.1924763 -0.08078604
#> 2   direct        1 -0.4043844 -0.25937576
#> 3 indirect        1  0.1625160  0.21349156

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
confint(mc)
#>      effect interval        2.5 %        97.5 %
#> 1     total        1 -0.153289947 -0.0944838096
#> 2    direct        1 -0.363116708 -0.2825315170
#> 3  indirect        1  0.159559140  0.2185506557
#> 4     total        2  0.052362907  0.1175562220
#> 5    direct        2 -0.431355872 -0.3168226290
#> 6  indirect        2  0.401797187  0.5151754085
#> 7     total        3  0.245513269  0.3081154359
#> 8    direct        3 -0.391916929 -0.2690027616
#> 9  indirect        3  0.562328216  0.6965984477
#> 10    total        4  0.340797280  0.4310212918
#> 11   direct        4 -0.322434758 -0.2043057473
#> 12 indirect        4  0.584316592  0.7534560502
#> 13    total        5  0.362947131  0.4709010362
#> 14   direct        5 -0.252996099 -0.1459567379
#> 15 indirect        5  0.540618311  0.7238971349
#> 16    total        6  0.340576832  0.4534093965
#> 17   direct        6 -0.193966712 -0.1008283621
#> 18 indirect        6  0.467100573  0.6469835621
#> 19    total        7  0.298571915  0.4050368021
#> 20   direct        7 -0.146763759 -0.0681862768
#> 21 indirect        7  0.385691277  0.5510687905
#> 22    total        8  0.251189175  0.3443958916
#> 23   direct        8 -0.110136644 -0.0454653914
#> 24 indirect        8  0.310275178  0.4538391078
#> 25    total        9  0.205938031  0.2828229149
#> 26   direct        9 -0.082249110 -0.0300246057
#> 27 indirect        9  0.245590897  0.3648283331
#> 28    total       10  0.165942594  0.2265026104
#> 29   direct       10 -0.061244976 -0.0196949657
#> 30 indirect       10  0.192498981  0.2877140670
#> 31    total       11  0.131547275  0.1786259796
#> 32   direct       11 -0.045525560 -0.0128579248
#> 33 indirect       11  0.150048840  0.2235160380
#> 34    total       12  0.101372208  0.1388003550
#> 35   direct       12 -0.033805568 -0.0083659507
#> 36 indirect       12  0.116116593  0.1715592233
#> 37    total       13  0.074490100  0.1065757378
#> 38   direct       13 -0.025087158 -0.0054300431
#> 39 indirect       13  0.087172261  0.1303832146
#> 40    total       14  0.054026588  0.0810345441
#> 41   direct       14 -0.018610371 -0.0035182722
#> 42 indirect       14  0.063164182  0.0986686556
#> 43    total       15  0.038870760  0.0611122347
#> 44   direct       15 -0.013802747 -0.0022766953
#> 45 indirect       15  0.045495842  0.0741756159
#> 46    total       16  0.027793922  0.0461670497
#> 47   direct       16 -0.010235837 -0.0014719126
#> 48 indirect       16  0.032470593  0.0554490876
#> 49    total       17  0.019780381  0.0362721011
#> 50   direct       17 -0.007590196 -0.0009509795
#> 51 indirect       17  0.023062123  0.0412505401
#> 52    total       18  0.014028139  0.0284846014
#> 53   direct       18 -0.005628196 -0.0006141192
#> 54 indirect       18  0.016331230  0.0305605036
#> 55    total       19  0.009923596  0.0224638666
#> 56   direct       19 -0.004173314 -0.0003964470
#> 57 indirect       19  0.011540130  0.0231322935
#> 58    total       20  0.007007834  0.0177398282
#> 59   direct       20 -0.003094523 -0.0002558654
#> 60 indirect       20  0.008142701  0.0180956731