Monte Carlo Method Confidence Intervals

Monte Carlo Method Confidence Intervals

Usage

# S3 method for ctmedmc
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object

Object of class ctmedmc.

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

...

additional arguments.

Value

Returns a matrix of confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.002704274, -0.001475275, 0.000949122,
    -0.001619422, 0.000885122, -0.000569404,
    0.00085493, -0.000465824, 0.000297815,
    -0.001475275, 0.004428442, -0.002642303,
    0.000980573, -0.00271817, 0.001618805,
    -0.000586921, 0.001478421, -0.000871547,
    0.000949122, -0.002642303, 0.006402668,
    -0.000697798, 0.001813471, -0.004043138,
    0.000463086, -0.001120949, 0.002271711,
    -0.001619422, 0.000980573, -0.000697798,
    0.002079286, -0.001152501, 0.000753,
    -0.001528701, 0.000820587, -0.000517524,
    0.000885122, -0.00271817, 0.001813471,
    -0.001152501, 0.00342605, -0.002075005,
    0.000899165, -0.002532849, 0.001475579,
    -0.000569404, 0.001618805, -0.004043138,
    0.000753, -0.002075005, 0.004984032,
    -0.000622255, 0.001634917, -0.003705661,
    0.00085493, -0.000586921, 0.000463086,
    -0.001528701, 0.000899165, -0.000622255,
    0.002060076, -0.001096684, 0.000686386,
    -0.000465824, 0.001478421, -0.001120949,
    0.000820587, -0.002532849, 0.001634917,
    -0.001096684, 0.003328692, -0.001926088,
    0.000297815, -0.000871547, 0.002271711,
    -0.000517524, 0.001475579, -0.003705661,
    0.000686386, -0.001926088, 0.004726235
  ),
  nrow = 9
)

# Specific time interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
confint(mc)
#>     effect interval      2.5 %      97.5 %
#> 1    total        1 -0.1666356 -0.04401672
#> 2   direct        1 -0.3567261 -0.18628781
#> 3 indirect        1  0.1272750  0.20061066

# Range of time intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:5,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
confint(mc)
#>      effect interval       2.5 %     97.5 %
#> 1     total        1 -0.16476025 -0.0320114
#> 2    direct        1 -0.33969758 -0.1737522
#> 3  indirect        1  0.12388402  0.1982442
#> 4     total        2  0.01098553  0.1541882
#> 5    direct        2 -0.41779752 -0.2098965
#> 6  indirect        2  0.30898862  0.4662702
#> 7     total        3  0.18841270  0.3099660
#> 8    direct        3 -0.38632534 -0.1918376
#> 9  indirect        3  0.43511469  0.6285914
#> 10    total        4  0.28231133  0.3989232
#> 11   direct        4 -0.32719487 -0.1538166
#> 12 indirect        4  0.46682812  0.6888180
#> 13    total        5  0.30907920  0.4401860
#> 14   direct        5 -0.26357998 -0.1131929
#> 15 indirect        5  0.44035670  0.6682153