Print Method for Object of Class ctmedmc

Print Method for Object of Class ctmedmc

Usage

# S3 method for ctmedmc
print(x, alpha = 0.05, digits = 4, ...)

Arguments

x

an object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
print(mc)
#> $`1`
#>          interval     est     se R    2.5%   97.5%
#> total           1 -0.1175 0.0266 5 -0.1455 -0.0815
#> direct          1 -0.3052 0.0429 5 -0.3440 -0.2380
#> indirect        1  0.1877 0.0182 5  0.1565  0.1997
#> 

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
print(mc)
#> $`1`
#>          interval     est     se R    2.5%   97.5%
#> total           1 -0.1175 0.0530 5 -0.1992 -0.0653
#> direct          1 -0.3052 0.0534 5 -0.3722 -0.2492
#> indirect        1  0.1877 0.0176 5  0.1705  0.2094
#> 
#> $`2`
#>          interval     est     se R    2.5%   97.5%
#> total           2  0.0896 0.0669 5 -0.0224  0.1462
#> direct          2 -0.3547 0.0631 5 -0.4366 -0.2864
#> indirect        2  0.4444 0.0276 5  0.4108  0.4759
#> 
#> $`3`
#>          interval     est     se R    2.5%   97.5%
#> total           3  0.2856 0.0613 5  0.1753  0.3274
#> direct          3 -0.3146 0.0598 5 -0.3946 -0.2513
#> indirect        3  0.6001 0.0280 5  0.5659  0.6316
#> 
#> $`4`
#>          interval     est     se R    2.5%   97.5%
#> total           4  0.3966 0.0471 5  0.3053  0.4183
#> direct          4 -0.2521 0.0530 5 -0.3250 -0.1993
#> indirect        4  0.6487 0.0310 5  0.6162  0.6834
#> 
#> $`5`
#>          interval     est     se R    2.5%   97.5%
#> total           5  0.4311 0.0348 5  0.3634  0.4463
#> direct          5 -0.1924 0.0456 5 -0.2566 -0.1505
#> indirect        5  0.6235 0.0405 5  0.5816  0.6688
#> 
#> $`6`
#>          interval     est     se R    2.5%   97.5%
#> total           6  0.4150 0.0334 5  0.3669  0.4454
#> direct          6 -0.1429 0.0387 5 -0.1983 -0.1106
#> indirect        6  0.5579 0.0522 5  0.5076  0.6245
#> 
#> $`7`
#>          interval     est     se R    2.5%   97.5%
#> total           7  0.3717 0.0407 5  0.3381  0.4158
#> direct          7 -0.1045 0.0325 5 -0.1516 -0.0801
#> indirect        7  0.4762 0.0619 5  0.4200  0.5604
#> 
#> $`8`
#>          interval     est     se R    2.5%   97.5%
#> total           8  0.3174 0.0484 5  0.2764  0.3746
#> direct          8 -0.0758 0.0271 5 -0.1164 -0.0574
#> indirect        8  0.3932 0.0680 5  0.3339  0.4884
#> 
#> $`9`
#>          interval     est     se R    2.5%   97.5%
#> total           9  0.2622 0.0533 5  0.2162  0.3279
#> direct          9 -0.0546 0.0224 5 -0.0904 -0.0410
#> indirect        9  0.3168 0.0701 5  0.2571  0.4161
#> 
#> $`10`
#>          interval     est     se R    2.5%   97.5%
#> total          10  0.2113 0.0547 5  0.1638  0.2795
#> direct         10 -0.0392 0.0184 5 -0.0701 -0.0291
#> indirect       10  0.2505 0.0687 5  0.1929  0.3479
#> 
#> $`11`
#>          interval     est     se R    2.5%   97.5%
#> total          11  0.1671 0.0534 5  0.1207  0.2331
#> direct         11 -0.0281 0.0151 5 -0.0544 -0.0207
#> indirect       11  0.1952 0.0648 5  0.1414  0.2861
#> 
#> $`12`
#>          interval     est     se R    2.5%   97.5%
#> total          12  0.1302 0.0499 5  0.0869  0.1909
#> direct         12 -0.0201 0.0122 5 -0.0422 -0.0146
#> indirect       12  0.1503 0.0591 5  0.1015  0.2320
#> 
#> $`13`
#>          interval     est     se R    2.5%   97.5%
#> total          13  0.1002 0.0451 5  0.0611  0.1538
#> direct         13 -0.0144 0.0099 5 -0.0328 -0.0104
#> indirect       13  0.1146 0.0524 5  0.0715  0.1856
#> 
#> $`14`
#>          interval     est     se R    2.5%   97.5%
#> total          14  0.0764 0.0396 5  0.0421  0.1220
#> direct         14 -0.0103 0.0080 5 -0.0254 -0.0073
#> indirect       14  0.0867 0.0454 5  0.0494  0.1467
#> 
#> $`15`
#>          interval     est     se R    2.5%   97.5%
#> total          15  0.0578 0.0340 5  0.0277  0.0955
#> direct         15 -0.0074 0.0064 5 -0.0197 -0.0052
#> indirect       15  0.0652 0.0386 5  0.0335  0.1146
#> 
#> $`16`
#>          interval     est     se R    2.5%   97.5%
#> total          16  0.0434 0.0286 5  0.0169  0.0746
#> direct         16 -0.0053 0.0051 5 -0.0153 -0.0037
#> indirect       16  0.0487 0.0321 5  0.0214  0.0885
#> 
#> $`17`
#>          interval     est     se R    2.5%   97.5%
#> total          17  0.0324 0.0237 5  0.0096  0.0583
#> direct         17 -0.0038 0.0041 5 -0.0119 -0.0026
#> indirect       17  0.0362 0.0264 5  0.0127  0.0675
#> 
#> $`18`
#>          interval     est     se R    2.5%   97.5%
#> total          18  0.0241 0.0193 5  0.0047  0.0452
#> direct         18 -0.0027 0.0032 5 -0.0092 -0.0018
#> indirect       18  0.0268 0.0213 5  0.0070  0.0509
#> 
#> $`19`
#>          interval     est     se R    2.5%   97.5%
#> total          19  0.0178 0.0154 5  0.0017  0.0349
#> direct         19 -0.0019 0.0026 5 -0.0072 -0.0013
#> indirect       19  0.0197 0.0170 5  0.0033  0.0381
#> 
#> $`20`
#>          interval     est     se R    2.5%   97.5%
#> total          20  0.0131 0.0122 5  0.0000  0.0267
#> direct         20 -0.0014 0.0020 5 -0.0056 -0.0009
#> indirect       20  0.0145 0.0134 5  0.0011  0.0291
#>