Print Method for Object of Class ctmedmc

Print Method for Object of Class ctmedmc

Usage

# S3 method for ctmedmc
print(x, alpha = 0.05, digits = 4, ...)

Arguments

x

an object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.002704274, -0.001475275, 0.000949122,
    -0.001619422, 0.000885122, -0.000569404,
    0.00085493, -0.000465824, 0.000297815,
    -0.001475275, 0.004428442, -0.002642303,
    0.000980573, -0.00271817, 0.001618805,
    -0.000586921, 0.001478421, -0.000871547,
    0.000949122, -0.002642303, 0.006402668,
    -0.000697798, 0.001813471, -0.004043138,
    0.000463086, -0.001120949, 0.002271711,
    -0.001619422, 0.000980573, -0.000697798,
    0.002079286, -0.001152501, 0.000753,
    -0.001528701, 0.000820587, -0.000517524,
    0.000885122, -0.00271817, 0.001813471,
    -0.001152501, 0.00342605, -0.002075005,
    0.000899165, -0.002532849, 0.001475579,
    -0.000569404, 0.001618805, -0.004043138,
    0.000753, -0.002075005, 0.004984032,
    -0.000622255, 0.001634917, -0.003705661,
    0.00085493, -0.000586921, 0.000463086,
    -0.001528701, 0.000899165, -0.000622255,
    0.002060076, -0.001096684, 0.000686386,
    -0.000465824, 0.001478421, -0.001120949,
    0.000820587, -0.002532849, 0.001634917,
    -0.001096684, 0.003328692, -0.001926088,
    0.000297815, -0.000871547, 0.002271711,
    -0.000517524, 0.001475579, -0.003705661,
    0.000686386, -0.001926088, 0.004726235
  ),
  nrow = 9
)

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
print(mc)
#> $`1`
#>          interval     est     se   R    2.5%   97.5%
#> total           1 -0.1000 0.0318 100 -0.1611 -0.0431
#> direct          1 -0.2675 0.0425 100 -0.3409 -0.1894
#> indirect        1  0.1674 0.0199 100  0.1350  0.2109
#> 

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
print(mc)
#> $`1`
#>          interval     est     se   R    2.5%   97.5%
#> total           1 -0.1000 0.0317 100 -0.1592 -0.0366
#> direct          1 -0.2675 0.0379 100 -0.3258 -0.1916
#> indirect        1  0.1674 0.0165 100  0.1282  0.1974
#> 
#> $`2`
#>          interval     est     se   R    2.5%   97.5%
#> total           2  0.0799 0.0388 100  0.0002  0.1525
#> direct          2 -0.3209 0.0446 100 -0.3937 -0.2289
#> indirect        2  0.4008 0.0340 100  0.3247  0.4600
#> 
#> $`3`
#>          interval     est     se   R    2.5%   97.5%
#> total           3  0.2508 0.0393 100  0.1650  0.3261
#> direct          3 -0.2914 0.0426 100 -0.3686 -0.2055
#> indirect        3  0.5423 0.0439 100  0.4516  0.6187
#> 
#> $`4`
#>          interval     est     se   R    2.5%   97.5%
#> total           4  0.3449 0.0366 100  0.2674  0.4052
#> direct          4 -0.2374 0.0386 100 -0.3108 -0.1667
#> indirect        4  0.5823 0.0489 100  0.4917  0.6628
#> 
#> $`5`
#>          interval     est     se   R    2.5%   97.5%
#> total           5  0.3693 0.0335 100  0.2991  0.4197
#> direct          5 -0.1828 0.0342 100 -0.2558 -0.1246
#> indirect        5  0.5521 0.0508 100  0.4663  0.6328
#> 
#> $`6`
#>          interval     est     se   R    2.5%   97.5%
#> total           6  0.3482 0.0315 100  0.2843  0.4043
#> direct          6 -0.1363 0.0297 100 -0.2050 -0.0890
#> indirect        6  0.4846 0.0504 100  0.4028  0.5648
#> 
#> $`7`
#>          interval     est     se   R    2.5%   97.5%
#> total           7  0.3041 0.0303 100  0.2479  0.3601
#> direct          7 -0.0996 0.0254 100 -0.1604 -0.0617
#> indirect        7  0.4037 0.0479 100  0.3201  0.4875
#> 
#> $`8`
#>          interval     est     se   R    2.5%   97.5%
#> total           8  0.2524 0.0292 100  0.2018  0.3067
#> direct          8 -0.0718 0.0213 100 -0.1234 -0.0414
#> indirect        8  0.3242 0.0439 100  0.2500  0.4063
#> 
#> $`9`
#>          interval     est     se   R    2.5%   97.5%
#> total           9  0.2020 0.0277 100  0.1563  0.2542
#> direct          9 -0.0513 0.0176 100 -0.0958 -0.0267
#> indirect        9  0.2533 0.0392 100  0.1927  0.3309
#> 
#> $`10`
#>          interval     est     se   R    2.5%   97.5%
#> total          10  0.1574 0.0260 100  0.1162  0.2092
#> direct         10 -0.0364 0.0144 100 -0.0744 -0.0169
#> indirect       10  0.1938 0.0344 100  0.1415  0.2645
#> 
#> $`11`
#>          interval     est     se   R    2.5%   97.5%
#> total          11  0.1202 0.0241 100  0.0830  0.1656
#> direct         11 -0.0257 0.0116 100 -0.0577 -0.0105
#> indirect       11  0.1459 0.0298 100  0.1035  0.2081
#> 
#> $`12`
#>          interval     est     se   R    2.5%   97.5%
#> total          12  0.0903 0.0221 100  0.0567  0.1334
#> direct         12 -0.0181 0.0093 100 -0.0448 -0.0065
#> indirect       12  0.1085 0.0257 100  0.0724  0.1613
#> 
#> $`13`
#>          interval     est     se   R    2.5%   97.5%
#> total          13  0.0670 0.0200 100  0.0369  0.1077
#> direct         13 -0.0128 0.0074 100 -0.0347 -0.0039
#> indirect       13  0.0798 0.0221 100  0.0486  0.1281
#> 
#> $`14`
#>          interval     est     se   R    2.5%   97.5%
#> total          14  0.0493 0.0179 100  0.0239  0.0868
#> direct         14 -0.0090 0.0058 100 -0.0268 -0.0024
#> indirect       14  0.0582 0.0189 100  0.0328  0.0993
#> 
#> $`15`
#>          interval     est     se   R    2.5%   97.5%
#> total          15  0.0359 0.0159 100  0.0152  0.0699
#> direct         15 -0.0063 0.0046 100 -0.0206 -0.0014
#> indirect       15  0.0422 0.0163 100  0.0204  0.0797
#> 
#> $`16`
#>          interval     est     se   R    2.5%   97.5%
#> total          16  0.0260 0.0139 100  0.0092  0.0562
#> direct         16 -0.0044 0.0036 100 -0.0158 -0.0008
#> indirect       16  0.0304 0.0139 100  0.0132  0.0627
#> 
#> $`17`
#>          interval     est     se   R    2.5%   97.5%
#> total          17  0.0187 0.0121 100  0.0037  0.0453
#> direct         17 -0.0031 0.0028 100 -0.0121 -0.0005
#> indirect       17  0.0218 0.0119 100  0.0080  0.0493
#> 
#> $`18`
#>          interval     est     se   R    2.5%   97.5%
#> total          18  0.0134 0.0104 100  0.0004  0.0365
#> direct         18 -0.0022 0.0022 100 -0.0093 -0.0003
#> indirect       18  0.0156 0.0101 100  0.0040  0.0394
#> 
#> $`19`
#>          interval     est     se   R    2.5%   97.5%
#> total          19  0.0096 0.0089 100 -0.0021  0.0295
#> direct         19 -0.0015 0.0017 100 -0.0071 -0.0001
#> indirect       19  0.0111 0.0085 100  0.0016  0.0315
#> 
#> $`20`
#>          interval     est     se   R    2.5%   97.5%
#> total          20  0.0068 0.0075 100 -0.0028  0.0238
#> direct         20 -0.0011 0.0013 100 -0.0054 -0.0001
#> indirect       20  0.0079 0.0072 100  0.0002  0.0252
#>