Print Method for Object of Class ctmedmcphi

Print Method for Object of Class ctmedmcphi

Usage

# S3 method for ctmedmcphi
print(x, digits = 4, ...)

Arguments

x

an object of class ctmedmcphi.

digits

Integer indicating the number of decimal places to display.

...

further arguments.

Value

Returns the structure of the output.

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
mc <- MCPhi(
  phi = phi,
  vcov_phi_vec = 0.1 * diag(9),
  R = 100L # use a large value for R in actual research
)
print(mc)
#> [[1]]
#>         x       m       y
#> x -1.2172  0.4919 -0.1929
#> m  0.5275 -0.3601 -0.6022
#> y -0.4472  0.4735 -0.7843
#> 
#> [[2]]
#>         x       m       y
#> x -0.1469 -0.0061  0.2046
#> m  0.6589 -0.6010 -0.5694
#> y -0.2909  0.6768 -1.0335
#> 
#> [[3]]
#>         x       m       y
#> x -0.3268  0.3411  0.3774
#> m  0.6648 -0.9441 -0.0834
#> y -0.6912  0.6889 -0.4232
#> 
#> [[4]]
#>         x       m       y
#> x -0.3572  0.0728 -0.3330
#> m  0.5656 -0.4131 -0.2556
#> y -0.3138  1.0532 -0.3379
#> 
#> [[5]]
#>         x       m       y
#> x -0.6029 -0.2135 -0.2440
#> m  0.6596 -0.7103 -0.1610
#> y -0.5020  0.7928 -0.7927
#> 
#> [[6]]
#>         x       m       y
#> x -0.2427  0.3232  0.6557
#> m  0.6457 -0.9032  0.1401
#> y -0.3878  0.5709 -1.4226
#> 
#> [[7]]
#>         x       m       y
#> x -0.5370  0.0156  0.3618
#> m  0.3190 -0.7951 -0.2277
#> y -0.0542  1.3713 -1.1177
#> 
#> [[8]]
#>         x       m       y
#> x -0.4552 -0.0286 -0.1295
#> m  1.0440 -0.5918  0.1796
#> y -0.4036 -0.0903 -0.8073
#> 
#> [[9]]
#>         x       m       y
#> x -0.2958 -0.5262  0.0737
#> m  0.6193 -0.4638 -0.3797
#> y  0.0814  0.9849 -0.7178
#> 
#> [[10]]
#>         x       m       y
#> x -0.1039 -0.4525 -0.1440
#> m  0.9437 -0.6492  0.6602
#> y -0.6863  1.1777 -1.0098
#> 
#> [[11]]
#>         x       m       y
#> x -0.2326 -0.4706 -0.3473
#> m  1.0277 -0.6215  0.1983
#> y -1.2210  1.2009 -0.1973
#> 
#> [[12]]
#>         x       m       y
#> x -0.4326 -0.0417 -0.2028
#> m  0.4779 -0.1082  0.1613
#> y  0.0071  0.3428 -0.8414
#> 
#> [[13]]
#>         x       m       y
#> x -0.7316  0.2577  0.0062
#> m  1.1601 -0.6382  0.1982
#> y -0.4084  0.7628 -0.8988
#> 
#> [[14]]
#>         x       m       y
#> x -0.4596 -0.0327  0.0672
#> m  0.6559 -0.7274 -0.0929
#> y -0.1569  0.8768 -0.4709
#> 
#> [[15]]
#>         x       m       y
#> x -0.2638 -0.1336 -0.3878
#> m  1.4643 -0.3181  0.5830
#> y  0.1441  0.7112 -1.2214
#> 
#> [[16]]
#>         x       m       y
#> x -0.7764  0.2291 -0.1583
#> m  1.0623 -0.8213 -0.2970
#> y -0.6717  0.7516 -0.9436
#> 
#> [[17]]
#>         x       m       y
#> x -0.4852  0.0157  0.0420
#> m  0.8917 -0.8979 -0.3787
#> y -0.9517  0.7118 -0.6513
#> 
#> [[18]]
#>         x       m       y
#> x -0.1140 -0.4394 -0.0162
#> m  0.5869 -0.0183 -0.0830
#> y -0.3946  0.5892 -0.8671
#> 
#> [[19]]
#>         x       m       y
#> x -0.3163 -0.5258  0.3955
#> m  0.8855 -0.3874 -0.0832
#> y  0.2263  0.3631 -0.5809
#> 
#> [[20]]
#>         x       m       y
#> x -0.3721 -0.4929  0.0022
#> m  0.6941 -0.3057 -0.1620
#> y -0.5582  0.7966 -0.2772
#> 
#> [[21]]
#>         x       m       y
#> x -0.3956 -0.0874 -0.7891
#> m  0.6373 -0.3887 -0.4503
#> y -0.3377  0.5391 -0.9641
#> 
#> [[22]]
#>         x       m       y
#> x -0.1952 -0.1310 -0.5027
#> m  0.7944 -0.5916  0.1647
#> y -0.4999  0.8565 -1.0302
#> 
#> [[23]]
#>         x       m       y
#> x -0.7055  0.3022 -0.2534
#> m  0.6804 -1.0679  0.2892
#> y -0.6880  0.5388 -0.7500
#> 
#> [[24]]
#>         x      m       y
#> x -0.1055 -0.054  0.1198
#> m  0.9818 -0.826  0.2905
#> y -0.5751  0.917 -0.7239
#> 
#> [[25]]
#>         x       m       y
#> x -0.8262 -0.6729  0.5037
#> m  0.3675 -0.5961 -0.5110
#> y -0.4603  0.6033 -0.5822
#> 
#> [[26]]
#>         x       m       y
#> x -0.4946  0.0971 -0.0519
#> m  0.6173 -0.8353  0.0690
#> y -0.3678  0.7143 -0.5841
#> 
#> [[27]]
#>         x       m       y
#> x -0.5165 -0.0873  0.2034
#> m  1.0837 -0.5755 -0.4350
#> y -0.5371  0.1620 -0.8398
#> 
#> [[28]]
#>         x       m       y
#> x -0.7633 -0.0024  0.2340
#> m  0.6573 -0.7798 -0.0967
#> y -0.3473  0.8367 -1.4571
#> 
#> [[29]]
#>         x       m       y
#> x -0.2639  0.1110 -0.6344
#> m  1.0274 -0.5936 -0.4582
#> y -0.3495  0.8292 -0.6289
#> 
#> [[30]]
#>         x       m       y
#> x -0.3511  0.2161 -0.1186
#> m  0.5944 -1.2240  0.3725
#> y  0.2009  0.5099 -1.0644
#> 
#> [[31]]
#>         x       m       y
#> x -0.2371 -0.2815 -0.0509
#> m  0.3030 -0.7689 -0.2563
#> y -0.9043  0.5676 -1.0311
#> 
#> [[32]]
#>         x       m       y
#> x -0.4634 -0.0319 -0.0895
#> m  0.7665 -0.3260 -0.0043
#> y -0.0855  0.7239 -1.1413
#> 
#> [[33]]
#>         x       m       y
#> x -0.6282 -0.4019 -0.2183
#> m  0.8706 -0.5780 -0.3740
#> y -0.5571  1.1056 -0.7528
#> 
#> [[34]]
#>         x       m       y
#> x -0.4458 -0.1668 -0.4417
#> m  1.2123 -1.0409  0.1503
#> y -0.5921  0.3628 -0.7737
#> 
#> [[35]]
#>         x       m       y
#> x -0.4865  0.0540 -0.1405
#> m  0.5490 -0.3911 -0.3253
#> y -0.1894  0.8203 -0.7982
#> 
#> [[36]]
#>         x       m       y
#> x -0.7010 -0.1751  0.3510
#> m  0.6111 -0.2954  0.0117
#> y -1.2426  0.7247 -0.5638
#> 
#> [[37]]
#>         x      m       y
#> x -0.7584 -0.308 -0.2499
#> m  0.1954 -0.814 -0.0211
#> y -0.7423  0.577 -0.4350
#> 
#> [[38]]
#>         x       m       y
#> x -0.5522  0.0955 -0.3275
#> m  0.8697 -0.2208 -0.1311
#> y -0.8315  1.0540 -1.0057
#> 
#> [[39]]
#>         x       m       y
#> x -0.3162 -0.2114 -0.5513
#> m  0.8375 -0.9010  0.0718
#> y -0.5343  1.4000 -1.2868
#> 
#> [[40]]
#>         x       m       y
#> x -0.4829 -0.4096 -0.1592
#> m  0.8196 -0.5283  0.0553
#> y -0.3907  0.4510 -0.6091
#> 
#> [[41]]
#>         x       m       y
#> x -0.2512 -0.4291  0.0623
#> m  0.6293 -0.4242 -0.0677
#> y  0.2471  0.7670 -0.5933
#> 
#> [[42]]
#>         x       m       y
#> x -0.3261  0.1321 -0.0052
#> m  1.4503 -0.5574 -0.4600
#> y -0.2843  0.9506 -0.5055
#> 
#> [[43]]
#>         x       m       y
#> x -0.4327 -0.7625 -0.5049
#> m  0.4249 -0.4464 -0.2808
#> y -0.3763  0.8766 -0.3498
#> 
#> [[44]]
#>         x       m       y
#> x -0.3821  0.1696 -0.1162
#> m  1.0972 -0.3101 -0.2028
#> y  0.2021  1.6205 -1.0924
#> 
#> [[45]]
#>         x       m       y
#> x -0.7117 -0.1330  0.1926
#> m  1.0603 -0.4183 -0.1596
#> y -0.6207  0.6832 -0.6556
#> 
#> [[46]]
#>         x       m       y
#> x -0.4716 -0.1449 -0.1319
#> m  0.9354 -1.3138 -0.4519
#> y -0.0609  0.7722 -0.7343
#> 
#> [[47]]
#>         x       m       y
#> x -1.1031  0.0051 -0.1910
#> m  0.8300 -0.4186 -0.0772
#> y -0.1282  1.1730 -0.6704
#> 
#> [[48]]
#>         x       m       y
#> x -0.1419 -0.1596 -0.0327
#> m  1.3562 -0.2028 -0.1193
#> y -0.4559  1.3424 -1.1398
#> 
#> [[49]]
#>         x       m       y
#> x -0.6659 -0.0655 -0.3940
#> m  1.0763 -0.2137  0.3715
#> y -0.4288  1.0932 -0.7463
#> 
#> [[50]]
#>         x       m       y
#> x -0.3634 -0.1231 -0.2175
#> m  0.6415 -0.2317  0.2243
#> y -0.1493  0.5829 -0.9934
#> 
#> [[51]]
#>         x       m       y
#> x -1.1339 -0.0204 -0.0185
#> m  0.9068 -0.8162 -0.1040
#> y -0.4963  0.8611 -0.9517
#> 
#> [[52]]
#>         x       m       y
#> x -0.3093 -0.1571 -0.3162
#> m  0.8738 -0.0776 -0.2072
#> y -0.1177  0.6247 -0.4599
#> 
#> [[53]]
#>         x       m       y
#> x -0.3037  0.1186  0.0297
#> m  1.1143 -0.8160  0.4402
#> y -0.9191  0.6275 -0.7760
#> 
#> [[54]]
#>         x       m       y
#> x -0.3969  0.4908  0.0854
#> m  0.1754 -0.4692  0.0469
#> y -0.2205  1.1180 -0.6608
#> 
#> [[55]]
#>         x       m       y
#> x -0.7153  0.1023  0.1439
#> m  0.9419 -0.4754  0.2073
#> y -0.7439  0.0754 -0.3520
#> 
#> [[56]]
#>         x       m       y
#> x -0.7777 -0.1474 -0.2044
#> m  0.9964 -0.5002  0.0157
#> y -0.2804  0.7292 -1.1278
#> 
#> [[57]]
#>         x       m       y
#> x -0.4153 -0.4723  0.0927
#> m  0.4449 -0.6409  0.0515
#> y -0.1754  1.1809 -0.8690
#> 
#> [[58]]
#>         x       m       y
#> x -0.4934  0.3519 -0.2327
#> m  0.0039 -0.7045 -0.0309
#> y -0.3794  0.4292 -0.4458
#> 
#> [[59]]
#>         x       m       y
#> x -0.5421 -0.2487 -0.1491
#> m  0.5411 -0.5718 -0.0581
#> y -0.6074  0.7933 -0.8436
#> 
#> [[60]]
#>         x       m       y
#> x -0.3456  0.0936 -0.4913
#> m  0.1003 -0.5067 -0.3937
#> y -0.2379  1.0693 -0.8348
#> 
#> [[61]]
#>         x       m       y
#> x -1.3000  0.4276  0.0283
#> m  0.8540 -0.9041 -0.3988
#> y -0.7188  1.0053 -0.5657
#> 
#> [[62]]
#>         x       m       y
#> x -0.6453 -0.4789 -0.0069
#> m  0.5584 -0.8307  0.0346
#> y -0.7198  0.4933 -1.0572
#> 
#> [[63]]
#>         x       m       y
#> x -0.9251  0.5665  0.1412
#> m  0.8678 -0.7354 -0.3602
#> y -0.9364  0.2652 -0.3053
#> 
#> [[64]]
#>         x       m       y
#> x -0.5211 -0.1954 -0.0587
#> m  0.3778 -0.6142  0.0842
#> y -0.1764  0.5220 -0.9322
#> 
#> [[65]]
#>         x       m       y
#> x -0.6479  0.4017 -0.4261
#> m  1.2861 -0.8367  0.0545
#> y -0.4887  0.4285 -1.5571
#> 
#> [[66]]
#>         x       m       y
#> x -0.6375  0.2963  0.2367
#> m  0.9633 -0.9012  0.1863
#> y -0.3656  0.6219 -0.8178
#> 
#> [[67]]
#>         x       m       y
#> x -0.7017 -0.1143 -0.2240
#> m  0.6863 -0.5099  0.1880
#> y  0.0391  0.4070 -1.2275
#> 
#> [[68]]
#>         x       m       y
#> x -0.4962 -0.0001 -0.4282
#> m  1.3859 -0.9315 -0.4332
#> y -0.2218  1.1474 -0.9036
#> 
#> [[69]]
#>         x       m       y
#> x -0.1635 -0.1879 -0.0938
#> m  0.8478 -0.2801 -0.6321
#> y -0.0992  0.9694 -0.4547
#> 
#> [[70]]
#>         x       m       y
#> x -0.2051 -0.6411 -0.2252
#> m  1.2024 -0.5647 -0.3734
#> y -0.6945  1.1056 -1.1280
#> 
#> [[71]]
#>         x       m       y
#> x -0.1659 -0.4218  0.2516
#> m  0.7825 -0.7446 -0.3221
#> y -0.6604  0.7688 -0.1262
#> 
#> [[72]]
#>         x       m       y
#> x -0.0661 -0.1647  0.0870
#> m  0.2358 -0.3202 -0.2502
#> y -0.3998  0.2208 -0.8981
#> 
#> [[73]]
#>         x       m       y
#> x -0.0457 -0.1890  0.2112
#> m  0.6732 -0.4487  0.0859
#> y -0.7282  0.4222 -1.1479
#> 
#> [[74]]
#>         x       m       y
#> x -0.7651 -0.1990  0.0821
#> m  1.4303 -0.6414  0.0986
#> y -0.5074  1.0661 -0.9035
#> 
#> [[75]]
#>         x       m       y
#> x -0.3076 -0.1549  0.5191
#> m  0.6565 -0.9777  0.2289
#> y -0.5923  0.6305 -0.5769
#> 
#> [[76]]
#>         x       m       y
#> x -0.6982 -0.3826 -0.2044
#> m  1.0703 -0.6639 -0.2764
#> y -0.6121  0.3555 -0.4009
#> 
#> [[77]]
#>         x       m       y
#> x -1.1768 -0.0786  0.3476
#> m  0.9151 -0.4544  0.0446
#> y -0.0419  0.5937 -0.6709
#> 
#> [[78]]
#>         x       m       y
#> x -0.4723 -0.1234 -0.2054
#> m  0.3114 -0.8065  0.0575
#> y -0.8986  0.4148 -0.4691
#> 
#> [[79]]
#>         x       m       y
#> x -0.1186 -0.7618 -0.3396
#> m  1.1290 -0.3067 -0.7211
#> y -0.7179  1.3569 -0.5743
#> 
#> [[80]]
#>         x       m       y
#> x -0.6726 -0.1856  0.2652
#> m  0.5196 -0.7446 -0.0715
#> y -0.8199  0.0227 -0.3660
#> 
#> [[81]]
#>         x       m       y
#> x -0.0358  0.0232 -0.3710
#> m  0.7497 -0.3799 -0.0236
#> y -0.4844  0.7677 -1.1273
#> 
#> [[82]]
#>         x       m       y
#> x -0.3283 -0.3251  0.1527
#> m  0.6645 -0.4221  0.2346
#> y -0.3549  0.2996 -0.9707
#> 
#> [[83]]
#>         x       m       y
#> x -0.3166  0.1852 -0.0933
#> m  0.2354 -0.7932 -0.3560
#> y -0.4553  0.4647 -0.7704
#> 
#> [[84]]
#>         x       m       y
#> x -0.6358 -0.1143  0.1549
#> m  0.5104 -0.3494 -0.2068
#> y  0.0857  0.4257 -0.6200
#> 
#> [[85]]
#>         x       m       y
#> x -0.6242  0.0001  0.0071
#> m  0.9168 -0.8359 -0.1133
#> y  0.0473  0.4489 -0.4692
#> 
#> [[86]]
#>         x       m       y
#> x -0.7359 -0.0226 -0.4199
#> m  0.2538 -0.7933  0.2333
#> y -0.6921  1.7188 -0.6349
#> 
#> [[87]]
#>         x       m       y
#> x -0.4809  0.0244 -0.1051
#> m  1.1310 -0.4623  0.1339
#> y -0.4358  0.9036 -0.9574
#> 
#> [[88]]
#>         x       m       y
#> x -0.3950  0.2013  0.0882
#> m  0.5138 -0.4353 -0.3617
#> y  0.0878  0.6523 -0.7692
#> 
#> [[89]]
#>         x       m       y
#> x -0.4429 -0.1148 -0.0416
#> m  0.8045 -0.6687  0.3551
#> y  0.2499  0.7287 -0.7859
#> 
#> [[90]]
#>         x       m       y
#> x -0.3358 -0.2334  0.2430
#> m  0.6873 -0.3681  0.0479
#> y -0.1486  0.3990 -1.0369
#> 
#> [[91]]
#>         x       m       y
#> x -0.2639 -0.2639  0.0445
#> m  0.9109 -0.3741 -0.2463
#> y -0.8524  0.5051 -0.0676
#> 
#> [[92]]
#>         x       m       y
#> x -0.4575 -0.4511 -0.0684
#> m  1.0681 -0.3706  0.4236
#> y -0.6680  0.4215 -0.5603
#> 
#> [[93]]
#>         x       m       y
#> x -0.6091  0.3016 -0.0679
#> m  0.4534 -0.5213 -0.3286
#> y  0.1478  0.7912 -0.5862
#> 
#> [[94]]
#>         x       m       y
#> x -0.7929 -0.1626  0.2671
#> m  0.7873 -0.2396 -0.0121
#> y -0.9847  0.0437 -1.0839
#> 
#> [[95]]
#>         x       m       y
#> x -0.6353  0.2781  0.0343
#> m  0.3844 -0.5588 -0.2128
#> y -0.1787  1.3626 -0.6419
#> 
#> [[96]]
#>         x       m       y
#> x -0.4619 -0.0436 -0.0618
#> m  1.0200 -0.1637 -0.2742
#> y -0.2229  0.8496 -0.5927
#> 
#> [[97]]
#>         x       m       y
#> x -0.4584  0.2681  0.0069
#> m  1.0567 -0.3348 -0.4434
#> y -0.0884  0.9062 -0.1896
#> 
#> [[98]]
#>         x       m       y
#> x -0.3539 -0.1828  0.1273
#> m  1.1478 -0.1777 -0.0039
#> y  0.0514  0.5102 -0.4233
#> 
#> [[99]]
#>         x       m       y
#> x -0.4849  0.1206  0.1912
#> m  0.7250 -0.3485 -0.1506
#> y  0.0221  0.9021 -1.0319
#> 
#> [[100]]
#>         x       m       y
#> x -0.5183  0.0804 -0.2036
#> m  0.5293 -0.5450 -0.5212
#> y -0.5604  0.7025 -1.1414
#>