Summary Method for an Object of Class ctmedmc

Summary Method for an Object of Class ctmedmc

Usage

# S3 method for class 'ctmedmc'
summary(object, alpha = 0.05, ...)

Arguments

object

Object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

...

additional arguments.

Value

Returns a data frame of effects, time intervals, estimates, standard errors, number of Monte Carlo replications, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.00843, 0.00040, -0.00151,
    -0.00600, -0.00033, 0.00110,
    0.00324, 0.00020, -0.00061,
    0.00040, 0.00374, 0.00016,
    -0.00022, -0.00273, -0.00016,
    0.00009, 0.00150, 0.00012,
    -0.00151, 0.00016, 0.00389,
    0.00103, -0.00007, -0.00283,
    -0.00050, 0.00000, 0.00156,
    -0.00600, -0.00022, 0.00103,
    0.00644, 0.00031, -0.00119,
    -0.00374, -0.00021, 0.00070,
    -0.00033, -0.00273, -0.00007,
    0.00031, 0.00287, 0.00013,
    -0.00014, -0.00170, -0.00012,
    0.00110, -0.00016, -0.00283,
    -0.00119, 0.00013, 0.00297,
    0.00063, -0.00004, -0.00177,
    0.00324, 0.00009, -0.00050,
    -0.00374, -0.00014, 0.00063,
    0.00495, 0.00024, -0.00093,
    0.00020, 0.00150, 0.00000,
    -0.00021, -0.00170, -0.00004,
    0.00024, 0.00214, 0.00012,
    -0.00061, 0.00012, 0.00156,
    0.00070, -0.00012, -0.00177,
    -0.00093, 0.00012, 0.00223
  ),
  nrow = 9
)

# Specific time interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
summary(mc)
#>     effect interval        est         se   R       2.5%       97.5%
#> 1    total        1 -0.1000384 0.03485068 100 -0.1727980 -0.03444335
#> 2   direct        1 -0.2674539 0.04544184 100 -0.3530971 -0.18029584
#> 3 indirect        1  0.1674155 0.01890677 100  0.1310468  0.19907046

# Range of time intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:5,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
summary(mc)
#>      effect interval        est         se   R        2.5%       97.5%
#> 1     total        1 -0.1000384 0.02828969 100 -0.15232781 -0.04123191
#> 2    direct        1 -0.2674539 0.03785126 100 -0.34098924 -0.19575030
#> 3  indirect        1  0.1674155 0.01813653 100  0.13034214  0.19730024
#> 4     total        2  0.0799008 0.03177536 100  0.01734713  0.13816481
#> 5    direct        2 -0.3209035 0.05331547 100 -0.43347383 -0.22496835
#> 6  indirect        2  0.4008043 0.04715986 100  0.32090375  0.48438931
#> 7     total        3  0.2508138 0.03411898 100  0.17891544  0.31595326
#> 8    direct        3 -0.2914426 0.06004235 100 -0.42496332 -0.19529856
#> 9  indirect        3  0.5422564 0.07226284 100  0.43134204  0.68245524
#> 10    total        4  0.3449279 0.03955381 100  0.27762014  0.42528849
#> 11   direct        4 -0.2373900 0.06142283 100 -0.37132486 -0.14987686
#> 12 indirect        4  0.5823179 0.08884600 100  0.45427525  0.74867661
#> 13    total        5  0.3692538 0.04543535 100  0.30966378  0.47707320
#> 14   direct        5 -0.1828447 0.05923691 100 -0.31560524 -0.11390776
#> 15 indirect        5  0.5520985 0.09636387 100  0.42321469  0.75325267