Summary Method for an Object of Class ctmedmc

Summary Method for an Object of Class ctmedmc

Usage

# S3 method for ctmedmc
summary(object, alpha = 0.05, ...)

Arguments

object

Object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

...

additional arguments.

Value

Returns a matrix of effects, time-interval, estimates, standard errors, test statistics, p-values, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
summary(mc)
#>     effect interval        est         se R       2.5%       97.5%
#> 1    total        1 -0.1174847 0.03268602 5 -0.1454790 -0.07027877
#> 2   direct        1 -0.3052152 0.05384875 5 -0.3592152 -0.24520672
#> 3 indirect        1  0.1877306 0.02800216 5  0.1459421  0.21373620

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
summary(mc)
#>      effect interval          est          se R          2.5%        97.5%
#> 1     total        1 -0.117484662 0.018731759 5 -0.1604456212 -0.119503443
#> 2    direct        1 -0.305215245 0.029689961 5 -0.3757433772 -0.307074441
#> 3  indirect        1  0.187730583 0.014603550 5  0.1795963047  0.215843461
#> 4     total        2  0.089649670 0.018853439 5  0.0508882911  0.097758433
#> 5    direct        2 -0.354707332 0.029207088 5 -0.4134417371 -0.353007241
#> 6  indirect        2  0.444357002 0.021420021 5  0.4248454867  0.478993355
#> 7     total        3  0.285551674 0.014958547 5  0.2522474777  0.290350904
#> 8    direct        3 -0.314584160 0.028346965 5 -0.3746295227 -0.304659880
#> 9  indirect        3  0.600135834 0.027637078 5  0.5732483514  0.642233772
#> 10    total        4  0.396605704 0.015836611 5  0.3646482571  0.405143092
#> 11   direct        4 -0.252121939 0.028250606 5 -0.3096999424 -0.238757852
#> 12 indirect        4  0.648727643 0.039998361 5  0.6209654478  0.714380648
#> 13    total        5  0.431122443 0.025220472 5  0.4045834306  0.460465868
#> 14   direct        5 -0.192366877 0.027031878 5 -0.2457352357 -0.178855142
#> 15 indirect        5  0.623489320 0.050867478 5  0.5866357709  0.705637001
#> 16    total        6  0.415038108 0.033521943 5  0.3803944972  0.456777884
#> 17   direct        6 -0.142904381 0.024641761 5 -0.1906186108 -0.130532678
#> 18 indirect        6  0.557942490 0.056497596 5  0.5125485531  0.645775117
#> 19    total        7  0.371690575 0.037678575 5  0.3295806304  0.416170939
#> 20   direct        7 -0.104537400 0.021627375 5 -0.1460732516 -0.094072901
#> 21 indirect        7  0.476227975 0.056699177 5  0.4252464010  0.561115033
#> 22    total        8  0.317412261 0.037856153 5  0.2686310977  0.356961670
#> 23   direct        8 -0.075772453 0.018465985 5 -0.1111877300 -0.067339186
#> 24 indirect        8  0.393184713 0.052698022 5  0.3410139772  0.467175596
#> 25    total        9  0.262156401 0.035250004 5  0.2096141358  0.291560501
#> 26   direct        9 -0.054617018 0.015452706 5 -0.0843173778 -0.048019096
#> 27 indirect        9  0.316773419 0.046130277 5  0.2666898025  0.375054421
#> 28    total       10  0.211258400 0.031248855 5  0.1581996254  0.228197692
#> 29   direct       10 -0.039233252 0.012737479 5 -0.0638075700 -0.034167923
#> 30 indirect       10  0.250491651 0.038553522 5  0.2036594311  0.291320861
#> 31    total       11  0.167073060 0.027031187 5  0.1161236951  0.175822660
#> 32   direct       11 -0.028122716 0.010377505 5 -0.0482312964 -0.024282067
#> 33 indirect       11  0.195195777 0.031215864 5  0.1492280698  0.219213792
#> 34    total       12  0.130190355 0.023368023 5  0.0832006636  0.138818304
#> 35   direct       12 -0.020131937 0.008376590 5 -0.0364349540 -0.017244277
#> 36 indirect       12  0.150322292 0.024965408 5  0.1072825016  0.160405732
#> 37    total       13  0.100248253 0.020555577 5  0.0583193144  0.108747763
#> 38   direct       13 -0.014399758 0.006710445 5 -0.0275151261 -0.012241311
#> 39 indirect       13  0.114648012 0.020218673 5  0.0758268381  0.123214561
#> 40    total       14  0.076439602 0.018487227 5  0.0397423234  0.084332138
#> 41   direct       14 -0.010294390 0.005341947 5 -0.0207760757 -0.008687797
#> 42 indirect       14  0.086733992 0.016955722 5  0.0527684277  0.094530253
#> 43    total       15  0.057809232 0.016857452 5  0.0258753209  0.064839277
#> 44   direct       15 -0.007357084 0.004229969 5 -0.0156868803 -0.006165004
#> 45 indirect       15  0.065166316 0.014804605 5  0.0361438894  0.072092061
#> 46    total       16  0.043415873 0.015376361 5  0.0145242450  0.049485203
#> 47   direct       16 -0.005256817 0.003334254 5 -0.0118444378 -0.004374455
#> 48 indirect       16  0.048672691 0.013268020 5  0.0237460103  0.054641941
#> 49    total       17  0.032411079 0.013874346 5  0.0045720135  0.037524465
#> 50   direct       17 -0.003755649 0.002617899 5 -0.0089435561 -0.003103812
#> 51 indirect       17  0.036166728 0.011955126 5  0.0135155696  0.041190289
#> 52    total       18  0.024069586 0.012300555 5 -0.0015904682  0.028293366
#> 53   direct       18 -0.002682950 0.002048422 5 -0.0067535317 -0.002202198
#> 54 indirect       18  0.026752536 0.010662124 5  0.0051630635  0.030899044
#> 55    total       19  0.017792908 0.010679396 5 -0.0050109563  0.021225321
#> 56   direct       19 -0.001916543 0.001598016 5 -0.0051001166 -0.001562468
#> 57 indirect       19  0.019709451 0.009331417 5  0.0000891603  0.023077313
#> 58    total       20  0.013099439 0.009068094 5 -0.0065397449  0.015850610
#> 59   direct       20 -0.001369024 0.001243342 5 -0.0038517603 -0.001108569
#> 60 indirect       20  0.014468463 0.007983967 5 -0.0026879847  0.017166860