Summary Method for an Object of Class ctmedmc

Summary Method for an Object of Class ctmedmc

Usage

# S3 method for ctmedmc
summary(object, alpha = 0.05, ...)

Arguments

object

Object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

...

additional arguments.

Value

Returns a matrix of effects, time-interval, estimates, standard errors, test statistics, p-values, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

data("deboeck2015phi", package = "cTMed")
phi <- deboeck2015phi$dynr$phi
vcov_phi_vec <- deboeck2015phi$dynr$vcov

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
summary(mc)
#>     effect interval        est         se R       2.5%       97.5%
#> 1    total        1 -0.1174847 0.02889313 5 -0.1492508 -0.08189395
#> 2   direct        1 -0.3052152 0.04710723 5 -0.3410064 -0.22054112
#> 3 indirect        1  0.1877306 0.02426227 5  0.1364945  0.19254548

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:20,
  from = "x",
  to = "y",
  med = "m",
  R = 5 # use a large value for R in actual research
)
summary(mc)
#>      effect interval          est          se R         2.5%         97.5%
#> 1     total        1 -0.117484662 0.021280637 5 -0.162258752 -0.1107995621
#> 2    direct        1 -0.305215245 0.039232005 5 -0.346177587 -0.2568493133
#> 3  indirect        1  0.187730583 0.028601616 5  0.138355132  0.2104747871
#> 4     total        2  0.089649670 0.036998344 5  0.006978106  0.0981458302
#> 5    direct        2 -0.354707332 0.051118483 5 -0.425983361 -0.3034545159
#> 6  indirect        2  0.444357002 0.056423859 5  0.339707959  0.4804444661
#> 7     total        3  0.285551674 0.043471291 5  0.195820642  0.3011455794
#> 8    direct        3 -0.314584160 0.061004255 5 -0.421369974 -0.2727891755
#> 9  indirect        3  0.600135834 0.074469142 5  0.472145469  0.6452761198
#> 10    total        4  0.396605704 0.043529968 5  0.301265466  0.4067774898
#> 11   direct        4 -0.252121939 0.066588247 5 -0.376828788 -0.2210024196
#> 12 indirect        4  0.648727643 0.090673199 5  0.522267885  0.7505440210
#> 13    total        5  0.431122443 0.045893895 5  0.341841980  0.4496797944
#> 14   direct        5 -0.192366877 0.066900355 5 -0.320533475 -0.1692998901
#> 15 indirect        5  0.623489320 0.102954221 5  0.511894571  0.7701419745
#> 16    total        6  0.415038108 0.049684906 5  0.339369620  0.4642193111
#> 17   direct        6 -0.142904381 0.063285210 5 -0.265272579 -0.1258192419
#> 18 indirect        6  0.557942490 0.108342372 5  0.466511692  0.7294918905
#> 19    total        7  0.371690575 0.051823919 5  0.312239878  0.4389634062
#> 20   direct        7 -0.104537400 0.057415189 5 -0.216075281 -0.0900767068
#> 21 indirect        7  0.476227975 0.106936168 5  0.405690767  0.6550386874
#> 22    total        8  0.317412261 0.051430699 5  0.273963043  0.3926541492
#> 23   direct        8 -0.075772453 0.050588963 5 -0.174338803 -0.0628505822
#> 24 indirect        8  0.393184713 0.100567854 5  0.341931996  0.5667331146
#> 25    total        9  0.262156401 0.048996895 5  0.227322307  0.3387230829
#> 26   direct        9 -0.054617018 0.043652402 5 -0.139858862 -0.0436378103
#> 27 indirect        9  0.316773419 0.091332421 5  0.281902808  0.4775747492
#> 28    total       10  0.211258400 0.045312810 5  0.179538585  0.2847854550
#> 29   direct       10 -0.039233252 0.037094534 5 -0.111809545 -0.0302099142
#> 30 indirect       10  0.250491651 0.080914548 5  0.218938664  0.3950937247
#> 31    total       11  0.167073060 0.041032073 5  0.137787038  0.2351827311
#> 32   direct       11 -0.028122716 0.031161628 5 -0.089200571 -0.0208782136
#> 33 indirect       11  0.195195777 0.070442097 5  0.166758827  0.3232115964
#> 34    total       12  0.130190355 0.036589618 5  0.103868189  0.1918214484
#> 35   direct       12 -0.020131937 0.025947718 5 -0.071077657 -0.0144149233
#> 36 indirect       12  0.150322292 0.060567608 5  0.124638916  0.2619958058
#> 37    total       13  0.100248253 0.032243561 5  0.077165298  0.1551381115
#> 38   direct       13 -0.014399758 0.021457404 5 -0.056599124 -0.0099472021
#> 39 indirect       13  0.114648012 0.051608766 5  0.092034782  0.2110474012
#> 40    total       14  0.076439602 0.028136720 5  0.056637884  0.1247730470
#> 41   direct       14 -0.010294390 0.017646618 5 -0.045054965 -0.0068624482
#> 42 indirect       14  0.086733992 0.043674950 5  0.067272325  0.1693838850
#> 43    total       15  0.057809232 0.024343682 5  0.041149685  0.1000007683
#> 44   direct       15 -0.007357084 0.014447936 5 -0.035860878 -0.0047339245
#> 45 indirect       15  0.065166316 0.036758846 5  0.048750262  0.1356083786
#> 46    total       16  0.043415873 0.020899509 5  0.029638104  0.0799979264
#> 47   direct       16 -0.005256817 0.011785716 5 -0.028542888 -0.0032656873
#> 48 indirect       16  0.048672691 0.030794768 5  0.035067916  0.1084014983
#> 49    total       17  0.032411079 0.017815419 5  0.021187436  0.0639664261
#> 50   direct       17 -0.003755649 0.009584745 5 -0.022719806 -0.0022530393
#> 51 indirect       17  0.036166728 0.025692413 5  0.025065295  0.0865859602
#> 52    total       18  0.024069586 0.015087171 5  0.015047841  0.0511267122
#> 53   direct       18 -0.002682950 0.007774843 5 -0.018086643 -0.0015546168
#> 54 indirect       18  0.026752536 0.021354868 5  0.017816764  0.0691412002
#> 55    total       19  0.017792908 0.012699966 5  0.010626501  0.0408637829
#> 56   direct       19 -0.001916543 0.006293022 5 -0.014400117 -0.0010728806
#> 57 indirect       19  0.019709451 0.017687503 5  0.012603334  0.0552119871
#> 58    total       20  0.013099439 0.010631804 5  0.007466658  0.0326676874
#> 59   direct       20 -0.001369024 0.005084188 5 -0.011466519 -0.0007405629
#> 60 indirect       20  0.014468463 0.014601988 5  0.008877858  0.0440968614