Summary Method for an Object of Class ctmedmc

Summary Method for an Object of Class ctmedmc

Usage

# S3 method for ctmedmc
summary(object, alpha = 0.05, ...)

Arguments

object

Object of class ctmedmc.

alpha

Numeric vector. Significance level \(\alpha\).

...

additional arguments.

Value

Returns a matrix of effects, time-interval, estimates, standard errors, test statistics, p-values, and confidence intervals.

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
colnames(phi) <- rownames(phi) <- c("x", "m", "y")
vcov_phi_vec <- matrix(
  data = c(
    0.002704274, -0.001475275, 0.000949122,
    -0.001619422, 0.000885122, -0.000569404,
    0.00085493, -0.000465824, 0.000297815,
    -0.001475275, 0.004428442, -0.002642303,
    0.000980573, -0.00271817, 0.001618805,
    -0.000586921, 0.001478421, -0.000871547,
    0.000949122, -0.002642303, 0.006402668,
    -0.000697798, 0.001813471, -0.004043138,
    0.000463086, -0.001120949, 0.002271711,
    -0.001619422, 0.000980573, -0.000697798,
    0.002079286, -0.001152501, 0.000753,
    -0.001528701, 0.000820587, -0.000517524,
    0.000885122, -0.00271817, 0.001813471,
    -0.001152501, 0.00342605, -0.002075005,
    0.000899165, -0.002532849, 0.001475579,
    -0.000569404, 0.001618805, -0.004043138,
    0.000753, -0.002075005, 0.004984032,
    -0.000622255, 0.001634917, -0.003705661,
    0.00085493, -0.000586921, 0.000463086,
    -0.001528701, 0.000899165, -0.000622255,
    0.002060076, -0.001096684, 0.000686386,
    -0.000465824, 0.001478421, -0.001120949,
    0.000820587, -0.002532849, 0.001634917,
    -0.001096684, 0.003328692, -0.001926088,
    0.000297815, -0.000871547, 0.002271711,
    -0.000517524, 0.001475579, -0.003705661,
    0.000686386, -0.001926088, 0.004726235
  ),
  nrow = 9
)

# Specific time-interval ----------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
summary(mc)
#>     effect interval        est         se   R       2.5%       97.5%
#> 1    total        1 -0.1000384 0.03423846 100 -0.1666356 -0.04401672
#> 2   direct        1 -0.2674539 0.04395758 100 -0.3567261 -0.18628781
#> 3 indirect        1  0.1674155 0.02013482 100  0.1272750  0.20061066

# Range of time-intervals ---------------------------------------------------
mc <- MCMed(
  phi = phi,
  vcov_phi_vec = vcov_phi_vec,
  delta_t = 1:30,
  from = "x",
  to = "y",
  med = "m",
  R = 100L # use a large value for R in actual research
)
summary(mc)
#>      effect interval           est           se   R          2.5%         97.5%
#> 1     total        1 -1.000384e-01 0.0339589876 100 -1.647602e-01 -3.201140e-02
#> 2    direct        1 -2.674539e-01 0.0470974310 100 -3.396976e-01 -1.737522e-01
#> 3  indirect        1  1.674155e-01 0.0200762629 100  1.238840e-01  1.982442e-01
#> 4     total        2  7.990080e-02 0.0352031964 100  1.098553e-02  1.541882e-01
#> 5    direct        2 -3.209035e-01 0.0555265317 100 -4.177975e-01 -2.098965e-01
#> 6  indirect        2  4.008043e-01 0.0412536801 100  3.089886e-01  4.662702e-01
#> 7     total        3  2.508138e-01 0.0336890023 100  1.884127e-01  3.099660e-01
#> 8    direct        3 -2.914426e-01 0.0523990077 100 -3.863253e-01 -1.918376e-01
#> 9  indirect        3  5.422564e-01 0.0523393012 100  4.351147e-01  6.285914e-01
#> 10    total        4  3.449279e-01 0.0324431282 100  2.823113e-01  3.989232e-01
#> 11   direct        4 -2.373900e-01 0.0464474361 100 -3.271949e-01 -1.538166e-01
#> 12 indirect        4  5.823179e-01 0.0574698571 100  4.668281e-01  6.888180e-01
#> 13    total        5  3.692538e-01 0.0325493806 100  3.090792e-01  4.401860e-01
#> 14   direct        5 -1.828447e-01 0.0401998392 100 -2.635800e-01 -1.131929e-01
#> 15 indirect        5  5.520985e-01 0.0593964646 100  4.403567e-01  6.682153e-01
#> 16    total        6  3.482375e-01 0.0335613326 100  2.925000e-01  4.286184e-01
#> 17   direct        6 -1.363138e-01 0.0343119913 100 -2.112707e-01 -8.062251e-02
#> 18 indirect        6  4.845513e-01 0.0588841968 100  3.838320e-01  6.026638e-01
#> 19    total        7  3.041486e-01 0.0343265270 100  2.539171e-01  3.884809e-01
#> 20   direct        7 -9.957130e-02 0.0289490530 100 -1.643848e-01 -5.687611e-02
#> 21 indirect        7  4.037199e-01 0.0562456181 100  3.161108e-01  5.285489e-01
#> 22    total        8  2.523884e-01 0.0342054467 100  2.073452e-01  3.407276e-01
#> 23   direct        8 -7.176890e-02 0.0241604548 100 -1.251473e-01 -3.924749e-02
#> 24 indirect        8  3.241573e-01 0.0520016384 100  2.495136e-01  4.551718e-01
#> 25    total        9  2.019841e-01 0.0331649094 100  1.593458e-01  2.895433e-01
#> 26   direct        9 -5.126787e-02 0.0199601732 100 -9.451215e-02 -2.571360e-02
#> 27 indirect        9  2.532520e-01 0.0468405155 100  1.909764e-01  3.768626e-01
#> 28    total       10  1.573830e-01 0.0314410027 100  1.205796e-01  2.418289e-01
#> 29   direct       10 -3.639891e-02 0.0163386710 100 -7.427992e-02 -1.662282e-02
#> 30 indirect       10  1.937819e-01 0.0414079800 100  1.436636e-01  3.014980e-01
#> 31    total       11  1.201652e-01 0.0292921661 100  8.677298e-02  2.025630e-01
#> 32   direct       11 -2.573253e-02 0.0132658635 100 -5.954627e-02 -1.116566e-02
#> 33 indirect       11  1.458977e-01 0.0361687645 100  1.038572e-01  2.429403e-01
#> 34    total       12  9.031231e-02 0.0269083624 100  5.947531e-02  1.707654e-01
#> 35   direct       12 -1.813772e-02 0.0106958998 100 -4.635929e-02 -7.596136e-03
#> 36 indirect       12  1.084500e-01 0.0313763903 100  7.489748e-02  1.942669e-01
#> 37    total       13  6.703448e-02 0.0244112840 100  3.900538e-02  1.412058e-01
#> 38   direct       13 -1.275769e-02 0.0085733315 100 -3.600941e-02 -4.692290e-03
#> 39 indirect       13  7.979217e-02 0.0271153043 100  5.101539e-02  1.548317e-01
#> 40    total       14  4.926167e-02 0.0218822732 100  2.401886e-02  1.125737e-01
#> 41   direct       14 -8.960190e-03 0.0068389830 100 -2.795686e-02 -2.864173e-03
#> 42 indirect       14  5.822186e-02 0.0233688672 100  3.409160e-02  1.280901e-01
#> 43    total       15  3.590906e-02 0.0193850233 100  1.427067e-02  8.816289e-02
#> 44   direct       15 -6.286452e-03 0.0054345577 100 -2.169868e-02 -1.738400e-03
#> 45 indirect       15  4.219552e-02 0.0200794213 100  2.216665e-02  1.063014e-01
#> 46    total       16  2.600312e-02 0.0169752566 100  6.575845e-03  7.133252e-02
#> 47   direct       16 -4.407267e-03 0.0043057334 100 -1.683837e-02 -1.049831e-03
#> 48 indirect       16  3.041039e-02 0.0171856980 100  1.369692e-02  8.453960e-02
#> 49    total       17  1.872757e-02 0.0147013105 100  1.847440e-03  5.856743e-02
#> 50   direct       17 -3.088175e-03 0.0034039082 100 -1.306534e-02 -6.222844e-04
#> 51 indirect       17  2.181574e-02 0.0146383175 100  7.693625e-03  6.524935e-02
#> 52    total       18  1.342697e-02 0.0126014274 100 -1.586529e-03  4.807236e-02
#> 53   direct       18 -2.163065e-03 0.0026869119 100 -1.013712e-02 -3.627938e-04
#> 54 indirect       18  1.559004e-02 0.0124013552 100  4.411704e-03  5.123419e-02
#> 55    total       19  9.590557e-03 0.0107014575 100 -2.789502e-03  3.945144e-02
#> 56   direct       19 -1.514677e-03 0.0021190104 100 -7.864924e-03 -2.089769e-04
#> 57 indirect       19  1.110523e-02 0.0104480888 100  2.343225e-03  4.076531e-02
#> 58    total       20  6.828830e-03 0.0090143362 100 -3.281684e-03  3.237366e-02
#> 59   direct       20 -1.060441e-03 0.0016704786 100 -6.098840e-03 -1.191145e-04
#> 60 indirect       20  7.889270e-03 0.0087560687 100  8.731756e-04  3.326979e-02
#> 61    total       21  4.849589e-03 0.0075412711 100 -3.053383e-03  2.656445e-02
#> 62   direct       21 -7.423229e-04 0.0013169483 100 -4.702842e-03 -7.053738e-05
#> 63 indirect       21  5.591912e-03 0.0073036852 100 -8.760382e-05  2.717717e-02
#> 64    total       22  3.436380e-03 0.0062740211 100 -2.600330e-03  2.179720e-02
#> 65   direct       22 -5.195851e-04 0.0010386734 100 -3.626784e-03 -4.165815e-05
#> 66 indirect       22  3.955965e-03 0.0060685705 100 -5.264391e-04  2.221609e-02
#> 67    total       23  2.430443e-03 0.0051976037 100 -2.093720e-03  1.788532e-02
#> 68   direct       23 -3.636553e-04 0.0008197988 100 -2.796908e-03 -2.454106e-05
#> 69 indirect       23  2.794098e-03 0.0050273912 100 -7.157622e-04  1.817169e-02
#> 70    total       24  1.716259e-03 0.0042929271 100 -1.682706e-03  1.467546e-02
#> 71   direct       24 -2.545080e-04 0.0006476864 100 -2.156910e-03 -1.442368e-05
#> 72 indirect       24  1.970767e-03 0.0041564280 100 -6.461769e-04  1.487125e-02
#> 73    total       25  1.210316e-03 0.0035390465 100 -1.294831e-03  1.204167e-02
#> 74   direct       25 -1.781137e-04 0.0005123194 100 -1.663353e-03 -8.458950e-06
#> 75 indirect       25  1.388430e-03 0.0034324419 100 -4.983056e-04  1.217555e-02
#> 76    total       26  8.525525e-04 0.0029149109 100 -1.011755e-03  9.880593e-03
#> 77   direct       26 -1.246471e-04 0.0004057928 100 -1.282734e-03 -4.845382e-06
#> 78 indirect       26  9.771996e-04 0.0028335051 100 -3.360370e-04  9.972151e-03
#> 79    total       27  5.999624e-04 0.0024005835 100 -7.885456e-04  8.107375e-03
#> 80   direct       27 -8.722859e-05 0.0003218896 100 -9.892110e-04 -2.721592e-06
#> 81 indirect       27  6.871910e-04 0.0023396291 100 -2.265022e-04  8.170004e-03
#> 82    total       28  4.218618e-04 0.0019779894 100 -6.439520e-04  6.652404e-03
#> 83   direct       28 -6.104217e-05 0.0002557311 100 -7.628551e-04 -1.367154e-06
#> 84 indirect       28  4.829040e-04 0.0019331335 100 -1.838051e-04  6.695254e-03
#> 85    total       29  2.964234e-04 0.0016312811 100 -4.492215e-04  5.458560e-03
#> 86   direct       29 -4.271663e-05 0.0002034958 100 -5.882963e-04 -6.495358e-07
#> 87 indirect       29  3.391400e-04 0.0015987716 100 -1.667492e-04  5.487884e-03
#> 88    total       30  2.081593e-04 0.0013469176 100 -2.800054e-04  4.478974e-03
#> 89   direct       30 -2.989242e-05 0.0001621926 100 -4.536818e-04 -2.813071e-07
#> 90 indirect       30  2.380517e-04 0.0013236636 100 -1.274000e-04  4.499048e-03